refactor(proxy): ProxyManager 抽 SingBoxConfigBuilder（config-gen 模块化）+ index.ts 拆分
ProxyManager 7999→5093（-36%）：generateSingBoxConfig 全 config-gen 逐字抽到 9 个模块——
singbox-config-types（14 接口）/ config-helpers（共享纯 helper）/ outbound-builder / dns-builder /
custom-rules / route-builder / inbounds-builder / log-builder。

离线等价保证：先建 config-snapshot 字节级 characterization 网（26 快照，「config 字节相同 = sing-box
行为相同」），每步抽取后快照 0 written 即离线证等价；启停/TUN 热切换真机另验（已过）。
注入模式：值注入（coreVersion/probe/lanResolver）/ 读回调（log）/ 副作用写回调（onDegraded）/
双载体返回值回写（pendingEndpoints/pendingRuleSelectors）/ gateInvalidNodes 就地改 Map。

index.ts 1798→1508：抽 tray-actions（13 托盘动作回调，getter 注入 live-ref）/ startup-tasks（自动连接+
内核落位+自动检查更新）/ config-change-handler（CONFIG_CHANGED 监听，getPrivacyMode 注入防循环依赖）；
断 TrayManager→index 循环依赖（删 handleEnterPrivacyMode 死 fallback）。

单测：各 builder 纯逻辑单测（含三平台 mock 矩阵）+ 提取期已存测试改 import 新模块；gate 全绿。
