+++ src/main/services/ProxyManager.ts
@@ -27,67 +27,48 @@ import { PlatformPrivilegeService } from './PlatformPrivilegeService';
 import { type ISystemProxyManager, SystemProxyBase } from './SystemProxyManager';
 import { type ISystemDnsManager } from './SystemDnsManager';
 import { localProxyPort, controlApiPort } from '../../shared/proxy-ports';
+import {
+  isIpv4Host,
+  isIpv6Host,
+  effectiveAppRules,
+  applyRuleSetPrune,
+} from './singbox-config-helpers';
+import { buildDnsConfig } from './singbox-dns-builder';
+import { buildRouteConfig } from './singbox-route-builder';
+import { buildInbounds } from './singbox-inbounds-builder';
+import { buildLogConfig } from './singbox-log-builder';
 import { IPC_CHANNELS } from '../../shared/ipc-channels';
 import { LOGIN_ITEMS_SETTINGS_URL } from '../../shared/constants';
 import { resourceManager } from './ResourceManager';
+import { seedBuiltinRuleSets } from './builtin-geo-rulesets';
+import type { SingBoxConfig, SingBoxOutbound, SingBoxEndpoint } from './singbox-config-types';
 import {
-  BUILTIN_GEO_RULESETS,
-  getRuleSetRuntimeDir as getRuntimeRulesDir,
-  seedBuiltinRuleSets,
-  isValidSrsFile,
-  resolveBuiltinRuleSetRefMeta,
-} from './builtin-geo-rulesets';
+  buildProxyOutbound,
+  isNodeUsable,
+  buildWireGuardEndpoint,
+  prunedSelectorDefault,
+  buildOutbounds,
+  type PendingRuleSelector,
+} from './singbox-outbound-builder';
 import { retry } from '../utils/retry';
 import { coreVersionAtLeast } from '../../shared/version';
 import { parseTailscaleAuthLine } from '../../shared/tailscale';
-import { endpointForcedRouteCidrs, isEndpointProtocol } from '../../shared/endpoint-routes';
-import { effectiveLogLevel } from '../../shared/log-level';
 import {
   getUserDataPath,
   getSingBoxConfigPath,
   getSingBoxLogPath,
   getSingBoxPidPath,
   getCachePath,
-  getRuleResourcesPath,
   getCustomRulesDir,
 } from '../utils/paths';
-import { getAppPreset } from '../../shared/app-rules-preset';
-import {
-  parseDnsServerSpec,
-  type ParsedDnsServer,
-  BOOTSTRAP_DIRECT_DNS_IPS,
-} from '../../shared/dns';
 import { ruleConditions } from '../../shared/rules';
-import {
-  EXT_TYPES,
-  planCustomRule,
-  customRuleFileBase,
-  buildCustomRuleFiles,
-  usesFakeIp,
-  condMatcherFields,
-} from './custom-rule-files';
+import { planCustomRule, buildCustomRuleFiles, condMatcherFields } from './custom-rule-files';
 import coreManifest from '../../shared/core-manifest.json';
 
 /**
  * 私有 IP 地址段（CIDR 格式）
  * 用于路由规则中的直连配置
  */
-const PRIVATE_IP_CIDRS = [
-  // IPv4 私有地址
-  '10.0.0.0/8',
-  '172.16.0.0/12',
-  '192.168.0.0/16',
-  '127.0.0.0/8',
-  '169.254.0.0/16',
-  '224.0.0.0/4',
-  '240.0.0.0/4',
-  // IPv6 私有地址
-  '::1/128', // loopback
-  'fc00::/7', // unique local address (ULA)
-  'fe80::/10', // link-local
-  'ff00::/8', // multicast
-];
-
 /**
  * 私有 IP 地址正则表达式
  * 用于日志过滤中识别内网请求
@@ -99,373 +80,6 @@ const PRIVATE_IP_PATTERNS = [
   /\b127\.\d{1,3}\.\d{1,3}\.\d{1,3}/,
   /\b169\.254\.\d{1,3}\.\d{1,3}/,
 ];
-
-/**
- * 国内常见网银 U盾插件及本地证券/炒股软件的专属域名
- * 用于绕过代理，防止被 FakeIP 劫持或因协议不兼容（如二进制协议通过 HTTP 代理）被阻断
- */
-const DOMESTIC_BANK_AND_STOCK_DOMAINS = [
-  // U盾及网银相关（通常指向 127.0.0.1）
-  '.microdone.cn', // 微动（杭州银行、中信银行等地方和股份制网银插件常用）
-  '.icbc.com.cn', // 工商银行
-  '.boc.cn', // 中国银行
-  '.ccb.com', // 建设银行
-  '.abchina.com',
-  '.abchina.com.cn', // 农业银行
-  '.bankcomm.com', // 交通银行
-  '.cmbchina.com', // 招商银行
-  '.psbc.com', // 邮储银行
-  '.spdb.com.cn', // 浦发银行
-  '.cebbank.com', // 光大银行
-  '.citicbank.com', // 中信银行
-  '.pingan.com', // 平安银行
-  '.cib.com.cn', // 兴业银行
-  '.hxb.com.cn', // 华夏银行
-  '.cmbc.com.cn', // 民生银行
-  '.hzbank.com.cn', // 杭州银行
-
-  // 证券炒股软件相关（经常使用定制化的 TCP 二进制协议通信，在 SOCKS/HTTP 系统代理模式下会导致握手失败并被代理核心主动断开）
-  '.10jqka.com.cn',
-  '.thsi.cn', // 同花顺
-  '.eastmoney.com',
-  '.1234567.com.cn', // 东方财富
-  '.gw.com.cn', // 大智慧
-  '.tdx.com.cn', // 通达信
-];
-
-/**
- * fake-ip-filter 默认清单：FakeIP 模式下被假 IP（198.18/15）弄坏的通用域名 → 强制走真实 DNS 解析（绕过 FakeIP）。
- * **刻意内联硬编码**——fake-ip-filter 是启动必经的 DNS 规则，绝不能依赖可能 404 的远程 rule_set（重蹈 geoip-telegram FATAL）。
- * 默认注入（config.fakeIpFilter !== false 时），按"应路由到哪个真实解析器"分组：
- *   - 连通性探测 / Captive Portal → dns-local（系统解析，反映真实本地网络，否则系统误判断网 / 锁屏登录态卡死）
- *   - NTP / STUN / TURN → dns-domestic（真实 DoH；校时握手与 WebRTC/P2P 需真实 IP）
- * 本地域（.local/.arpa/.lan/.home.arpa）已由上方 dns-local 规则覆盖，不重复。
- */
-const FAKEIP_FILTER_CAPTIVE_DOMAINS = [
-  'captive.apple.com', // Apple 连通性探测
-  'connectivitycheck.gstatic.com', // Android 连通性探测
-  'connectivitycheck.android.com',
-  'msftconnecttest.com', // Windows NCSI
-  'www.msftconnecttest.com',
-  'msftncsi.com',
-  'www.msftncsi.com',
-  'dns.msftncsi.com',
-  'detectportal.firefox.com', // Firefox captive 检测
-  'network-test.debian.org',
-  'connect.rom.miui.com', // 小米连通性
-];
-/** NTP 时间同步域名（FakeIP 下校时失败）。 */
-const FAKEIP_FILTER_NTP_SUFFIXES = [
-  'ntp.org', // pool.ntp.org 及各区域子域
-  'time.windows.com',
-  'time.apple.com',
-  'time.cloudflare.com',
-  'time.nist.gov',
-  'time.android.com',
-];
-// NTP/STUN 关键字：domain_keyword 是**裸子串**匹配，故只保留误伤面极小的 'ntp'/'stun'；刻意去掉 'turn'
-// （会误命中 saturn/return/nocturne/overturn 等正常域名 → 被强制走 dns-domestic 真实解析、绕过 FakeIP，海外域可能被污染）。
-// TURN 服务器通常同时是 STUN（stun.* 已覆盖）；NTP 域另由 FAKEIP_FILTER_NTP_SUFFIXES 的 domain_suffix 兜底。
-const FAKEIP_FILTER_NTP_STUN_KEYWORDS = ['ntp', 'stun'];
-
-/**
- * sing-box 1.12.x / 1.13.x 配置类型定义
- */
-
-interface SingBoxLogConfig {
-  level: string;
-  timestamp: boolean;
-  output?: string;
-  disabled?: boolean;
-}
-
-interface SingBoxDnsServer {
-  tag: string;
-  type?: string;
-  server?: string;
-  server_port?: number;
-  /** DoH path, e.g. "/dns-query" */
-  path?: string;
-  /** Bootstrap resolver tag: required when server is a domain name (sing-box 1.12+ new format) */
-  domain_resolver?: string;
-  detour?: string;
-  // Legacy / compat fields (not emitted in new format)
-  address?: string;
-  address_resolver?: string;
-  // FakeIP specific
-  inet4_range?: string;
-  inet6_range?: string;
-}
-
-interface SingBoxDnsRule {
-  rule_set?: string;
-  query_type?: string[];
-  domain?: string[];
-  domain_suffix?: string[];
-  domain_keyword?: string[];
-  server: string;
-}
-
-interface SingBoxFakeIPConfig {
-  enabled: boolean;
-  inet4_range?: string;
-  inet6_range?: string;
-}
-
-interface SingBoxDnsConfig {
-  servers: SingBoxDnsServer[];
-  rules?: SingBoxDnsRule[];
-  final?: string;
-  strategy?: string;
-  fakeip?: SingBoxFakeIPConfig;
-  // 关 FakeIP 时注入：用 DNS 解析结果反查域名补无 SNI/ECH 流量的路由匹配（不改节点收 IP 事实）。
-  reverse_mapping?: boolean;
-}
-
-interface SingBoxInbound {
-  type: string;
-  tag: string;
-  listen?: string;
-  listen_port?: number;
-  // TUN 模式
-  interface_name?: string;
-  address?: string[];
-  mtu?: number;
-  auto_route?: boolean;
-  strict_route?: boolean;
-  stack?: string;
-  sniff?: boolean;
-  sniff_override_destination?: boolean; // Keep for interface compatibility if needed by types, but won't be used for 1.13+
-  route_exclude_address?: string[];
-  platform?: {
-    http_proxy?: {
-      enabled: boolean;
-      server: string;
-      server_port: number;
-    };
-  };
-}
-
-interface SingBoxOutbound {
-  type: string;
-  tag: string;
-  detour?: string; // 代理链
-  server?: string;
-  server_port?: number;
-  override_address?: string;
-  // Shadowsocks
-  method?: string;
-  password?: string;
-  username?: string;
-  plugin?: string;
-  plugin_opts?: string;
-  // VLESS / VMess
-  uuid?: string;
-  security?: string; // vmess specific
-  alter_id?: number; // vmess specific
-  flow?: string;
-  packet_encoding?: string;
-  // Trojan and Hysteria2
-  // password?: string; // Shared with SS
-  // Hysteria2 specific
-  up_mbps?: number;
-  down_mbps?: number;
-  obfs?: {
-    type: string;
-    password: string;
-  };
-  network?: string;
-  // naive specific: 走 HTTP/3 (QUIC) 传输
-  quic?: boolean;
-  // TUIC specific
-  congestion_control?: string;
-  udp_relay_mode?: string;
-  zero_rtt_handshake?: boolean;
-  heartbeat?: string;
-  // ShadowTLS specific
-  version?: number;
-  // AnyTLS specific
-  idle_session_check_interval?: string;
-  idle_session_timeout?: string;
-  min_idle_session?: number;
-  // TLS
-  tls?: {
-    enabled: boolean;
-    server_name?: string;
-    insecure?: boolean;
-    alpn?: string[];
-    utls?: {
-      enabled: boolean;
-      fingerprint: string;
-    };
-    reality?: {
-      enabled: boolean;
-      public_key: string;
-      short_id: string;
-    };
-    ech?: { enabled: boolean; config?: string[] };
-    fragment?: boolean;
-  };
-  // Transport
-  transport?: {
-    type: string;
-    path?: string;
-    host?: string;
-    headers?: Record<string, string | string[]>;
-    service_name?: string;
-    max_early_data?: number;
-    early_data_header_name?: string;
-  };
-  // Multiplex 多路复用
-  multiplex?: {
-    enabled: boolean;
-    protocol?: string;
-    max_connections?: number;
-    min_streams?: number;
-    padding?: boolean;
-  };
-  // Hysteria2 端口跳跃
-  server_ports?: string[];
-  hop_interval?: string;
-  // DNS resolver for outbound server domain
-  domain_resolver?: string;
-  // UDP over TCP (UoT)
-  udp_over_tcp?: {
-    enabled: boolean;
-    version: number;
-  };
-  // Direct outbound: UDP fragmentation (also used to mark outbound as "non-empty" for sing-box 1.13+ validation)
-  udp_fragment?: boolean;
-  // SSH specific
-  user?: string;
-  private_key?: string;
-  private_key_path?: string;
-  private_key_passphrase?: string;
-  host_key?: string[];
-  host_key_algorithms?: string[];
-  client_version?: string;
-  // selector specific（用于 clash_api 热切换节点：default=当前选中，interrupt_exist_connections=切换时是否中断现有连接）
-  outbounds?: string[];
-  default?: string;
-  interrupt_exist_connections?: boolean;
-}
-
-interface SingBoxWireGuardPeer {
-  address: string;
-  port: number;
-  public_key: string;
-  pre_shared_key?: string;
-  allowed_ips: string[];
-  persistent_keepalive_interval?: number;
-  reserved?: number[];
-}
-
-// sing-box endpoint（1.11+）：独立于 outbound 的顶层 endpoints[] 元素，其 tag 可被 route/selector 当 outbound 引用。
-// WireGuard + Tailscale（默认用户态：WG system=false / TS system_interface=false，零提权）。
-interface SingBoxEndpoint {
-  type: string;
-  tag: string;
-  // WireGuard
-  system?: boolean;
-  mtu?: number;
-  address?: string[];
-  private_key?: string;
-  listen_port?: number;
-  peers?: SingBoxWireGuardPeer[];
-  udp_timeout?: string;
-  workers?: number;
-  // Tailscale（账号制 mesh；默认 tsnet 用户态。system_interface=Phase 2 反向 mesh）
-  auth_key?: string;
-  state_directory?: string;
-  control_url?: string;
-  hostname?: string;
-  exit_node?: string;
-  exit_node_allow_lan_access?: boolean;
-  accept_routes?: boolean;
-  ephemeral?: boolean;
-  advertise_routes?: string[];
-  system_interface?: boolean;
-}
-
-interface SingBoxRouteRule {
-  protocol?: string;
-  network?: string[];
-  rule_set?: string | string[];
-  domain?: string[];
-  domain_suffix?: string[];
-  domain_keyword?: string[];
-  domain_regex?: string[];
-  geosite?: string[];
-  ip_cidr?: string[];
-  source_ip_cidr?: string[];
-  port?: number | number[];
-  port_range?: string[];
-  source_port?: number | number[];
-  source_port_range?: string[];
-  process_name?: string | string[];
-  process_path?: string | string[];
-  process_name_not?: string | string[]; // sing-box 1.13+
-  inbound?: string | string[]; // sing-box 1.13+
-  action?: string; // logical 子规则为纯 matcher 无 action；default/logical 外层显式设 'route'
-  outbound?: string;
-  sniffer?: string[];
-  rewrite_target?: boolean; // sing-box 1.12+
-  timeout?: string;
-  domain_resolver?: string; // sing-box 1.13+: 指定该规则使用的 DNS 解析器
-  override_address?: string; // sing-box 1.13+: 在规则层强制修改目标地址
-  // logical 规则（多条件跨维度 OR / AND）：type:'logical' + mode + rules(纯 matcher 子规则，无 action/outbound)
-  type?: string;
-  mode?: string;
-  rules?: SingBoxRouteRule[];
-}
-
-interface SingBoxRuleSet {
-  tag: string;
-  type: string;
-  format: string;
-  path?: string;
-  url?: string;
-  download_detour?: string;
-  // remote rule_set 更新周期；不填 sing-box 用隐式默认，显式设定避免长期不更新 / 频繁拉取
-  update_interval?: string;
-}
-
-interface SingBoxRouteConfig {
-  rule_set?: SingBoxRuleSet[];
-  rules: SingBoxRouteRule[];
-  default_domain_resolver?: string;
-  auto_detect_interface?: boolean;
-  final?: string;
-}
-
-/**
- * 浏览器隐私 DoH 泄漏域名（DoH-over-HTTPS / DoH-over-QUIC）。route reject 与 DNS 拦截须用同一份清单，
- * 避免某处漏掉某域名导致 DoH 绕过 hijack-dns / FakeIP 体系。改这一处即两处同步。
- */
-const DOH_LEAK_DOMAIN_KEYWORDS = [
-  'dns.google',
-  'cloudflare-dns.com',
-  'doh.opendns.com',
-  'dns.quad9.net',
-  'one.one.one.one',
-];
-
-/** 主机字符串是否为 IPv4 字面量（与 DNS/route 生成各处保持同一判定，避免分类不一致）。 */
-const isIpv4Host = (host: string): boolean => /^(?:[0-9]{1,3}\.){3}[0-9]{1,3}$/.test(host);
-
-/** 主机字符串是否为 IPv6 字面量。 */
-const isIpv6Host = (host: string): boolean => /^[0-9a-fA-F:]+$/.test(host) && host.includes(':');
-
-/**
- * QUIC(UDP/443) reject 规则工厂：可选叠加域名/进程等匹配器。route 与各处 blockQuic 共用，
- * 保证 network/port/action 字面量始终一致（避免某处漏写 network 导致行为漂移）。
- */
-const udp443RejectRule = (matcher: Record<string, unknown> = {}): SingBoxRouteRule => ({
-  ...matcher,
-  network: ['udp'],
-  port: [443],
-  action: 'reject',
-});
 
 /**
  * 从 sing-box `check` 的 stderr 解析出错出站的数组下标。覆盖两种措辞：
@@ -490,35 +104,6 @@ export function parseCheckEndpointIndex(stderr: string): number | null {
   if (!m) return null;
   const idx = Number(m[1]);
   return Number.isInteger(idx) && idx >= 0 ? idx : null;
-}
-
-interface SingBoxExperimental {
-  cache_file?: {
-    enabled: boolean;
-    path: string;
-    store_fakeip?: boolean;
-    store_rdrc?: boolean;
-  };
-}
-
-interface SingBoxConfig {
-  log: SingBoxLogConfig;
-  dns?: SingBoxDnsConfig;
-  inbounds: SingBoxInbound[];
-  outbounds: SingBoxOutbound[];
-  endpoints?: SingBoxEndpoint[];
-  route?: SingBoxRouteConfig;
-  experimental?: SingBoxExperimental & {
-    clash_api?: {
-      external_controller: string;
-      external_ui?: string;
-      secret?: string;
-      external_ui_download_url?: string;
-      external_ui_download_detour?: string;
-      default_mode?: string;
-      cache_file?: string;
-    };
-  };
 }
 
 export interface IProxyManager {
@@ -624,12 +209,7 @@ export class ProxyManager extends EventEmitter implements IProxyManager {
   // 生成侧暂存：本轮 generateOutbounds 产出的 rule-sel selector 列表（selectorTag/memberTag/targetServerId），
   // 供 generateSingBoxConfig 末尾按 targetServerId → currentConfig 解析回 ruleKey 填充 currentRuleTargetMap。
   // 仅 generateRuleSelectors 写、generateSingBoxConfig 末尾读，跨这两步的临时载体（非持久态）。
-  private pendingRuleSelectors: {
-    ruleKey: string;
-    selectorTag: string;
-    memberTag: string;
-    targetServerId?: string;
-  }[] = [];
+  private pendingRuleSelectors: PendingRuleSelector[] = [];
   // 生成侧暂存：本轮 generateOutbounds 产出的 endpoint（WireGuard 等，sing-box 顶层 endpoints[]）。
   // 仅 generateOutbounds 写、generateSingBoxConfig 末尾读，注入 SingBoxConfig.endpoints（非持久态）。
   private pendingEndpoints: SingBoxEndpoint[] = [];
@@ -822,7 +402,7 @@ export class ProxyManager extends EventEmitter implements IProxyManager {
       const dns = require('dns').promises;
       const allServerIds = new Set([
         config.selectedServerId as string,
-        ...this.effectiveAppRules(config).map((r) => r.targetServerId),
+        ...effectiveAppRules(config).map((r) => r.targetServerId),
       ]);
 
       const resolvePromises = Array.from(allServerIds).map(async (serverId) => {
@@ -2056,7 +1636,7 @@ done
       throw new Error('Selected server not found');
     }
     // 选中节点不可用（naive 缺 libcronet）→ 明确报错，不静默切到别的节点（修 review M1）
-    if (!this.isNodeUsable(selectedServer)) {
+    if (!isNodeUsable(selectedServer)) {
       throw new Error(this.naiveUnavailableReason(selectedServer));
     }
 
@@ -2096,12 +1676,38 @@ done
     // 记录本次生成的 id→tag 映射，供 clash_api 热切换定位 selector 成员 tag（见 hotSwitchNode）
     this.currentIdToTagMap = idToTagMap;
 
+    // outbounds 须先于 route/endpoints 生成：其产出的两载体（pendingEndpoints / pendingRuleSelectors）
+    // 由下方 route deps + endpoints 注入 + 末尾 currentRuleTargetMap 回填消费。回写 this.* 维持原时序。
+    const outboundsResult = buildOutbounds(selectedServer, config, idToTagMap, {
+      coreVersion: this.coreVersion,
+      gateInvalidNodes: this.gateInvalidNodes,
+      log: (level, message) => this.logToManager(level, message),
+    });
+    this.pendingEndpoints = outboundsResult.pendingEndpoints;
+    this.pendingRuleSelectors = outboundsResult.pendingRuleSelectors;
+
     const singboxConfig: SingBoxConfig = {
-      log: this.generateLogConfig(config),
-      dns: this.generateDnsConfig(config, 'proxy-selector'),
-      inbounds: this.generateInbounds(config, resolvedIps),
-      outbounds: this.generateOutbounds(selectedServer, config, idToTagMap),
-      route: this.generateRouteConfig(config, idToTagMap),
+      log: buildLogConfig(config, this.privacyProvider()),
+      dns: buildDnsConfig(config, 'proxy-selector', this.lanResolverForDns, (level, message) =>
+        this.logToManager(level, message)
+      ),
+      inbounds: buildInbounds(config, resolvedIps, {
+        coreVersion: this.coreVersion,
+        probeDirectPort: this.probeDirectPort,
+        probeProxyPort: this.probeProxyPort,
+      }),
+      outbounds: outboundsResult.outbounds,
+      route: buildRouteConfig(config, idToTagMap, {
+        coreVersion: this.coreVersion,
+        probeDirectPort: this.probeDirectPort,
+        probeProxyPort: this.probeProxyPort,
+        lanResolverForDns: this.lanResolverForDns,
+        pendingEndpoints: this.pendingEndpoints,
+        log: (level, message) => this.logToManager(level, message),
+        onDegraded: () => {
+          this.customRuleFilesDegraded = true;
+        },
+      }),
       experimental: {
         cache_file: {
           enabled: true,
@@ -2168,37 +1774,6 @@ done
   }
 
   /**
-   * 生成日志配置
-   */
-  private generateLogConfig(config: UserConfig): SingBoxLogConfig {
-    // 日志级别由用户配置（默认 info）。level 影响是否记录访问域名/SNI（info/debug 会记，warn+ 不记）。
-    // 隐私模式经 effectiveLogLevel 抬到 ≥warn，从源头不让 sing-box 记录连接明细到 singbox.log。
-    const logConfig: SingBoxLogConfig = {
-      level: effectiveLogLevel(config.logLevel || 'info', this.privacyProvider()),
-      timestamp: true,
-    };
-
-    // 用户关闭日志写盘：整体禁用 sing-box 日志（隐私/省盘），不再写文件
-    if (config.disableLogFile) {
-      logConfig.disabled = true;
-      return logConfig;
-    }
-
-    // 在 TUN 模式下（macOS 和 Windows），使用权限提升运行时无法捕获 stdout
-    // 需要将日志输出到文件，然后通过文件监控读取
-    // 注意：这里直接根据 config 参数判断，而不是 this.currentConfig
-    const isTunMode = config.proxyModeType?.toLowerCase() !== 'systemproxy';
-    const isMacTunMode = process.platform === 'darwin' && isTunMode;
-    const isWindowsTunMode = process.platform === 'win32' && isTunMode;
-
-    if (isMacTunMode || isWindowsTunMode) {
-      logConfig.output = this.getLogFilePath();
-    }
-
-    return logConfig;
-  }
-
-  /**
    * 获取 sing-box 日志文件路径
    */
   private getLogFilePath(): string {
@@ -2223,569 +1798,7 @@ done
     }
   }
 
-  /** 用户自定义国内 DNS 若为 IP（非 DoH 域名），返回 {ip, port} 供 TUN 直连放行/排除集；否则 null。 */
-  private getCustomDomesticDnsEndpoint(config: UserConfig): { ip: string; port: number } | null {
-    const p = parseDnsServerSpec(config.dnsConfig?.domesticDns);
-    return p && !p.isDomain ? { ip: p.server, port: p.port } : null;
-  }
-
-  /**
-   * #57 节点域名解析器档位 → 实际使用的 DNS server tag。
-   * 同时供节点 outbound 的 domain_resolver（ctx='dial'）与节点域名 DNS rule1（ctx='rule'）取值，
-   * 保证 dial 与 rule1 用同一档（outbound 级统一 tag），不破坏 selector 热切换。
-   *
-   * 档位（缺省 / 'auto' = 零行为变化，dial 与 rule1 忠实保留各自基线解析器）：
-   *  - auto   → dial=dns-bootstrap（AliDNS IP-DoH 223.5.5.5，保 outbound.domain_resolver 现状）；
-   *             rule=dns-domestic（doh.pub DoH，保节点域名 DNS rule1 现状）。
-   *             两路径基线本就不同档，统一会把 rule1 从 doh.pub 悄改成 AliDNS，违反「默认零行为变化」。
-   *  - dnspod → dns-node（DNSPod IP-DoH 1.12.12.12）
-   *  - system → dns-local（系统 DNS）；但 INV-1：TUN 下 rule ctx 强制 dns-node（IP-DoH）防递归——
-   *             节点域名查询若落 dns-local，其上游可能再经 TUN 被 hijack-dns 劫持回 DNS rules 形成软死循环。
-   */
-  private getNodeResolverTag(config: UserConfig | null | undefined, ctx: 'dial' | 'rule'): string {
-    const mode = config?.dnsConfig?.nodeDomainResolver ?? 'auto';
-    if (mode === 'dnspod') return 'dns-node';
-    if (mode === 'system') {
-      if (ctx === 'rule' && config?.proxyModeType === 'tun') return 'dns-node'; // INV-1
-      return 'dns-local';
-    }
-    // auto（含缺省）：忠实保留两路径各自基线解析器，逐字节回现状。
-    return ctx === 'dial' ? 'dns-bootstrap' : 'dns-domestic';
-  }
-
-  private generateDnsConfig(config: UserConfig, selectedServerTag: string): SingBoxDnsConfig {
-    const proxyMode = (config.proxyMode || 'smart').toLowerCase();
-
-    // 获取用户 DNS 配置，不存在则使用默认值
-    const userDnsConfig = config.dnsConfig || {
-      domesticDns: 'https://doh.pub/dns-query',
-      foreignDns: 'https://dns.google/dns-query',
-      enableFakeIp: false,
-    };
-
-    // 决定是否开启 FakeIP。
-    // 在 TUN 模式下强制开启 FakeIP。
-    // 原因：很多第三方机场的节点防滥用严格，如果收到纯 IP 地址而非域名，会直接拒绝连接并抛出无效证书或拦截页面。
-    // 配合我们刚刚修复的 macOS gvisor strict_route DHCP DNS 劫持逻辑，
-    // FakeIP 现在能够 100% 完美的用内部 cache 把假 IP 还原成真域名丢给代理节点。
-    // 从而完美避开机场对纯 IP 请求的无情封杀！
-    // 单一真值：与 custom-rule-files.usesFakeIp 同源（外化判定一致，避免漂移）
-    const enableFakeIp = usesFakeIp(config);
-
-    // 用户自定义 DNS：解析 domesticDns/foreignDns（https DoH / tls DoT / udp / 裸 IP），非法或空回退默认并告警。
-    const DEFAULT_DOMESTIC: ParsedDnsServer = {
-      type: 'https',
-      server: 'doh.pub',
-      port: 443,
-      path: '/dns-query',
-      isDomain: true,
-    };
-    const DEFAULT_FOREIGN: ParsedDnsServer = {
-      type: 'https',
-      server: 'dns.google',
-      port: 443,
-      path: '/dns-query',
-      isDomain: true,
-    };
-    const domestic = parseDnsServerSpec(userDnsConfig.domesticDns) ?? DEFAULT_DOMESTIC;
-    const foreign = parseDnsServerSpec(userDnsConfig.foreignDns) ?? DEFAULT_FOREIGN;
-    if (userDnsConfig.domesticDns && !parseDnsServerSpec(userDnsConfig.domesticDns)) {
-      this.logToManager(
-        'warn',
-        `国内 DNS 无法解析，已回退默认 doh.pub: ${userDnsConfig.domesticDns}`
-      );
-    }
-    if (userDnsConfig.foreignDns && !parseDnsServerSpec(userDnsConfig.foreignDns)) {
-      this.logToManager(
-        'warn',
-        `境外 DNS 无法解析，已回退默认 dns.google: ${userDnsConfig.foreignDns}`
-      );
-    }
-    // 由解析结果构造 dns server：域名型 DNS 需 domain_resolver 引导解析；DoH 带 path；remote 走代理 detour。
-    const buildUserDns = (tag: string, p: ParsedDnsServer, detour?: string): SingBoxDnsServer => ({
-      tag,
-      type: p.type,
-      server: p.server,
-      server_port: p.port,
-      ...(p.type === 'https' ? { path: p.path || '/dns-query' } : {}),
-      ...(p.isDomain ? { domain_resolver: 'dns-bootstrap' } : {}),
-      ...(detour ? { detour } : {}),
-    });
-
-    // sing-box 1.13+ 新格式：每个 server 必须有显式 type 字段
-    //
-    // 关键架构说明：
-    // - 在 TUN 下，Windows 的系统 DNS (svchost) 发出的解析请求会被 TUN 劫持。如果该系统 DNS 配置为公共 IP，
-    //   此时 type: 'local' (调用系统 getaddrinfo) 就会进入死循环。
-    // - 为了彻底解决这个问题，同时避免 UDP 53 屏蔽（之前使用 223.5.5.5 UDP 的缺陷），
-    //   我们引入一个坚不可摧的 DoH IP Bootstrap (dns-bootstrap)：向 223.5.5.5:443 直接发 DoH 包。
-    //   它绕过 TUN 不是靠 DNS detour，而是靠 route 规则把 223.5.5.5:443 直连放行（见 generateRouteConfig）。
-    //   关键路径（节点域名 / doh.pub / default_domain_resolver）均以它为 resolver，免疫 UDP 53 限速/劫持/投毒。
-    const dnsServers: SingBoxDnsServer[] = [
-      {
-        // 引导解析：专门用于解析代理节点的 IP 解析器（UDP，最稳健）
-        tag: 'dns-bootstrap-udp',
-        type: 'udp',
-        server: '223.5.5.5',
-        server_port: 53,
-      },
-      {
-        // 引导解析（DoH over IP）：关键路径的解析器。server 已是 IP，无需 domain_resolver。
-        // 相比 UDP 53，对运营商的 UDP 53 限速/劫持/投毒免疫，避免节点域名解析失败导致断流。
-        tag: 'dns-bootstrap',
-        type: 'https',
-        server: '223.5.5.5',
-        server_port: 443,
-        path: '/dns-query',
-      },
-      {
-        // 节点域名解析器可选档（#57，DNSPod IP-DoH）：向 1.12.12.12:443 直接发 DoH 包。
-        // 与 dns-bootstrap 同为 IP-based DoH（免疫 UDP 53 限速/劫持）、靠 route 把 1.12.12.12:443 直连放行绕过 TUN。
-        // 仅当用户把节点域名解析器切到「DNSPod」或 system 档（TUN 下 rule ctx 防递归）时被引用；
-        // 恒加进 servers——未被引用的 server sing-box 不会主动连接，零成本，避免按档增删 server 破坏热切换 tag 稳定性。
-        tag: 'dns-node',
-        type: 'https',
-        server: '1.12.12.12',
-        server_port: 443,
-        path: '/dns-query',
-      },
-      {
-        // 兼容性和兜底的系统 DNS
-        tag: 'dns-local',
-        type: 'local',
-      },
-      // 国内直连 DNS（用户可自定义，默认 doh.pub DoH）
-      buildUserDns('dns-domestic', domestic),
-      // 远程代理 DNS（用户可自定义，默认 dns.google DoH）。必须走代理 detour，
-      // 否则在境内直接发起会因 GFW 拦截/污染导致 FakeIP 映射失败或 TTL 极短产生大量无效解析。
-      buildUserDns('dns-remote', foreign, selectedServerTag),
-    ];
-
-    if (enableFakeIp) {
-      dnsServers.push({
-        // FakeIP 服务器：返回虚假 IP，由 sniff 识别真实域名
-        tag: 'fakeip',
-        type: 'fakeip',
-        inet4_range: '198.18.0.0/15',
-        // §B：关 IPv6(enableIPv6=false)时不给 FakeIP 分配 v6 段 → fakeip 对 AAAA 无段可分配即返空（真机实测，见
-        // docs/design E-1：无需改规则 query_type）；配合下方 strategy=ipv4_only → 客户端拿不到任何 v6，杜绝
-        // "双栈站 + 节点无 v6"时浏览器试 v6 失败致 ERR_CONNECTION_CLOSED。
-        ...(config.enableIPv6 ? { inet6_range: 'fc00::/18' } : {}),
-      });
-    }
-
-    // 方案B：DNS 接管把系统 DNS 改成公网 8.8.8.8 后，dns-local(type:local) 解不了内网域名(.lan/.home.arpa/内网域)、
-    // 反查(.arpa) 与 captive portal。接管激活且读到内网 LAN 解析器(私网 IPv4)时，建 dns-lan(udp:53 指向它)，把这些
-    // 域名重定向到它解析；其 :53 由 generateRouteConfig rule C 直连放行(防被 hijack-dns 抢走成环)。
-    const lanResolver = this.lanResolverForDns;
-    if (lanResolver) {
-      dnsServers.push({ tag: 'dns-lan', type: 'udp', server: lanResolver, server_port: 53 });
-    }
-    // Q1 死循环防护（**仅 Windows**）：Win TUN 的 strict_route(WFP) 把所有 :53 逼进 TUN；type:local 经 svchost(Win DNS
-    // 客户端，不在 route 直连进程白名单)发出的查询 → 进 TUN → hijack-dns → 命中 dns-local 规则 → 又 svchost → ∞。
-    // **死环源于 strict_route + type:local 本身，与系统 DNS 是否被改无关**（Win setDns 已收敛 no-op，见 SystemDnsManager）。
-    // macOS 安全(dns-local→mDNSResponder，在直连白名单，查询不被 hijack)；Linux 不接管。故仅 Win 下把会落 type:local 的
-    // unicast 域名改路由：银行公网→dns-domestic、内网/反查/captive→dns-lan(有 LAN 解析器)否则 dns-domestic。
-    // .local 走 mDNS/LLMNR 组播(不发 unicast)，留 dns-local。
-    // 注：winLoopRisk 暂以 takeoverSystemDns 开关为代理键(默认 on)；理想是解耦为「Win+TUN」恒判(死环与开关无关)，
-    // 待真机 nslookup 实证后再改（见 docs/design/dns-ipv6-takeover.md §A）。
-    const dnsTakeoverActive =
-      config.proxyModeType === 'tun' && config.dnsConfig?.takeoverSystemDns !== false;
-    const winLoopRisk = process.platform === 'win32' && dnsTakeoverActive;
-    // 内网/反查/captive 解析器：优先 dns-lan(直连放行的 LAN IP)；无则 Win 退 dns-domestic(避免 type:local 死环、
-    // 内网域名 NXDOMAIN 但不挂)，非 Win 退 dns-local(系统解析器=真实 LAN，正常)。
-    const internalResolverTag = lanResolver
-      ? 'dns-lan'
-      : winLoopRisk
-        ? 'dns-domestic'
-        : 'dns-local';
-    // 银行/U盾公网域名解析器：仅 Win 改 dns-domestic 绕死环；其余 dns-local（mac 经 mDNSResponder 直连解公网正常）。
-    const bankResolverTag = winLoopRisk ? 'dns-domestic' : 'dns-local';
-
-    const dnsConfig: SingBoxDnsConfig = {
-      servers: dnsServers,
-      rules: [],
-      // 默认使用国内 DNS 解析
-      final: 'dns-domestic',
-      // §B strategy 随 enableIPv6 收敛:
-      //  - 开 IPv6 → prefer_ipv4(允许 AAAA,IPv6-only 目标/CDN v6 优选可达;有 v6 时 v6 可能被系统选用)。
-      //  - 关 IPv6 → ipv4_only(抑制 AAAA,客户端只拿 A)→ 杜绝"双栈站 + 节点无 v6"时浏览器试 v6 失败致
-      //    ERR_CONNECTION_CLOSED(#57:TUN 本地握手骗过 Happy Eyeballs、不回落 v4)。关 IPv6 即明示放弃 v6,
-      //    IPv6-only 目标不可达是该选择的预期代价(故不再有当初移除 ipv4_only 的顾虑——那是 v6 开着时的副作用)。
-      strategy: config.enableIPv6 ? 'prefer_ipv4' : 'ipv4_only',
-      // 关 FakeIP：补 reverse_mapping，让无 SNI/ECH 的连接也能用 DNS 反查域名做路由匹配（不改节点收 IP 事实，见设计 T3）。
-      // 开 FakeIP 时不加（FakeIP 本身已提供域名↔假 IP 的双向映射）。
-      ...(enableFakeIp ? {} : { reverse_mapping: true }),
-    };
-    const dnsRules: SingBoxDnsRule[] = [];
-
-    // 代理服务器域名必须使用真实 DNS 解析（避免 FakeIP 劫持产生死循环）。
-    // #57 rule1 全量化：遍历【全部】节点的 address + tlsSettings.serverName（过滤 IP 字面量），而非仅当前选中节点——
-    //   修热切换 query 侧缺口：原先只含 selectedServer，热切换到其它节点后该节点域名查询会落 FakeIP/dns-local。
-    // 去掉过宽的 domain_keyword（exact domain + suffix 已对节点域名全覆盖，keyword 易误伤同名子串）。
-    // server 取 getNodeResolverTag(config,'rule')：auto 忠实回 dns-domestic（doh.pub DoH，与基线一致，
-    //   不随 dial 的 dns-bootstrap 统一）；dnspod=dns-node；TUN+system 强制 dns-node 防递归。
-    const nodeDomains = Array.from(
-      new Set(
-        config.servers.flatMap((s) => {
-          const ds: string[] = [];
-          if (s.address) ds.push(s.address);
-          if (s.tlsSettings?.serverName) ds.push(s.tlsSettings.serverName);
-          return ds;
-        })
-      )
-    ).filter((d) => !!d && !isIpv4Host(d) && !isIpv6Host(d));
-    if (nodeDomains.length) {
-      // 观测：超大订阅下 rule1 域名规模（不设上限，trie 匹配千级无压力；异常膨胀时据此定位）。
-      this.logToManager('info', `DNS rule1 节点域名规则: ${nodeDomains.length} 个域名`);
-      dnsRules.push({
-        domain: nodeDomains,
-        domain_suffix: nodeDomains.flatMap((d) => [d, `.${d}`]),
-        server: this.getNodeResolverTag(config, 'rule'),
-      } as SingBoxDnsRule);
-    }
-
-    // 处理基础 DNS 服务的地址解析，确保它们走引导解析器（含用户自定义的 DoH 域名）
-    const bootstrapDomains = ['doh.pub', 'dns.google', 'cloudflare-dns.com', 'one.one.one.one'];
-    if (domestic.isDomain) bootstrapDomains.push(domestic.server);
-    if (foreign.isDomain) bootstrapDomains.push(foreign.server);
-    dnsRules.push({
-      domain: Array.from(new Set(bootstrapDomains)),
-      server: 'dns-bootstrap-udp',
-    } as SingBoxDnsRule);
-
-    // 解决 mDNS 和本地反向解析导致的 context deadline exceeded 超时问题。
-    // 方案B + Q1：内网/反查 → internalResolverTag、银行 → bankResolverTag、.local 恒 dns-local（组播安全）。
-    // 三者全为 dns-local（即非 dns-lan、非 Win+takeover）→ 合并回原单条规则，byte-diff 零变化（Linux baseline/系统代理/
-    // 关接管均走此路）。否则拆三条（mac/Win 有 LAN 解析器，或 Win+takeover 无 LAN 解析器）。
-    if (internalResolverTag === 'dns-local' && bankResolverTag === 'dns-local') {
-      dnsRules.push({
-        domain_suffix: [
-          '.local',
-          '.arpa',
-          '.lan',
-          '.home.arpa',
-          ...DOMESTIC_BANK_AND_STOCK_DOMAINS,
-        ],
-        server: 'dns-local',
-      } as SingBoxDnsRule);
-    } else {
-      dnsRules.push({ domain_suffix: ['.local'], server: 'dns-local' } as SingBoxDnsRule);
-      dnsRules.push({
-        domain_suffix: [...DOMESTIC_BANK_AND_STOCK_DOMAINS],
-        server: bankResolverTag,
-      } as SingBoxDnsRule);
-      dnsRules.push({
-        domain_suffix: ['.arpa', '.lan', '.home.arpa'],
-        server: internalResolverTag,
-      } as SingBoxDnsRule);
-    }
-
-    // fake-ip-filter 默认清单（仅 FakeIP 开启时有意义；config.fakeIpFilter === false 可关）：NTP/STUN/Captive 等
-    // 在 FakeIP 下会坏的域名 → 在 fakeip catch-all 之前命中、改走真实解析器，绕过 FakeIP。内联硬编码、无下载依赖。
-    if (enableFakeIp && config.fakeIpFilter !== false) {
-      // 连通性探测 / Captive Portal → internalResolverTag（需接入网/LAN 解析器反映真实本地网络，否则系统误判断网 /
-      // 锁屏登录卡死）：有 LAN 解析器→dns-lan；Win+takeover 无 LAN→dns-domestic(避死环,captive 检测降级但不挂);否则 dns-local。
-      dnsRules.push({
-        domain: FAKEIP_FILTER_CAPTIVE_DOMAINS,
-        server: internalResolverTag,
-      } as SingBoxDnsRule);
-      // NTP / STUN / TURN → dns-domestic（真实 DoH；校时与 P2P 需真实 IP）。同规则内 domain_suffix/domain_keyword 为 OR。
-      dnsRules.push({
-        domain_suffix: FAKEIP_FILTER_NTP_SUFFIXES.flatMap((d) => [d, `.${d}`]),
-        domain_keyword: FAKEIP_FILTER_NTP_STUN_KEYWORDS,
-        server: 'dns-domestic',
-      } as SingBoxDnsRule);
-    }
-
-    // 处理自定义规则中的 bypassFakeIP（仅 domain / domainSuffix / domainKeyword 三类域名规则有效）。
-    // 可外化规则（smart/global + 已落盘 <base>.dns.json）→ 引用其 <base>-dns rule_set，值不进 DNS 配置
-    // （改值原子替换文件 → fswatch 热重载、零重启）；inline / direct 模式 / 文件缺失降级 → 仍按值提取合并（现状）。
-    // 用户路由仅 smart 生效（effectiveCustomRules）：global/direct 下 route 侧不 emit 自定义规则，其 bypassFakeIP DNS 也不应注入
-    const dnsCustomRules = this.effectiveCustomRules(config);
-    if (dnsCustomRules.length && enableFakeIp) {
-      const bypassDomains: string[] = []; // type 'domain'
-      const bypassSuffixes: string[] = []; // type 'domainSuffix'
-      const bypassKeywords: string[] = []; // type 'domainKeyword'
-      const dnsTags: string[] = [];
-      const externalize = proxyMode !== 'direct'; // direct 不外化（route 侧 generateCustomRules 不执行）；此处 dnsCustomRules 已 smart-only
-      for (const rule of dnsCustomRules) {
-        if (!rule.enabled || !rule.bypassFakeIP) continue;
-        if (externalize) {
-          const plan = planCustomRule(rule);
-          if (plan.kind !== 'inline' && plan.dnsRules) {
-            const base = customRuleFileBase(rule.id);
-            const dnsPath = path.join(getCustomRulesDir(), `${base}.dns.json`);
-            if (require('fs').existsSync(dnsPath)) {
-              if (!dnsTags.includes(`${base}-dns`)) dnsTags.push(`${base}-dns`);
-              continue; // 已外化：域名值走文件，不在此提取
-            }
-          }
-        }
-        // inline / direct / 文件缺失降级：取所有 domain/domainSuffix/domainKeyword 条件的值并集
-        for (const cond of ruleConditions(rule)) {
-          const vals = (cond.values || []).map((v) => v.trim()).filter(Boolean);
-          if (vals.length === 0) continue;
-          if (cond.type === 'domain') {
-            bypassDomains.push(...vals);
-          } else if (cond.type === 'domainSuffix') {
-            bypassSuffixes.push(...vals.map((d) => (d.startsWith('*.') ? d.slice(2) : d)));
-          } else if (cond.type === 'domainKeyword') {
-            bypassKeywords.push(...vals);
-          }
-        }
-      }
-
-      if (bypassDomains.length || bypassSuffixes.length || bypassKeywords.length) {
-        const bypassRule: Record<string, unknown> = { server: 'dns-bootstrap' };
-        if (bypassDomains.length) bypassRule.domain = bypassDomains;
-        if (bypassSuffixes.length) {
-          bypassRule.domain_suffix = bypassSuffixes.flatMap((d) => [d, `.${d}`]);
-        }
-        if (bypassKeywords.length) bypassRule.domain_keyword = bypassKeywords;
-        dnsRules.push(bypassRule as unknown as SingBoxDnsRule);
-      }
-      // 外化规则的域名匹配走 rule_set 引用（与上面 inline 合并规则相邻，OR 语义等价）
-      if (dnsTags.length) {
-        dnsRules.push({ rule_set: dnsTags, server: 'dns-bootstrap' } as unknown as SingBoxDnsRule);
-      }
-    }
-
-    // 智能分流/全局代理模式下的 DNS 规则
-    if (proxyMode === 'smart' || proxyMode === 'global') {
-      if (enableFakeIp) {
-        // [原版 Fork 核心精髓：Clash-style 全局 FakeIP]
-        // 让所有的 A/AAAA（IPv4/IPv6）解析无脑走 FakeIP 返回 198.18 的伪装 IP。
-        // 等浏览器连过来以后，Sing-box 靠伪装 IP 查缓存恢复域名，然后交给下面的 Route 引擎。
-        // Route 引擎看到域名，如果命中 geosite-cn，就走 direct 出口，走 direct 时再真正发起本地 DNS 查询拿到淘宝的真实 IP。
-        // 如果查不到 cn 规则，自然落入 proxy，连同域名一起完好无损发给代理节点！极其稳如泰山！
-        dnsRules.push({
-          query_type: ['A', 'AAAA'],
-          server: 'fakeip',
-        } as SingBoxDnsRule);
-      } else {
-        // 如果实在没开 FakeIP（比如系统代理模式），那就用 geosite 规则让它各自拿正确的 IP 吧（但也容易被墙污染）
-        if (proxyMode === 'smart') {
-          dnsRules.push({
-            rule_set: 'geosite-cn',
-            server: 'dns-domestic',
-          } as SingBoxDnsRule);
-
-          // 此处移除了 rule_set: 'geosite-geolocation-!cn'，因为 1.12 的 singbox 在
-          // dns block 里跑规则集会导致某些内置不支持的匹配失效或报错，一律 fallthrough 给 dns-remote
-          dnsRules.push({
-            server: 'dns-remote',
-          } as SingBoxDnsRule);
-        } else {
-          dnsRules.push({
-            query_type: ['A', 'AAAA'],
-            server: 'dns-remote',
-          } as SingBoxDnsRule);
-        }
-      }
-    }
-
-    dnsConfig.rules = dnsRules;
-    return dnsConfig;
-  }
-
-  /**
-   * 生成 AppRule 路由规则
-   */
-
-  /**
-   * 生成 Inbound 配置（sing-box 1.12.x / 1.13.x 兼容格式）
-   */
-  private generateInbounds(
-    config: UserConfig,
-    resolvedIps?: Record<string, string>
-  ): SingBoxInbound[] {
-    const inbounds: SingBoxInbound[] = [];
-
-    // 使用小写比较，兼容 SystemProxy/systemProxy 和 Tun/tun
-    const modeType = (config.proxyModeType || 'systemProxy').toLowerCase();
-
-    const listenAddr = config.allowLan ? '::' : '127.0.0.1';
-
-    // 无论哪种模式，都添加 HTTP + SOCKS inbound
-    // 这样用户在终端配置的代理环境变量在切换模式后仍然可用
-    //
-    // 关键修复：必须启用流量嗅探（sniff），否则 sing-box 无法从 TLS ClientHello 中
-    // 提取域名（SNI），导致路由引擎只看到 IP 地址，无法匹配 geosite 规则正确分流。
-    // 症状：Instagram 消息中心无网络、WhatsApp 二维码无法扫码等 WebSocket 类应用异常。
-    // NekoBox 等 sing-box 客户端默认开启 sniff，FlowZ 之前遗漏了。
-    //
-    // 版本兼容：
-    //   1.12.x → sniff/sniff_override_destination 是 inbound 级别字段
-    //   1.13.x → 这两个字段均已移除。sniff（嗅出域名用于路由匹配）由路由层只 push {action:'sniff'} 替代；
-    //            sniff_override_destination（改写 outbound 目标让节点收到域名）在 1.13.0 已移除且无替代
-    //            （详见 generateRouteConfig A. 嗅探规则段注释）。
-    const useLegacySniff = !coreVersionAtLeast(this.coreVersion, 1, 13);
-
-    // mixed-only：单个 mixed inbound 同口服务 HTTP + SOCKS（取代原 http-in + socks-in + 可选 mixed-in）。
-    // 要 SOCKS 的 app 指同一端口即可。端口取单一真值 localProxyPort（mixedPort，旧配置回退 httpPort）。
-    const mixedInbound: SingBoxInbound = {
-      type: 'mixed',
-      tag: 'mixed-in',
-      listen: listenAddr,
-      listen_port: localProxyPort(config),
-    };
-    if (useLegacySniff) {
-      mixedInbound.sniff = true;
-      mixedInbound.sniff_override_destination = true;
-    }
-    inbounds.push(mixedInbound);
-
-    // 出口 IP 探针 inbound（仅本地回环，端口动态分配）：经 probe-direct-in 的请求由 route.rules 头部
-    // 钉死走 direct 出站、经 probe-proxy-in 钉死走 proxy-selector，从而无论接管/分流模式都能测出真实出口
-    // IP。loopback 不进 TUN，无回环风险。分配失败（probe*Port 为 null）则不注入，IP 卡显示「获取失败」。
-    if (this.probeDirectPort && this.probeProxyPort) {
-      inbounds.push(
-        {
-          type: 'http',
-          tag: 'probe-direct-in',
-          listen: '127.0.0.1',
-          listen_port: this.probeDirectPort,
-        },
-        {
-          type: 'http',
-          tag: 'probe-proxy-in',
-          listen: '127.0.0.1',
-          listen_port: this.probeProxyPort,
-        }
-      );
-    }
-
-    // TUN 模式额外添加 TUN inbound
-    if (modeType === 'tun') {
-      const shouldBypassLAN = config.bypassLAN !== false; // 默认为 true
-      // 恢复 3.3.18 能完美工作的排除列表。
-      // 注意：macOS 下绝对不能在底层排除物理局域网段，否则 macOS NetworkExtension 的路由逆向拦截机制会导致从 TUN (172.19.0.1) 发回 192.168.x.x 的 TCP 回执包被当作非法源 IP 丢弃，导致网页无限 HANG。
-      // 但是在 Windows 下，Wintun 如果不排除局域网物理网关，发往本地路由器的 DHCP/网关查询会被死循环拦截，导致全局断网。
-      const excludeAddr =
-        process.platform === 'win32' && shouldBypassLAN
-          ? [...PRIVATE_IP_CIDRS]
-          : ['127.0.0.0/8', '::1/128'];
-      // 【已知限制 / Windows 真机待验】Windows+bypassLAN 下这里用宽私网段(10/8、192.168/16 等)整体排除出 TUN，
-      // 会顺带把落在私网段内的 endpoint(WG/Tailscale) force-route 段(如 mesh 192.168.50.0/24)也排除 → 该段到不了
-      // 组网节点（走物理 LAN/直连）。mac/Linux 不排除私网（gvisor/系统栈走路由规则），force-route 正常生效。
-      // 旧版曾用 route_address 把 mesh 段以更具体前缀抢回 TUN，但该机制本身从未在 Windows 真机验证（且 sing-box
-      // route_address 非空即替换默认 0/0，处置不当会反而破坏 Windows 全局代理）→ 故此处不再投机重加，留待 Windows
-      // 真机抓包定论后于专项分支处理（tailnet 100.64.0.0/10 不在私网排除表，Windows 下本就可达，不受此限制影响）。
-
-      // Windows 下额外排除核心 DNS IP，防止 WFP 进程匹配失效时产生回流死循环
-      if (process.platform === 'win32') {
-        excludeAddr.push(
-          '223.5.5.5/32',
-          '223.6.6.6/32',
-          '1.12.12.12/32', // #57 DNSPod IP-DoH（节点域名解析器 DNSPod 档）：同 223.5.5.5 须排除防回流死循环
-          '119.29.29.29/32',
-          '119.28.28.28/32',
-          '114.114.114.114/32',
-          '8.8.8.8/32',
-          '1.1.1.1/32'
-        );
-        // 用户自定义的国内 DNS（IP 型）一并排除，防 WFP 进程匹配失效时回流死循环
-        const customDns = this.getCustomDomesticDnsEndpoint(config);
-        if (customDns) {
-          excludeAddr.push(`${customDns.ip}/${isIpv6Host(customDns.ip) ? 128 : 32}`);
-        }
-      }
-
-      // 绝杀级修复（多服务器版本）：如果在 应用分流 (App Policy) 中选择了其他节点，那么这些节点的 IP 也必须被排除。
-      // 否则，FlowZ 去连接这些次选节点的流量也会回流进入 TUN 产生死循环。
-      const allServerIds = new Set([
-        config.selectedServerId as string,
-        ...this.effectiveAppRules(config).map((r) => r.targetServerId),
-      ]);
-
-      // 去除会导致 macOS 崩溃的 shouldBypassLAN 全局排除逻辑，回到 3.3.18 时代的精简状态
-      for (const serverId of allServerIds) {
-        if (!serverId) continue;
-        const server = config.servers.find((s) => s.id === serverId);
-        if (server?.address) {
-          if (isIpv4Host(server.address)) {
-            excludeAddr.push(`${server.address}/32`);
-          } else if (isIpv6Host(server.address)) {
-            excludeAddr.push(`${server.address}/128`);
-          } else if (resolvedIps && resolvedIps[serverId]) {
-            // 使用预解析的 IP
-            const addr = resolvedIps[serverId];
-            excludeAddr.push(isIpv6Host(addr) ? `${addr}/128` : `${addr}/32`);
-          }
-        }
-      }
-
-      // 恢复至对应平台最稳定的网段。Windows 在 v3.4.0 使用 /16 时非常完美；Mac 在 v3.3.18 使用 /30 时最完美。
-      const tunAddress = [
-        config.tunConfig?.inet4Address ||
-          (process.platform === 'darwin' ? '172.19.0.1/30' : '172.19.0.1/16'),
-      ];
-      // macOS 默认分配 IPv6 以提高与本地网络服务的兼容性，与 3.3.18 保持一致
-      // （此前 darwin / 非 darwin 两分支逻辑完全相同，已合并消重）
-      if (config.enableIPv6) {
-        tunAddress.push(config.tunConfig?.inet6Address || 'fdfe:dcba:9876::1/126');
-      }
-
-      // macOS (3.3.18) 最稳定 MTU 为 1400。Windows (3.4.0) 下 MTU=1350 最完美。
-      // 9000 为历史默认值（UI 不暴露 MTU 设置项），等同"未自定义"，必须回退到平台最优值，
-      // 否则巨型 MTU 会让上面精心调优的平台值成为永不生效的死代码。
-      const platformDefaultMtu = process.platform === 'darwin' ? 1400 : 1350;
-      const userMtu = config.tunConfig?.mtu;
-      const effectiveMtu = !userMtu || userMtu === 9000 ? platformDefaultMtu : userMtu;
-
-      // macOS 必须 gvisor 栈(3.3.18)；'system' 是历史默认值（UI 不暴露 stack 设置项），在 macOS 上
-      // 等同"未自定义"，必须回退到 gvisor，否则同 MTU 一样平台判定成永不生效的死代码。Win/Linux 保持 system。
-      const platformDefaultStack = process.platform === 'darwin' ? 'gvisor' : 'system';
-      const userStack = config.tunConfig?.stack;
-      const effectiveStack =
-        !userStack || (process.platform === 'darwin' && userStack === 'system')
-          ? platformDefaultStack
-          : userStack;
-
-      const tunInbound: SingBoxInbound = {
-        type: 'tun',
-        tag: 'tun-in',
-        address: tunAddress,
-        mtu: effectiveMtu,
-        auto_route: config.tunConfig?.autoRoute ?? true,
-        strict_route: config.tunConfig?.strictRoute ?? true,
-        // macOS 必须使用 gvisor 栈(3.3.18)。Windows 下 system 栈配合 Wintun 性能最强且稳定(3.4.0)。
-        stack: effectiveStack,
-        route_exclude_address: excludeAddr,
-      };
-
-      // 兼容 sing-box 1.12.x 版本（打包核心现已全部 ≥1.13.13，此分支仅为向后兼容旧 userData 核心保留），必须在 inbound 定义 sniff 否则无法域名分流。
-      // 对于 1.13.0+，嗅探逻辑已经统一由后方 route.rules 承担，但在入站开启会报错，因此需精准版本判断。
-      if (!coreVersionAtLeast(this.coreVersion, 1, 13)) {
-        (tunInbound as any).sniff = true;
-      }
-
-      // macOS 平台特定配置
-      if (process.platform === 'darwin') {
-        tunInbound.platform = {
-          http_proxy: {
-            enabled: true,
-            server: '127.0.0.1',
-            server_port: localProxyPort(config), // mixed-only：指向本地 mixed 端口
-          },
-        };
-      }
-
-      inbounds.push(tunInbound);
-    }
-
-    return inbounds;
-  }
-
   /** 节点是否可用：naive 需要 libcronet 核心库，缺库时不可用（会被跳过、分流/选中回退到 selector）。 */
-  private isNodeUsable(server: ServerConfig): boolean {
-    if (server.protocol.toLowerCase() === 'naive' && !resourceManager.hasCronetLib()) {
-      return false;
-    }
-    return true;
-  }
-
   /**
    * naive 节点因缺 libcronet 不可用时的用户可读原因，按平台/真因区分文案：
    * - copy-failed：内置库存在但拷贝到核心目录失败（权限/磁盘/AV），提示重启或修权限，而非误报"无预编译库"
@@ -2803,401 +1816,6 @@ done
     return `选中的节点「${server.name}」是 NaiveProxy，但未找到 libcronet 核心库。请选择其它协议的节点。`;
   }
 
-  private generateOutbounds(
-    selectedServer: ServerConfig,
-    config: UserConfig,
-    idToTagMap: Map<string, string>
-  ): SingBoxOutbound[] {
-    const outbounds: SingBoxOutbound[] = [];
-    // endpoint（WireGuard 等）单独收进 pendingEndpoints（顶层 endpoints[]），但其 tag 仍与节点出站 tag 一同
-    // 进入 nodeTags → proxy-selector / rule-sel / route，与普通节点一视同仁（clash_api 热切换已实测兼容）。
-    this.pendingEndpoints = [];
-    const nodeTags: string[] = [];
-
-    if (config) {
-      // 生成【全部】节点的 Outbound：selector 需要列出所有可切换节点；detour 前置节点亦在 config.servers
-      // 中，一并生成、通过 detour 字段链接。单个节点配置异常不应拖垮整体配置，逐节点 try/catch 跳过。
-      // （app/custom 分流规则指向的固定节点 tag 不变，仍直接命中其节点出站，不经 selector。）
-      for (const server of config.servers) {
-        const tag = idToTagMap.get(server.id) || `proxy-${server.id}`;
-        if (nodeTags.includes(tag)) continue; // 去重（含 endpoint）
-        // 不可用节点跳过：naive 缺 libcronet 时，sing-box 启动会预初始化全部出站、缺库的 naive 会让
-        // 整个代理启动 FATAL（连非 naive 节点也用不了）。跳过后，路由层对该节点 tag 的死引用会在
-        // generateSingBoxConfig 末尾被统一修正为 selector（见 H2 修复）。
-        if (!this.isNodeUsable(server)) {
-          this.logToManager(
-            'warn',
-            `跳过不可用节点「${server.name}」：NaiveProxy 缺少 libcronet 核心库`
-          );
-          continue;
-        }
-        // 启动前配置校验 gate 已标记为非法的节点：跳过、不进 outbounds/selector（防 onRetry 重生成复活）。
-        // 路由层对其 tag 的死引用由 generateSingBoxConfig 末尾 fixRouteDeadReferences 统一修正为 selector。
-        if (this.gateInvalidNodes.has(server.id)) {
-          continue;
-        }
-        // WireGuard：endpoint（非 outbound）。建进 pendingEndpoints，tag 仍入 nodeTags 参与选择器/路由。
-        if (server.protocol.toLowerCase() === 'wireguard') {
-          try {
-            this.pendingEndpoints.push(this.buildWireGuardEndpoint(server, tag));
-            nodeTags.push(tag);
-          } catch (e: any) {
-            this.logToManager(
-              'warn',
-              `生成 WireGuard endpoint 失败，已跳过: ${server.name} (${e?.message ?? e})`
-            );
-          }
-          continue;
-        }
-        // Tailscale：endpoint（账号制 mesh）。就绪门控——非选中且未就绪（无 authKey、无持久 state）不发射，
-        // 避免拖慢启动 + 多个未登录节点登录 URL 刷屏；选中节点即便未就绪也发射（触发交互登录 URL 流）。
-        if (server.protocol.toLowerCase() === 'tailscale') {
-          const ts = server.tailscaleSettings;
-          const ready = !!ts?.authKey?.trim() || this.tailscaleStateExists(server.id);
-          if (server.id !== selectedServer.id && !ready) {
-            this.logToManager(
-              'info',
-              `Tailscale 节点「${server.name}」未就绪(需登录)且非选中，已跳过`
-            );
-            continue;
-          }
-          try {
-            this.pendingEndpoints.push(this.buildTailscaleEndpoint(server, tag));
-            nodeTags.push(tag);
-          } catch (e: any) {
-            this.logToManager(
-              'warn',
-              `生成 Tailscale endpoint 失败，已跳过: ${server.name} (${e?.message ?? e})`
-            );
-          }
-          continue;
-        }
-        // 自定义协议标记为 endpoint 类型（第三方类 wireguard/tailscale 实现）→ 进 endpoints[]；
-        // 否则走下方普通 outbound 路径。tag 由 generateProxyOutbound/此处统一注入。
-        if (server.protocol.toLowerCase() === 'custom' && server.customSettings?.isEndpoint) {
-          const { detour: _d, ...epOutbound } = (server.customSettings.outbound || {}) as Record<
-            string,
-            unknown
-          >;
-          this.pendingEndpoints.push({
-            ...(epOutbound as unknown as SingBoxEndpoint),
-            tag,
-          });
-          nodeTags.push(tag);
-          continue;
-        }
-        try {
-          const ob = this.generateProxyOutbound(
-            server,
-            idToTagMap,
-            this.getNodeResolverTag(config, 'dial')
-          );
-          ob.tag = tag;
-          if (server.detour && config.servers.some((s) => s.id === server.detour)) {
-            // 环检测：沿 detour 链行进，若回到本节点即成环 → 不设 detour，避免 sing-box 报循环引用启动失败
-            const seen = new Set<string>([server.id]);
-            let cur: string | undefined = server.detour;
-            let looped = false;
-            while (cur) {
-              if (seen.has(cur)) {
-                looped = true;
-                break;
-              }
-              seen.add(cur);
-              cur = config.servers.find((s) => s.id === cur)?.detour;
-            }
-            if (looped) {
-              this.logToManager('warn', `检测到代理链成环，已跳过 detour: ${server.name}`);
-            } else {
-              const detourSrv = config.servers.find((s) => s.id === server.detour);
-              if (detourSrv && isEndpointProtocol(detourSrv.protocol)) {
-                // endpoint（WireGuard/Tailscale）不作为 outbound 的前置代理(detour)目标（dialer detour→endpoint 语义
-                // 未定，且本节点会被下方 detour 死引用预校验误剔）。丢弃 detour（本节点仍直连可用）+ 警告。UI 亦不列
-                // endpoint 为 detour 候选（route 规则指向 endpoint 仍合法，仅 detour 不合法——两者语义不同）。
-                this.logToManager(
-                  'warn',
-                  `endpoint 节点（WireGuard/Tailscale）不支持作为前置代理(detour)目标，已忽略「${server.name}」的代理链`
-                );
-              } else {
-                ob.detour = idToTagMap.get(server.detour);
-              }
-            }
-          }
-          outbounds.push(ob);
-          nodeTags.push(tag);
-        } catch (e: any) {
-          this.logToManager(
-            'warn',
-            `生成节点出站失败，已跳过: ${server.name} (${e?.message ?? e})`
-          );
-        }
-      }
-
-      // 全局 TLS 分片（PR-6）：开启后对所有已生成的 TCP-TLS 节点出站切分 ClientHello，抗 SNI-DPI。
-      // 跳过 hy2/tuic（QUIC 内 TLS、无 TCP ClientHello，死配置）与 naive（Cronet 自管 TLS，拒绝
-      // fragment 字段 → 启动 FATAL）。
-      // 机制选型：用 outbound 的 `tls.fragment`（按 TCP 段切分代理自身 ClientHello）而非 route action
-      // `route-options.tls_fragment`——后者作用于被嗅探出的流量（切内层），非代理自身握手入口。默认仅注入
-      // fragment=true，不注入 fragment_fallback_delay（用核心默认 500ms）/record_fragment（保持纯开关）。
-      if (config.tlsFragment) {
-        for (const ob of outbounds) {
-          if (ob.tls && ob.type !== 'hysteria2' && ob.type !== 'tuic' && ob.type !== 'naive') {
-            ob.tls.fragment = true;
-          }
-        }
-      }
-
-      // selector：列出已生成的全部节点 tag，default 指向当前选中节点；interrupt_exist_connections 由用户
-      // 开关决定（默认 false=优雅切换，现有连接保留至自然关闭）。clash_api `PUT /proxies/proxy-selector`
-      // 据此热切换、无需重启 sing-box（详见 switchMode）。路由的 final 与「→代理」规则统一指向本 selector。
-      // nodeTags 已在节点循环中按序累积（含 endpoint tag）；空=所有节点生成失败/不可用。
-      if (nodeTags.length === 0) {
-        throw new Error('没有可用的代理节点出站（所有节点配置生成失败）');
-      }
-      const selectedServerTag = idToTagMap.get(selectedServer.id) || 'proxy';
-      outbounds.push({
-        type: 'selector',
-        tag: 'proxy-selector',
-        outbounds: nodeTags,
-        default: nodeTags.includes(selectedServerTag) ? selectedServerTag : nodeTags[0],
-        interrupt_exist_connections: config.interruptConnectionsOnSwitch === true,
-      });
-
-      // rule-sel：恒优化后「所有 proxy 规则」均生成独立 selector（无论 targetServerId 有无）。
-      //   有 target → default=目标节点（anti-drift：固定规则不经共享 proxy-selector，热切全局节点后不漂走）；
-      //   无 target → default=proxy-selector（跟全局，嵌套）。「节点↔默认」切换 = rule-sel default 变（PUT 热切换），
-      //   非 outbound 结构变（重启）。members 复用 proxy-selector 的可用 nodeTags（已排除 naive 缺库 / gate-invalid /
-      //   detour 死引用剔除的节点）；目标无效/被 gate 剔除/无 target → default 落 proxy-selector 兜底
-      //   （rule-sel 仍建、不 FATAL，与 fixRouteDeadReferences 兜底语义一致）。
-      // 仅 smart 生成：用户路由仅 smart 生效，global/direct 下 effectiveCustomRules/effectiveAppRules 为空 → 无 proxy 规则、无消费者。
-      if ((config.proxyMode || 'smart').toLowerCase() === 'smart') {
-        this.generateRuleSelectors(config, idToTagMap, nodeTags, outbounds);
-      }
-    } else {
-      // Fallback if config is missing (shouldn't happen)
-      outbounds.push(
-        this.generateProxyOutbound(
-          selectedServer,
-          idToTagMap,
-          this.getNodeResolverTag(config, 'dial')
-        )
-      );
-    }
-
-    // 直连出站
-    outbounds.push({
-      type: 'direct',
-      tag: 'direct',
-    });
-
-    // 版本条件：sing-box 1.12.x 需要在 outbound 层面做 override_address
-    // 因为 1.12 的路由规则不支持 override_address 字段（会被静默忽略）。
-    // 1.13+ 已将此功能迁移到路由规则，不需要额外的 outbound。
-    if (!coreVersionAtLeast(this.coreVersion, 1, 13)) {
-      outbounds.push({
-        type: 'direct',
-        tag: 'direct-loopback',
-        override_address: '127.0.0.1',
-      });
-    }
-
-    // 阻断出站
-    outbounds.push({
-      type: 'block',
-      tag: 'block',
-    });
-
-    // Shadow-TLS 后处理：如果主节点或任意辅助节点使用了 Shadow-TLS，
-    // 为每个使用 Shadow-TLS 的节点插入内层 SS outbound
-    const stlsOutbounds: SingBoxOutbound[] = [];
-    for (const ob of outbounds) {
-      // 根据 tag（节点名称）反查对应的 ServerConfig；selector/direct/block 等非节点出站匹配不到 → 跳过
-      const srv = config?.servers.find((s) => idToTagMap.get(s.id) === ob.tag);
-      if (srv?.shadowTlsSettings) {
-        // 创建独立的外层 ShadowTLS outbound
-        const stlsTag = `stls-out-${srv.id}`;
-        const stlsOutbound: SingBoxOutbound = {
-          type: 'shadowtls',
-          tag: stlsTag,
-          server: srv.address,
-          server_port: srv.shadowTlsSettings.port || srv.port,
-          version: 3,
-          password: srv.shadowTlsSettings.password,
-          tls: {
-            enabled: true,
-            server_name: srv.shadowTlsSettings.sni || undefined,
-            utls: {
-              enabled: true,
-              fingerprint: srv.shadowTlsSettings.fingerprint || 'chrome',
-            },
-          },
-        };
-        stlsOutbounds.push(stlsOutbound);
-
-        // 主 outbound (原本的 shadowsocks) 必须作为应用的路由目标
-        // 所以我们保留它为 proxy (shadowsocks)，但将其 detour 指向新增的 shadowtls outbound
-        ob.detour = stlsTag;
-
-        // 当配置了 detour 后，sing-box 通常期望主 outbound 的 server/port 被忽略
-        // 但为了规范，我们可以保留 shadowsocks 的原参数或统一指向实际伪装的地址
-        // 在 ShadowTLS 架构中，外层负责 TLS 握手连接真实服务器地址，内层 SS 则是被保护的流量
-      }
-    }
-    outbounds.push(...stlsOutbounds);
-
-    // detour 引用预校验（补 check 唯一盲区：start-stage 引用解析——check 过、run 时报
-    // `dependency[X] not found for outbound[Y]` FATAL）。任何 outbound.detour 指向「生成集合内不存在的
-    // tag」→ 剔除该引用方（**不删 detour 字段**，与 pruneTagsClosure 同隐私语义；选中节点则 throw），
-    // 从 selector 删成员并记录 gateInvalidNodes；selector 剔空 throw。仅 config 存在时有 selector，需 guard。
-    if (config) {
-      const tagToServerId = (tag: string): string | undefined => {
-        if (tag.startsWith('stls-out-')) return tag.slice('stls-out-'.length);
-        for (const [id, t] of idToTagMap) {
-          if (t === tag) return id;
-        }
-        return undefined;
-      };
-      const selector = outbounds.find((o) => o.tag === 'proxy-selector');
-      const selectedTag = idToTagMap.get(selectedServer.id);
-      let mutated = false;
-      // 反复扫描：剔一个引用方可能让别的 detour 链断裂，收敛到不再有死引用。
-
-      while (true) {
-        const validTags = new Set(outbounds.map((o) => o.tag).filter((t): t is string => !!t));
-        const dead = outbounds.find(
-          (ob) =>
-            ob.detour !== undefined && !validTags.has(ob.detour) && ob.tag !== 'proxy-selector'
-        );
-        if (!dead) break;
-        if (dead.tag === selectedTag) {
-          throw new Error(
-            `选中节点「${dead.tag}」的代理链依赖的前置节点不存在，无法启动，请更换节点后重试`
-          );
-        }
-        // 删该引用方 outbound + selector 成员；记录 gateInvalidNodes。
-        const removedTag = dead.tag;
-        const sid = tagToServerId(removedTag);
-        outbounds.splice(outbounds.indexOf(dead), 1);
-        // 同步剔除所有 selector（proxy-selector + 各 rule-sel）的成员引用：rule-sel 的 members 与
-        // proxy-selector 同源（generateRuleSelectors 复用 nodeTags 快照），节点被剔后须一并清理，否则
-        // rule-sel 引用幽灵 tag → sing-box 启动 FATAL。default 命中被剔 → rule-sel 回 proxy-selector、
-        // proxy-selector 落剩余首节点（MED-1，见 prunedSelectorDefault；不 FATAL）。
-        const allSelectors = outbounds.filter(
-          (o) => o.type === 'selector' && Array.isArray(o.outbounds)
-        );
-        for (const sel of allSelectors) {
-          sel.outbounds = (sel.outbounds as string[]).filter((t) => t !== removedTag);
-          if (sel.default === removedTag) {
-            sel.default = this.prunedSelectorDefault(sel.tag, sel.outbounds as string[]);
-          }
-        }
-        if (sid) {
-          idToTagMap.delete(sid);
-          if (!this.gateInvalidNodes.has(sid)) {
-            this.gateInvalidNodes.set(sid, {
-              id: sid,
-              tag: removedTag,
-              reason: '代理链依赖的前置节点不存在（detour 引用无效）',
-            });
-          }
-        }
-        this.logToManager(
-          'warn',
-          `启动前配置校验：节点「${removedTag}」的 detour 引用无效，已剔除`
-        );
-        mutated = true;
-      }
-      if (
-        mutated &&
-        selector &&
-        Array.isArray(selector.outbounds) &&
-        selector.outbounds.length === 0
-      ) {
-        throw new Error('没有可用的代理节点出站（节点代理链依赖无效）');
-      }
-      // rule-sel 剔空（members 全被 detour 死引用剔除）：删该 selector outbound，对应 route 规则的
-      // outbound（rule-sel-<id>）成死引用 → fixRouteDeadReferences 兜底改写为 proxy-selector，启动不 FATAL。
-      if (mutated) {
-        for (let i = outbounds.length - 1; i >= 0; i--) {
-          const o = outbounds[i];
-          if (
-            o.type === 'selector' &&
-            o.tag !== 'proxy-selector' &&
-            o.tag &&
-            o.tag.startsWith('rule-sel') &&
-            Array.isArray(o.outbounds) &&
-            o.outbounds.length === 0
-          ) {
-            outbounds.splice(i, 1);
-          }
-        }
-      }
-    }
-
-    return outbounds;
-  }
-
-  /**
-   * 为「所有 proxy 的 customRule / appRule」生成独立 rule-sel selector（恒优化：不再限有 targetServerId 者）。
-   * members = 可用 nodeTags（与 proxy-selector 同源，已排除缺库/gate-invalid/detour 死引用节点）+ proxy-selector（嵌套全局）；
-   *   default = 规则目标 tag（有）；无 target / 目标无效/被剔除 → proxy-selector（跟全局嵌套，rule-sel 仍建、不 FATAL）。
-   * selectorTag 命名 `rule-sel-<ruleId>` / `rule-sel-app-<appId>`，与节点名撞车时加后缀去重（防启动 FATAL）。
-   * 同步填充 currentRuleTargetMap 供热切换定位（key=`custom:<id>` / `app:<appId>`）。
-   */
-  private generateRuleSelectors(
-    config: UserConfig,
-    idToTagMap: Map<string, string>,
-    nodeTags: string[],
-    outbounds: SingBoxOutbound[]
-  ): void {
-    const existingTags = new Set(outbounds.map((o) => o.tag).filter((t): t is string => !!t));
-    const interrupt = config.interruptConnectionsOnSwitch === true;
-    // (selectorTag, memberTag) 收集，供 generateSingBoxConfig 末尾回填 currentRuleTargetMap。
-    this.pendingRuleSelectors = [];
-
-    // rule-sel 恒存在（所有 proxy 规则，无论 targetServerId 有无）：members 含 proxy-selector（嵌套全局），
-    // default=target 节点（有）或 proxy-selector（无=跟全局）。使「节点↔默认」= rule-sel default 变化（PUT 热切换），
-    // 非 outbound 结构变（重启）。anti-drift 仍保留：固定节点规则 default=节点，不随全局漂。
-    const emit = (ruleKey: string, selectorTag: string, targetServerId: string | undefined) => {
-      // tag 防撞：节点名可能恰好叫 rule-sel-xxx → 加 (n) 后缀（与 getUniqueTag 同形）
-      let tag = selectorTag;
-      let n = 1;
-      while (existingTags.has(tag)) {
-        tag = `${selectorTag} (${n})`;
-        n++;
-      }
-      existingTags.add(tag);
-      const targetTag = targetServerId ? idToTagMap.get(targetServerId) : undefined;
-      // default：target 有效→节点 tag；target 无/无效→proxy-selector（跟全局，嵌套）
-      const defaultTag = targetTag && nodeTags.includes(targetTag) ? targetTag : 'proxy-selector';
-      outbounds.push({
-        type: 'selector',
-        tag,
-        outbounds: [...nodeTags, 'proxy-selector'],
-        default: defaultTag,
-        interrupt_exist_connections: interrupt,
-      });
-      this.pendingRuleSelectors.push({
-        ruleKey,
-        selectorTag: tag,
-        memberTag: defaultTag,
-        targetServerId,
-      });
-    };
-
-    for (const rule of this.effectiveCustomRules(config)) {
-      if (!rule.enabled) continue;
-      if (rule.action !== 'proxy') continue;
-      emit(`custom:${rule.id}`, `rule-sel-${rule.id}`, rule.targetServerId);
-    }
-    for (const appRule of this.effectiveAppRules(config)) {
-      if (!appRule.enabled) continue;
-      if (appRule.action !== 'proxy') continue;
-      // appRule 按 appId 去重（effectiveAppRules），用 appId 作 key；selectorTag 加 app- 前缀避免与 customRule 混淆。
-      emit(`app:${appRule.appId}`, `rule-sel-app-${appRule.appId}`, appRule.targetServerId);
-    }
-  }
-
   /**
    * 供 SpeedTestService 复用：为单个节点生成**完整出站**（全协议，非仅 hy2/tuic），供临时测速 sing-box 经代理做
    * 真实 urltest（端口通≠代理可用，裸 TCP ping 测不出鉴权/中继失败）。
@@ -3210,7 +1828,7 @@ done
     tag: string
   ): SingBoxOutbound | SingBoxEndpoint | null {
     try {
-      if (!this.isNodeUsable(server)) return null;
+      if (!isNodeUsable(server)) return null;
       // Tailscale：账号制/带认证状态的入网，非即起即测的临时隧道 → 排除测速。
       if (server.protocol.toLowerCase() === 'tailscale') return null;
       // 自定义 endpoint 类型：测速临时配置按 isEndpointProtocol(type) 分流 outbounds/endpoints[]，
@@ -3219,1491 +1837,15 @@ done
         return null;
       // WireGuard：endpoint（非 outbound）。SpeedTestService 据 type 放入测速临时配置的 endpoints[]。
       if (server.protocol.toLowerCase() === 'wireguard') {
-        return this.buildWireGuardEndpoint(server, tag);
+        return buildWireGuardEndpoint(server, tag);
       }
       const map = new Map<string, string>([[server.id, tag]]);
-      const ob = this.generateProxyOutbound(server, map, 'dns-direct');
+      const ob = buildProxyOutbound(server, map, 'dns-direct');
       ob.tag = tag;
       delete ob.detour;
       return ob;
     } catch {
       return null;
-    }
-  }
-
-  /**
-   * 构造 WireGuard endpoint（sing-box 1.11+ 顶层 endpoints[]）。单 peer：peer endpoint 用 server.address/port，
-   * 默认 gVisor 用户态栈（system=false，零提权）。base64/CIDR 字段在 UI 表单/.conf 解析已校验，此处直拼。
-   */
-  private buildWireGuardEndpoint(server: ServerConfig, tag: string): SingBoxEndpoint {
-    const s = server.wireguardSettings;
-    if (!s || !s.privateKey || !s.peerPublicKey || !s.localAddress?.length) {
-      throw new Error('WireGuard 配置缺少 privateKey/peerPublicKey/localAddress');
-    }
-    return {
-      type: 'wireguard',
-      tag,
-      system: false,
-      mtu: s.mtu && s.mtu > 0 ? s.mtu : 1408,
-      address: s.localAddress,
-      private_key: s.privateKey,
-      peers: [
-        {
-          address: server.address,
-          port: server.port,
-          public_key: s.peerPublicKey,
-          allowed_ips: s.allowedIPs?.length ? s.allowedIPs : ['0.0.0.0/0', '::/0'],
-          ...(s.preSharedKey ? { pre_shared_key: s.preSharedKey } : {}),
-          // keepalive 缺省强制 25s：WireGuard 是无连接 UDP，client 多在 NAT 后，无 keepalive 则 NAT 映射超时
-          // → 入向不可达、隧道静默断连。故未显式设置时兜 25s 维持映射，避免断连（用户可在表单调大）。
-          persistent_keepalive_interval:
-            s.persistentKeepalive && s.persistentKeepalive > 0 ? s.persistentKeepalive : 25,
-          ...(s.reserved?.length === 3 ? { reserved: s.reserved } : {}),
-        },
-      ],
-    };
-  }
-
-  /** Tailscale state 目录：`<userData>/tailscale/<serverId>`，跨重启免重认证、删节点时清理。 */
-  private tailscaleStateDir(serverId: string): string {
-    return path.join(getUserDataPath(), 'tailscale', serverId);
-  }
-
-  /** state 目录是否已有持久节点状态（免 authKey 重登录的判据）：目录存在且非空。 */
-  private tailscaleStateExists(serverId: string): boolean {
-    try {
-      // readdirSync 在目录缺失时抛 ENOENT → catch 返 false，无需先 existsSync（省一次 stat）。
-      return require('fs').readdirSync(this.tailscaleStateDir(serverId)).length > 0;
-    } catch {
-      return false;
-    }
-  }
-
-  /**
-   * 构造 Tailscale endpoint（sing-box 1.12+ 顶层 endpoints[]）。账号制 mesh，无 server address/port。
-   * 默认 tsnet 用户态（system_interface 省略=false，零提权）；auth_key 可选（无则核日志出登录 URL）。
-   */
-  private buildTailscaleEndpoint(server: ServerConfig, tag: string): SingBoxEndpoint {
-    const ts = server.tailscaleSettings || {};
-    const stateDir = this.tailscaleStateDir(server.id);
-    try {
-      require('fs').mkdirSync(stateDir, { recursive: true });
-    } catch {
-      /* best-effort；目录建失败时 sing-box 自身会报错 */
-    }
-    const ep: SingBoxEndpoint = { type: 'tailscale', tag, state_directory: stateDir };
-    const authKey = ts.authKey?.trim();
-    if (authKey) ep.auth_key = authKey;
-    const controlUrl = ts.controlUrl?.trim();
-    if (controlUrl) ep.control_url = controlUrl;
-    const hostname = ts.hostname?.trim();
-    if (hostname) ep.hostname = hostname;
-    const exitNode = ts.exitNode?.trim();
-    if (exitNode) {
-      ep.exit_node = exitNode;
-      if (ts.exitNodeAllowLanAccess) ep.exit_node_allow_lan_access = true;
-    }
-    if (ts.acceptRoutes) ep.accept_routes = true;
-    if (ts.ephemeral) ep.ephemeral = true;
-    const adv = (ts.advertiseRoutes || []).map((c) => c.trim()).filter(Boolean);
-    if (adv.length) ep.advertise_routes = adv;
-    // Phase 2：ts.reverseMesh → ep.system_interface=true（+ 选中节点门控 + 强制 helper）。Phase 1 用户态、省略。
-    return ep;
-  }
-
-  /**
-   * 生成代理 Outbound 配置（sing-box 1.12.x / 1.13.x 兼容格式）
-   */
-  private generateProxyOutbound(
-    server: ServerConfig,
-    idToTagMap: Map<string, string>,
-    // #57：节点域名 dial 解析器 tag，由调用方传 getNodeResolverTag(config,'dial')。
-    // 缺省 dns-bootstrap（AliDNS IP-DoH）= 现状，兼容无 config 上下文的兜底调用。
-    nodeResolverTag: string = 'dns-bootstrap'
-  ): SingBoxOutbound {
-    // sing-box 要求协议类型必须是小写
-    const protocol = server.protocol.toLowerCase();
-    const protocolLower = protocol;
-    const tlsProtocols = ['trojan', 'anytls', 'hysteria2', 'tuic'];
-
-    // 自定义协议（raw-JSON 透传）：原样下发用户的 outbound 对象，仅【强制覆盖 tag】（防撞/防注入），
-    // 语义与能否启用由 sing-box check（启动 gate / 表单 probe）判定，FlowZ 不解析 type。
-    if (protocol === 'custom') {
-      // 强制覆盖 tag、剥离 detour：内层 detour 会绕过 FlowZ 的代理链环检测（仅走 server.detour）且 sing-box check
-      // 测不出 detour 环 → 运行时拨号死循环。FlowZ 经 detour 字段统一管理链路，自定义透传不携带内层 detour。
-      const { detour: _drop, ...userOutbound } = (server.customSettings?.outbound || {}) as Record<
-        string,
-        unknown
-      >;
-      return {
-        ...userOutbound,
-        tag: idToTagMap.get(server.id) || `proxy-${server.id}`,
-      } as unknown as SingBoxOutbound;
-    }
-
-    const outbound: SingBoxOutbound = {
-      type: protocol,
-      tag: idToTagMap.get(server.id) || `proxy-${server.id}`,
-      server: server.address,
-      server_port: server.port,
-      // 代理节点域名经引导解析（默认 dns-bootstrap=AliDNS IP-DoH），免疫 UDP 53 限速/劫持，
-      // 避免节点解析失败导致全断流；同时防止 dns-local 死循环导致的连接挂起。
-      // #57：节点域名解析器档位可改为 dns-node(DNSPod)/dns-local(系统 DNS)，dial 与 rule1 同档（统一 tag）。
-      domain_resolver: nodeResolverTag,
-    };
-
-    // vless/vmess UDP 封装：默认 xudp；可经 server.packetEncoding 覆盖（兼容拒收 xudp 的旧核心/服务端，
-    // 否则 UDP 断流且无从调整）。显式设为空串则省略该字段（不下发 packet_encoding，由核心用其默认）。
-    const packetEncoding = server.packetEncoding ?? 'xudp';
-
-    // VLESS 特定配置
-    if (protocol === 'vless') {
-      outbound.uuid = server.uuid;
-      if (server.flow) {
-        outbound.flow = server.flow;
-      }
-      if (packetEncoding) {
-        outbound.packet_encoding = packetEncoding;
-      }
-    }
-
-    // VMess 特定配置
-    if (protocol === 'vmess') {
-      outbound.uuid = server.uuid;
-      outbound.security = server.vmessSecurity || 'auto';
-      outbound.alter_id = server.alterId || 0;
-      if (packetEncoding) {
-        outbound.packet_encoding = packetEncoding;
-      }
-    }
-
-    // Trojan 特定配置
-    if (protocol === 'trojan') {
-      outbound.password = server.password;
-    }
-
-    // Hysteria2 特定配置
-    if (protocol === 'hysteria2') {
-      outbound.password = server.password;
-
-      // 带宽限制
-      if (server.hysteria2Settings?.upMbps) {
-        outbound.up_mbps = server.hysteria2Settings.upMbps;
-      }
-      if (server.hysteria2Settings?.downMbps) {
-        outbound.down_mbps = server.hysteria2Settings.downMbps;
-      }
-
-      // 混淆配置
-      if (server.hysteria2Settings?.obfs?.type && server.hysteria2Settings?.obfs?.password) {
-        outbound.obfs = {
-          type: server.hysteria2Settings.obfs.type,
-          password: server.hysteria2Settings.obfs.password,
-        };
-      }
-
-      // 网络类型 (tcp/udp)
-      if (server.hysteria2Settings?.network) {
-        outbound.network = server.hysteria2Settings.network;
-      }
-    }
-
-    // AnyTLS 特定配置
-    if (protocol === 'anytls') {
-      outbound.password = server.password;
-      // AnyTLS 的 TLS 永远开启，这里不需要额外处理，类型检查结尾部分统一生成
-      // AnyTLS 会话参数
-      if (server.anyTlsSettings?.idleSessionCheckInterval) {
-        outbound.idle_session_check_interval = server.anyTlsSettings.idleSessionCheckInterval;
-      }
-      if (server.anyTlsSettings?.idleSessionTimeout) {
-        outbound.idle_session_timeout = server.anyTlsSettings.idleSessionTimeout;
-      }
-      if (server.anyTlsSettings?.minIdleSession !== undefined) {
-        outbound.min_idle_session = server.anyTlsSettings.minIdleSession;
-      }
-    }
-
-    // Shadowsocks 特定配置
-    if (protocol === 'shadowsocks') {
-      if (!server.shadowsocksSettings) {
-        throw new Error(`Shadowsocks server ${server.name} missing settings`);
-      }
-      outbound.method = server.shadowsocksSettings.method;
-      outbound.password = server.shadowsocksSettings.password;
-      if (server.shadowsocksSettings.plugin) {
-        outbound.plugin = server.shadowsocksSettings.plugin;
-        outbound.plugin_opts = server.shadowsocksSettings.pluginOptions;
-      }
-    }
-
-    // TUIC 特定配置
-    if (server.protocol === 'tuic') {
-      outbound.uuid = server.uuid;
-      outbound.password = server.password;
-
-      if (server.tuicSettings) {
-        if (server.tuicSettings.congestionControl) {
-          outbound.congestion_control = server.tuicSettings.congestionControl;
-        }
-        if (server.tuicSettings.udpRelayMode) {
-          outbound.udp_relay_mode = server.tuicSettings.udpRelayMode;
-        }
-        if (server.tuicSettings.zeroRttHandshake !== undefined) {
-          outbound.zero_rtt_handshake = server.tuicSettings.zeroRttHandshake;
-        }
-        if (server.tuicSettings.heartbeat) {
-          outbound.heartbeat = server.tuicSettings.heartbeat;
-        }
-      }
-    }
-
-    // NaiveProxy 特定配置
-    if (server.protocol === 'naive') {
-      outbound.username = server.username;
-      outbound.password = server.password;
-
-      // NaiveProxy specific configuration
-      // sing-box 的 naive outbound 由 Cronet 自管 TLS，仅支持 server_name / certificate /
-      // certificate_path / ech。下发 alpn 或 insecure:true 会让 sing-box 拒启：
-      //   FATAL: initialize outbound: alpn is not supported on naive outbound
-      //   FATAL: initialize outbound: insecure is not supported on naive outbound
-      // 故这里只下发 server_name（用户在节点上设置的 alpn / allowInsecure 对 naive 无效，忽略）。
-      // 参考：https://sing-box.sagernet.org/configuration/outbound/naive/ （TLS 字段限制）
-      //       SagerNet/sing-box protocol/naive/outbound.go (v1.13.x) ~L47 的 ALPN/insecure 校验
-      outbound.tls = {
-        enabled: true,
-        server_name: server.tlsSettings?.serverName || server.address,
-      };
-
-      // HTTP/3：naive 经 quic:true 走 h3(QUIC/UDP) 拨号传输，对应服务端 `--listen=quic://`。
-      // 注意这只改变"拨号传输"：naive 仍只过 TCP（HTTP CONNECT）、不能中继客户端 UDP（除非
-      // udp_over_tcp，FlowZ 不下发）。客户端 QUIC 若走到 naive，由 blockQuic（udp443 reject 逼回退 TCP）
-      // 或 sing-box 出站层（"UDP is not supported by outbound"）处理；该拨号本身受 fwmark 保护、
-      // 绕过 route 规则，不被 reject 误杀（已实测）。
-      if (server.naiveSettings?.useHttp3) {
-        outbound.quic = true;
-      }
-    }
-
-    // SOCKS 特定配置
-    if (server.protocol === 'socks') {
-      if (server.username) outbound.username = server.username;
-      if (server.password) outbound.password = server.password;
-      // 默认 SOCKS 版本
-      (outbound as any).version = '5';
-    }
-
-    // HTTP 特定配置
-    if (server.protocol === 'http') {
-      if (server.username) outbound.username = server.username;
-      if (server.password) outbound.password = server.password;
-
-      // HTTP outbound headers mapping can be added if needed via server.httpSettings.headers
-      if (server.httpSettings?.headers) {
-        if (!outbound.transport) outbound.transport = { type: 'http' };
-        outbound.transport.headers = server.httpSettings.headers;
-      }
-      if (server.httpSettings?.path) {
-        if (!outbound.transport) outbound.transport = { type: 'http' };
-        outbound.transport.path = server.httpSettings.path;
-      }
-    }
-
-    // SSH 特定配置
-    if (server.protocol === 'ssh') {
-      const ssh = server.sshSettings || {};
-      if (ssh.user) outbound.user = ssh.user;
-      if (ssh.password) outbound.password = ssh.password;
-      if (ssh.privateKey) outbound.private_key = ssh.privateKey;
-      if (ssh.privateKeyPath) outbound.private_key_path = ssh.privateKeyPath;
-      if (ssh.privateKeyPassphrase) outbound.private_key_passphrase = ssh.privateKeyPassphrase;
-      if (ssh.hostKey && ssh.hostKey.length > 0) outbound.host_key = ssh.hostKey;
-      if (ssh.hostKeyAlgorithms && ssh.hostKeyAlgorithms.length > 0)
-        outbound.host_key_algorithms = ssh.hostKeyAlgorithms;
-      if (ssh.clientVersion) outbound.client_version = ssh.clientVersion;
-
-      // SSH outbound 不需要 TLS 和传输层配置，直接返回
-      return outbound;
-    }
-
-    // TLS 配置 (非 Naive 协议，因为 Naive 已在前一段处理了 tls 结构)
-    if (
-      server.protocol !== 'naive' &&
-      (server.security === 'tls' || server.tlsSettings || tlsProtocols.includes(protocol))
-    ) {
-      // 为 Trojan 设置默认 ALPN ["http/1.1"] 以提高兼容性
-      let finalAlpn = server.tlsSettings?.alpn;
-      if (!finalAlpn && protocolLower === 'trojan') {
-        finalAlpn = ['http/1.1'];
-      }
-
-      outbound.tls = {
-        enabled: true,
-        server_name: server.tlsSettings?.serverName || server.address,
-        insecure: server.tlsSettings?.allowInsecure || false,
-        alpn: finalAlpn,
-      };
-
-      // uTLS 仅适用于基于 TCP 的协议，Hysteria2 和 TUIC 使用 QUIC (UDP) 不支持 uTLS
-      const fingerprint = server.tlsSettings?.fingerprint;
-
-      // 默认行为：VLESS 等协议默认开启 chrome 指纹，Trojan 默认不开启（none）以通过标准 TLS 握手
-      let finalFingerprint = fingerprint;
-      if (!finalFingerprint) {
-        if (protocolLower === 'vless' || protocolLower === 'anytls') {
-          finalFingerprint = 'chrome';
-        } else {
-          finalFingerprint = 'none';
-        }
-      }
-
-      if (
-        server.protocol !== 'hysteria2' &&
-        server.protocol !== 'tuic' &&
-        finalFingerprint !== 'none'
-      ) {
-        outbound.tls.utls = {
-          enabled: true,
-          fingerprint: finalFingerprint,
-        };
-      }
-
-      // ALPN 仅在支持的协议上设置
-      if (server.tlsSettings?.alpn) {
-        outbound.tls.alpn = server.tlsSettings.alpn;
-      }
-    }
-
-    // Reality 配置
-    if (server.security === 'reality' && server.realitySettings) {
-      outbound.tls = {
-        enabled: true,
-        server_name: server.tlsSettings?.serverName || undefined,
-        insecure: server.tlsSettings?.allowInsecure || false,
-        utls: {
-          enabled: true,
-          fingerprint: server.tlsSettings?.fingerprint || 'chrome',
-        },
-        reality: {
-          enabled: true,
-          public_key: server.realitySettings.publicKey,
-          short_id: server.realitySettings.shortId || '',
-        },
-      };
-    }
-
-    // 传输层配置（不适用于 hysteria2、anytls、naive）
-    if (
-      server.protocol !== 'hysteria2' &&
-      server.protocol !== 'anytls' &&
-      server.protocol !== 'naive' &&
-      server.network &&
-      server.network !== 'tcp'
-    ) {
-      outbound.transport = this.generateTransportConfig(server);
-    }
-
-    // PR-6 抗封增强（ECH / TLS 分片 / Multiplex / Hy2 端口跳跃）统一后处理
-    this.applyAntiCensorshipOptions(outbound, server);
-
-    return outbound;
-  }
-
-  /**
-   * PR-6 抗封增强统一后处理：ECH、每节点 TLS 分片、Multiplex(reality+vision 跳过)、Hy2 端口跳跃。
-   * 放在 outbound 构建完成后统一处理，避免散落到各协议的 tls/transport 构建点。
-   */
-  private applyAntiCensorshipOptions(outbound: SingBoxOutbound, server: ServerConfig): void {
-    const protocolLower = server.protocol.toLowerCase();
-
-    // fragment 仅对「标准 sing-box TCP-TLS 栈」有意义，以下协议必须排除（否则死配置或直接启动 FATAL）：
-    //   · hy2/tuic：TLS 在 QUIC 内、无 TCP ClientHello（死配置）；
-    //   · naive：TLS 由 Cronet 自管，naive 出站直接拒绝 fragment 字段（实测
-    //     "fragment is not supported on naive outbound" → 启动 FATAL），无论 h2/h3。
-    // 注：ECH 不受此限——QUIC 与 naive(Cronet) 均原生支持 ECH。
-    const fragmentUnsupported =
-      protocolLower === 'hysteria2' || protocolLower === 'tuic' || protocolLower === 'naive';
-
-    // ECH（隐藏 SNI）+ 每节点 TLS 分片（抗 SNI-DPI）：需已有 tls 块
-    if (outbound.tls) {
-      if (server.tlsSettings?.ech) {
-        // 可选 ECHConfigList：填了下发 tls.ech.config（PEM 按行拆数组）；留空则 sing-box 从 DNS(HTTPS RR) 自取。
-        const echCfg = server.tlsSettings.echConfig?.trim();
-        const echLines = echCfg
-          ? echCfg
-              .split(/\r?\n/)
-              .map((l) => l.trim())
-              .filter(Boolean)
-          : [];
-        outbound.tls.ech =
-          echLines.length > 0 ? { enabled: true, config: echLines } : { enabled: true };
-      }
-      if (server.tlsSettings?.fragment && !fragmentUnsupported) outbound.tls.fragment = true;
-    }
-
-    // Multiplex（vless/trojan/vmess/shadowsocks）；vision flow(xtls-rprx-vision) 自带流分帧、与 mux
-    // 不兼容（与是否 reality 无关，普通 TLS+vision 同样不兼容）→ 跳过
-    const mux = server.multiplexSettings;
-    const hasVisionFlow = (server.flow || '').toLowerCase().includes('vision');
-    if (
-      mux?.enabled &&
-      ['vless', 'trojan', 'vmess', 'shadowsocks'].includes(protocolLower) &&
-      !hasVisionFlow
-    ) {
-      outbound.multiplex = {
-        enabled: true,
-        protocol: mux.protocol || 'h2mux',
-        ...(mux.maxConnections ? { max_connections: mux.maxConnections } : {}),
-        ...(mux.minStreams ? { min_streams: mux.minStreams } : {}),
-        ...(mux.padding ? { padding: true } : {}),
-      };
-    }
-
-    // Hysteria2 端口跳跃（serverPorts 为逗号分隔的范围串，如 "20000:30000"，支持多段）
-    if (protocolLower === 'hysteria2' && server.hysteria2Settings?.serverPorts) {
-      const ports = server.hysteria2Settings.serverPorts
-        .split(',')
-        .map((s) => s.trim())
-        .filter(Boolean);
-      if (ports.length > 0) {
-        outbound.server_ports = ports;
-        if (server.hysteria2Settings.hopInterval) {
-          outbound.hop_interval = server.hysteria2Settings.hopInterval;
-        }
-      }
-    }
-  }
-
-  /**
-   * 生成传输层配置
-   */
-  private generateTransportConfig(server: ServerConfig): SingBoxOutbound['transport'] {
-    if (server.network === 'ws' && server.wsSettings) {
-      // 0-RTT early-data：订阅/分享链解析时已存入 wsSettings，此前未落运行时配置导致静默失效
-      return {
-        type: 'ws',
-        path: server.wsSettings.path || '/',
-        headers: server.wsSettings.headers,
-        max_early_data: server.wsSettings.maxEarlyData,
-        early_data_header_name: server.wsSettings.earlyDataHeaderName,
-      };
-    }
-
-    if (server.network === 'grpc' && server.grpcSettings) {
-      return {
-        type: 'grpc',
-        service_name: server.grpcSettings.serviceName || '',
-      };
-    }
-
-    // httpupgrade：较 ws 更隐蔽的 HTTP Upgrade 传输（复用 ws 的 path / Host）
-    if (server.network === 'httpupgrade') {
-      return {
-        type: 'httpupgrade',
-        path: server.wsSettings?.path || '/',
-        host: server.wsSettings?.headers?.['Host'] || server.tlsSettings?.serverName,
-      };
-    }
-
-    return undefined;
-  }
-
-  /**
-   * 生成路由配置（sing-box 1.12.x / 1.13.x 兼容格式）
-   */
-  /** 本地 geo 规则集运行时目录（内置 .srs 拷贝落地处）。copy 与 route 生成共用，单一真值（委托共享模块）。 */
-  private getRuleSetRuntimeDir(): string {
-    return getRuntimeRulesDir();
-  }
-
-  /**
-   * 内置 geo 规则集的单一真值表：tag、运行时文件名、内置源路径。
-   * copyRuleSetsToUserData（写入）、generateRouteConfig（引用 path）、RuleResourceManager（页面展示/更新）
-   * 共用 builtin-geo-rulesets 模块的 BUILTIN_GEO_RULESETS，避免多处硬编码目录+文件名导致漂移——
-   * 改一处而另一处仍指旧路径会让 sing-box 加载本地 rule_set 失败。
-   */
-  private getLocalGeoRuleSets(): { tag: string; fileName: string; srcPath: string }[] {
-    return BUILTIN_GEO_RULESETS.map((b) => ({
-      tag: b.tag,
-      fileName: b.fileName,
-      srcPath: b.bundledPath(),
-    }));
-  }
-
-  /**
-   * 用户路由规则 gate（单一真值）：自定义路由规则**仅 smart 模式生效**。
-   * global = 真·全局（一律走选中节点，忽略用户分流，仅保留功能性强制直连：防环 / LAN·私网 / 网银 U盾 / 节点排除）；
-   * direct = 全部直连。对齐业内（Clash/Surge/sing-box GUI）——全局模式即忽略所有用户分流规则。
-   * 消费点统一经此（route emit / 规则选择器 / geo 收集 / DNS bypassFakeIP），与 effectiveAppRules 对称。
-   */
-  private effectiveCustomRules(config: UserConfig): import('../../shared/types').Rule[] {
-    if ((config.proxyMode || 'smart').toLowerCase() !== 'smart') return [];
-    return config.customRules || [];
-  }
-
-  /**
-   * 应用分流总开关 gate：appRoutingEnabled !== true → 空（默认关）；非 smart 模式（global/direct）→ 空。
-   * 仅 smart + 显式启用才生效。单一真值点，4 个消费点（route 生成 / 规则选择器 / TUN 排除 / geo 收集）统一经此。
-   * global 忽略应用分流（真·全局走选中节点），与 effectiveCustomRules 对称。
-   */
-  private effectiveAppRules(config: UserConfig): import('../../shared/types').AppRule[] {
-    if (config.appRoutingEnabled !== true) return [];
-    if ((config.proxyMode || 'smart').toLowerCase() !== 'smart') return [];
-    return config.appRules || [];
-  }
-
-  private generateRouteConfig(
-    config: UserConfig,
-    idToTagMap: Map<string, string>
-  ): SingBoxRouteConfig {
-    const rules: SingBoxRouteRule[] = [];
-    const proxyMode = (config.proxyMode || 'smart').toLowerCase();
-
-    // 主代理出站统一走 selector(proxy-selector)：clash_api 热切换即改 selector 指向、路由无需重生成。
-    // 具体 targetServerId 的 app/custom 分流在各自逻辑里直指节点 tag，不经此变量。
-    const selectedServerTag = 'proxy-selector';
-
-    // blockQuic（节点无关）：开启时对"将走代理"的 QUIC(UDP443) 执行 reject，逼浏览器回退 TCP。
-    // 「禁 QUIC」即禁 QUIC，与选中节点的协议/中继能力无关，对所有节点一视同仁。两点实测保证安全：
-    //   · 节点自身的 UDP 拨号(naive-h3/hy2/tuic dial server)无害——拨号是 sing-box 进程自有 socket，
-    //     受 fwmark/auto_detect_interface 保护、绕过 route 规则；netns TUN 抓包实测：带 reject udp443
-    //     时 hy2 拨号包仍正常逸出（证伪旧假设"reject 经 strict_route 回流误杀拨号"）。
-    //   · 不下发"全 UDP reject"——只禁 QUIC(443)。非 QUIC 的代理向 UDP 若节点不能中继(naive/ssh/http)，
-    //     由 sing-box 出站层自动拒绝（实测日志 "UDP is not supported by outbound"，不漏 direct、不黑洞），
-    //     无需路由层按节点固化。这也使路由配置与选中节点解耦 → 支持 selector 跨协议无缝热切换。
-    // 节点无关：只要开了 blockQuic 且存在代理路径（非 direct 模式、有节点）就拦——不依赖 selectedServer
-    // 解析成功（避免 selectedServerId 失效但 selector default 仍出流量时 QUIC 漏过）。
-    const blockProxyQuic =
-      config.blockQuic === true && proxyMode !== 'direct' && config.servers.length > 0;
-
-    // 给定域名匹配器，返回应配对的 udp443 reject 规则（smart 模式放在每条 →代理 规则之前），否则 null。
-    const proxyUdpRejectFor = (matcher: Record<string, unknown>): SingBoxRouteRule | null =>
-      blockProxyQuic ? udp443RejectRule(matcher) : null;
-
-    // A. 嗅探规则（必须在前，用于识别域名）
-    // 1.13+ 必须在路由层开启 sniff，替代已移除的 inbound 级别 sniff 字段。
-    // 注意（旧注释「等效 sniff_override_destination」不准确）：sniff 只把嗅出的域名用于【路由匹配】这半边——
-    // sniff_override_destination（改写 outbound 目标，让节点收到域名）在 1.13.0 已移除且无替代。
-    // 故关 FakeIP 时节点仍收真实 IP，域名交付节点只能靠 FakeIP（见 generateDnsConfig / 设计 T1·T4）。
-    if (coreVersionAtLeast(this.coreVersion, 1, 13)) {
-      rules.push({
-        action: 'sniff',
-      });
-    }
-
-    // A2. 出口 IP 探针钉死路由（必须紧随 sniff、先于一切分流/进程规则，确保短路不受分流策略影响）：
-    //   probe-direct-in → direct（auto_detect_interface 绑物理网卡，TUN 模式下也是真直连出口）
-    //   probe-proxy-in  → proxy-selector（现有 selector，clash_api 热切换节点后探针自动跟随）
-    // 由此在「三种接管 × 三种分流」全矩阵下分别测出真实直连出口 IP 与代理出口 IP。
-    if (this.probeDirectPort && this.probeProxyPort) {
-      rules.push(
-        { inbound: ['probe-direct-in'], action: 'route', outbound: 'direct' },
-        { inbound: ['probe-proxy-in'], action: 'route', outbound: selectedServerTag }
-      );
-    }
-
-    // 1. 强制放行 sing-box 核心进程：防止流量回流死循环
-    // 必须放在最高优先级，确保核心组件的请求能直连物理网卡
-    // 注意：不要把 FlowZ (主进程) 放在直连里，否则会干扰 FlowZ 自身的 GitHub 核心下载和测速。
-    rules.push({
-      process_name: ['sing-box', 'sing-box.exe'],
-      action: 'route',
-      outbound: 'direct',
-    });
-
-    // C. 强制引导核心 DNS 直连（必须在 hijack-dns 之前！）
-    // 把已知 bootstrap DNS IP 放在 hijack-dns 之前，无论哪个进程发包都走直连，彻底断环。
-    // 注意：这里只应该放国内的 DNS IP。如果放 8.8.8.8，会导致用户去 ping 8.8.8.8 时走直连被墙！
-    const customDomesticDns = this.getCustomDomesticDnsEndpoint(config);
-    rules.push({
-      ip_cidr: [
-        // bootstrap-direct 国内 DNS（含 1.12.12.12 DNSPod IP-DoH）：单一真值 shared/dns#BOOTSTRAP_DIRECT_DNS_IPS，
-        // 与 SystemDnsManager 的受控 DNS IP 选择守卫共用（受控 IP 绝不能落此列表，否则逃逸 hijack）。
-        ...BOOTSTRAP_DIRECT_DNS_IPS.map((ip) => `${ip}/32`),
-        // 用户自定义的国内 DNS（IP 型）也须在 hijack-dns 之前直连放行，否则其 53 端口查询会被劫持成 FakeIP
-        ...(customDomesticDns
-          ? [`${customDomesticDns.ip}/${isIpv6Host(customDomesticDns.ip) ? 128 : 32}`]
-          : []),
-        // 方案B：DNS 接管的内网 LAN 解析器（dns-lan 指向它）必须在 hijack-dns 之前直连放行，否则其 :53 查询会被
-        // hijack-dns 抢走 → 内网域名解析成环。私网 IPv4 /32（getLanResolverForDns 已保证私网，亦经下方私网直连可达）。
-        ...(this.lanResolverForDns ? [`${this.lanResolverForDns}/32`] : []),
-      ],
-      port: Array.from(new Set([53, 443, ...(customDomesticDns ? [customDomesticDns.port] : [])])),
-      action: 'route',
-      outbound: 'direct',
-    });
-
-    // D. DNS 劫持（必须在引导 DNS IP 直连之后）
-    // 劫持所有其余 port 53 流量（浏览器/系统 DNS），返回 FakeIP
-    rules.push({
-      port: [53],
-      action: 'hijack-dns',
-    });
-
-    // F. 静默屏蔽 ICMP 流量（FakeIP 下常见，但代理节点通常不支持）
-    // 放置在靠前位置，防止 ICMP 流量误入不支持的代理出站引发报错
-    rules.push({
-      protocol: 'icmp',
-      action: 'reject',
-    });
-
-    rules.push({
-      process_name: [
-        'Surge',
-        'Surge 4',
-        'Surge 5',
-        'Clash',
-        'Clash for Windows',
-        'ClashX',
-        'ClashX Pro',
-        'clash-meta',
-        'Quantumult X',
-        'sing-box',
-        'sing-box.exe',
-        'mDNSResponder',
-        'apsd',
-        'nsurlsessiond',
-        'airportd',
-        'syspolicyd',
-        'trustd',
-        'ocspd',
-        'securityd',
-        'taskgated',
-        'findmydeviced',
-        'cloudd',
-      ],
-      action: 'route',
-      outbound: 'direct',
-    });
-
-    const routeConfig: SingBoxRouteConfig = {
-      rules,
-      // 核心修复：default_domain_resolver 使用 IP-based DoH 引导解析器 (dns-bootstrap)，
-      // 既避免解析 doh.pub 域名时的死循环，又免疫 UDP 53 限速/劫持（dns-bootstrap 同为 IP-based）。
-      default_domain_resolver: 'dns-bootstrap',
-      auto_detect_interface: true,
-      // 如果模式是全局代理 (global/proxy)，则最终出口是所选节点
-      final: proxyMode === 'direct' ? 'direct' : selectedServerTag,
-    };
-
-    // 【DNS 引导与辅助直连】：
-    // 确保以下公共 DNS IP 不会被后面的 block 规则拦截，从而保证 DoH 握手和初次域名解析。
-    // 海外 DNS 不应该强行直连（否则在国内会被黑洞）。移除原本强行直连 8.8.8.8 / 1.1.1.1 的设定。
-
-    // 【终极绝杀隐私 DoH 泄漏】：
-    // 现代浏览器会尝试通过常规 HTTPS 端口向特定域名发起 DoH 请求。
-    // 这里 reject 这些 DoH 域名（发 RST，让浏览器立即回退，而非 block 静默丢包等 21s 重传超时），
-    // 迫使浏览器退回系统标准 UDP 53，重新被 hijack-dns 捕获进入 DNS 分流/FakeIP 体系。
-    // 与下方同组 DoH 域名的 QUIC/UDP-443 reject 规则保持行为一致。
-    rules.push({
-      domain_keyword: DOH_LEAK_DOMAIN_KEYWORDS,
-      port: [443, 853],
-      action: 'reject',
-    });
-
-    // 排除全部代理节点的域名/IP，确保到任一节点的连接走直连（防回流死循环 + 兼容无缝切换/代理链）。
-    // CDN 安全：域名节点用纯域名规则(domain + domain_suffix，靠 sniff 出的 SNI 精确匹配节点域名)，
-    //   不预解析为共享 CDN IP（共享 IP 加直连会误伤同 IP 的被墙站点、且抗不住 IP 轮换）；
-    //   去掉过宽的 domain_keyword（会误匹配任意"含该域名串"的无关域名）。
-    // 仅用户显式填的 IP-literal 节点用 ip_cidr 排除（专用 IP、非共享，安全）。
-    // 扩展到全部节点(不止选中)：切节点 / detour 前置代理无需重生成配置即被豁免。
-    // 必须放在其他规则之前，否则可能被 geosite-cn 匹配导致死循环。
-    {
-      const ipSet = new Set<string>();
-      const domainSet = new Set<string>();
-      for (const s of config.servers) {
-        const hosts = [s.address, s.tlsSettings?.serverName].filter(
-          (h): h is string => !!h && h.length > 0
-        );
-        for (const host of hosts) {
-          if (isIpv4Host(host)) ipSet.add(`${host}/32`);
-          else if (isIpv6Host(host)) ipSet.add(`${host}/128`);
-          else domainSet.add(host);
-        }
-      }
-
-      if (domainSet.size > 0) {
-        const domains = Array.from(domainSet);
-        rules.push({
-          // domain(精确，= 节点 SNI) + domain_suffix(仅 .${d}，匹配子域)。不放裸 d 进 domain_suffix：
-          // 那是 raw 后缀匹配，会把共享 apex 下别的真实站点也沉降到直连。
-          domain: domains,
-          domain_suffix: domains.map((d) => `.${d}`),
-          action: 'route',
-          outbound: 'direct',
-        });
-      }
-
-      if (ipSet.size > 0) {
-        rules.push({
-          ip_cidr: Array.from(ipSet),
-          action: 'route',
-          outbound: 'direct',
-        });
-      }
-    }
-
-    // 0a. U盾/安全插件的本地伪域名 → 强制 127.0.0.1，完全跳过 DNS
-    // windows10.microdone.cn 等域名是 U盾厂商注册在本地的专用域名，公网 DNS 中不存在。
-    // 普通 direct outbound 会先做 DNS 解析 → NXDOMAIN → 连接失败。
-    // 版本分支：
-    //   1.12.x → 使用 direct-loopback outbound（outbound 层面 override_address）
-    //   1.13+  → 使用路由规则层面的 override_address（outbound 层面已移除此功能）
-    const UKEY_LOCAL_DOMAINS = ['.microdone.cn'];
-    const otherBankDomains = DOMESTIC_BANK_AND_STOCK_DOMAINS.filter(
-      (d) => !UKEY_LOCAL_DOMAINS.includes(d)
-    );
-
-    if (coreVersionAtLeast(this.coreVersion, 1, 13)) {
-      // 1.13+：路由规则支持 override_address
-      rules.push({
-        domain_suffix: UKEY_LOCAL_DOMAINS,
-        action: 'route',
-        outbound: 'direct',
-        override_address: '127.0.0.1',
-      });
-    } else {
-      // 1.12.x：使用专用的 direct-loopback outbound
-      rules.push({
-        domain_suffix: UKEY_LOCAL_DOMAINS,
-        action: 'route',
-        outbound: 'direct-loopback',
-      });
-    }
-
-    // 0b. 其余银行/证券域名 → 普通 direct（正常 DNS 解析，这些域名在公网真实存在）
-    if (otherBankDomains.length > 0) {
-      rules.push({
-        domain_suffix: otherBankDomains,
-        action: 'route',
-        outbound: 'direct',
-      });
-    }
-
-    // 0c. endpoint 节点（WireGuard/Tailscale）的「配置路由段」强制路由到**该节点自身 tag**——优先于下方私网直连、
-    //     **独立于 bypass-LAN 开关、独立于全局选中**。单一真值：节点路由由其 allowedIPs(WG) / routes+tailnet(TS)
-    //     决定（endpointForcedRouteCidrs），不再有独立「绕过局域网排除段」。指向节点自身 tag（非 selector）→ 该段
-    //     恒走其 mesh 节点、与全局选中无关，实现「某段走 WG/TS、上网走全局节点」。仅 userspace（system 由内核接口
-    //     路由——Phase 2）；仅非 direct。按 config.servers 顺序=隐式优先级；被跳过节点的死引用由
-    //     fixRouteDeadReferences 兜底改写 selector。
-    if (proxyMode !== 'direct') {
-      // 仅对【本轮实际发射】的 endpoint 节点强制路由：未就绪/被跳过的 Tailscale 节点不进 pendingEndpoints，
-      // 其 tag 若仍 force-route 会被末尾 fixRouteDeadReferences 改写成 selector → 该段误流向全局选中节点。
-      // emitted 集合天然只含 endpoint 协议（仅 WG/TS 进 pendingEndpoints），故无需再按协议预判。
-      const emittedEndpointTags = new Set(this.pendingEndpoints.map((e) => e.tag));
-      // 跨 endpoint 去重：同一 CIDR 被多个 endpoint 声明（如两个 Tailscale 节点都含 tailnet 100.64.0.0/10，
-      // 或两个 WG 节点 allowedIPs 重叠），sing-box 路由首条命中 → 后者静默失效。按 config.servers 顺序，
-      // 先声明者占有该段（隐式优先级），后者重复段跳过 + 累计告警，杜绝「无声误路由到另一节点」。
-      const claimedCidrs = new Set<string>();
-      let forceRouteConflicts = 0;
-      for (const s of config.servers) {
-        const tag = idToTagMap.get(s.id);
-        if (!tag || !emittedEndpointTags.has(tag)) continue;
-        const cidrs = endpointForcedRouteCidrs(s).filter((c) => {
-          if (claimedCidrs.has(c)) {
-            forceRouteConflicts++;
-            return false;
-          }
-          claimedCidrs.add(c);
-          return true;
-        });
-        if (cidrs.length > 0) {
-          rules.push({ ip_cidr: cidrs, action: 'route', outbound: tag });
-        }
-      }
-      if (forceRouteConflicts > 0) {
-        this.logToManager(
-          'warn',
-          `${forceRouteConflicts} 个 endpoint 路由段被多个节点重复声明，已按节点顺序去重（先声明者生效）`
-        );
-      }
-    }
-
-    // 1. 私有 IP 段直连（内网地址不应该经过代理）。仅当用户未关闭"绕过局域网"时添加。
-    if (config.bypassLAN !== false) {
-      rules.push({
-        ip_cidr: PRIVATE_IP_CIDRS,
-        action: 'route',
-        outbound: 'direct',
-      });
-      // 私有/本地**域名**直连（geosite-private，补 ip_cidr 的域名盲区，如路由器后台域）。geosite-private 由
-      // BUILTIN_GEO_RULESETS 随包 bundle + 自动更新 fswatch 热加载；仅在本地 .srs 有效时加规则，缺失则跳过
-      // （不引用不存在的 rule_set，避免 FATAL）——与上面 getLocalGeoRuleSets 的缺失即跳过一致。
-      if (isValidSrsFile(path.join(this.getRuleSetRuntimeDir(), 'geosite-private.srs'))) {
-        rules.push({ rule_set: 'geosite-private', action: 'route', outbound: 'direct' });
-      }
-    }
-
-    // Bug 4 修复：删除此处重复的 QUIC 阻断规则
-    // 第一条 QUIC reject 规则已在上方（生成 routeConfig 之前）添加，此处重复添加会造成规则冗余
-    // reject 比 block 更合适（发 TCP RST 让浏览器立即回退到 TCP，而不是静默丢弃造成等待超时）
-
-    // 3. 自定义规则 + 应用分流（用户路由）——**仅 smart 模式**：global=真·全局忽略用户分流（一律走选中节点，
-    //    下方 smart geo 也不加、final=proxy），direct=全直连。功能性强制直连（防环/LAN/网银/节点排除）在本块之外、不受影响。
-    if (proxyMode === 'smart') {
-      const { rules: customRules, ruleSets: customRuleSets } = this.generateCustomRules(
-        this.effectiveCustomRules(config),
-        config.customRuleSets || [],
-        config.selectedServerId || undefined,
-        idToTagMap,
-        selectedServerTag,
-        config.ruleResources || [],
-        usesFakeIp(config) // FakeIP 启用 → 注册外化 bypass 规则的 DNS rule_set 条目（供 generateDnsConfig 引用）
-      );
-      // 走代理的自定义规则同样要配对 udp443 reject（终止规则、在末尾兜底前命中）。逐条插入：
-      // 代理向规则前先放一条同匹配器的 udp443 reject；direct/block 规则不配对。
-      // udp443 reject matcher 提取：复制规则上除「动作/出站/目的端口/network」外的全部匹配字段，使 process/
-      // regex/source_ip/source_port 等各类代理向规则同样配对（修原先仅覆盖 5 字段的缺口）。
-      // 仅排除目的 port/port_range——它们与 udp443RejectRule 的 port:443 冲突；source_port 不冲突，可配对（修 P2-1）。
-      // type/mode/rules 也排除：logical 规则单独走嵌套 AND 路径（见下），default 规则本就无这些字段（防御）。
-      const UDP443_MATCHER_EXCLUDE = new Set([
-        'action',
-        'outbound',
-        'network',
-        'port',
-        'port_range',
-        'type',
-        'mode',
-        'rules',
-      ]);
-      for (const cr of customRules) {
-        const isProxyOut =
-          cr.action === 'route' &&
-          !!cr.outbound &&
-          cr.outbound !== 'direct' &&
-          cr.outbound !== 'block';
-        if (isProxyOut && blockProxyQuic) {
-          if (cr.type === 'logical') {
-            // logical 规则顶层不接受 network/port（sing-box 解码会 FATAL）→ 把原 logical matcher 与 udp443
-            // 条件再套一层 AND logical（headless 子规则可带 network/port）：(原 logical 命中) ∧ (udp:443) → reject。
-            rules.push({
-              action: 'reject',
-              type: 'logical',
-              mode: 'and',
-              rules: [
-                { type: 'logical', mode: cr.mode, rules: cr.rules },
-                { network: ['udp'], port: [443] },
-              ],
-            });
-          } else {
-            const matcher: Record<string, any> = {};
-            for (const [k, v] of Object.entries(cr)) {
-              if (!UDP443_MATCHER_EXCLUDE.has(k) && v != null) matcher[k] = v;
-            }
-            if (Object.keys(matcher).length > 0) {
-              rules.push(udp443RejectRule(matcher));
-            }
-          }
-        }
-        rules.push(cr);
-      }
-
-      if (customRuleSets.length > 0) {
-        if (!routeConfig.rule_set) {
-          routeConfig.rule_set = [];
-        }
-        routeConfig.rule_set.push(...customRuleSets);
-      }
-
-      // 排除进程：兼容旧配置的兜底（新数据已由 ConfigManager 迁移为 customRules 的 processName+direct 规则）。
-      // 位于自定义规则之后、应用分流之前；任意更早的自定义规则可覆盖它，并非"最高优先级"。
-      if (config.bypassProcesses && config.bypassProcesses.length > 0) {
-        rules.push({
-          process_name: config.bypassProcesses,
-          action: 'route',
-          outbound: 'direct',
-        });
-      }
-
-      // 应用分流规则（真·应用分流，基于进程名）
-      // 优先级高于后续的智能分流/全局分流，确保特定应用的流量始终走用户指定的出口
-      for (const appRule of this.effectiveAppRules(config)) {
-        if (!appRule.enabled) continue;
-        const preset = getAppPreset(appRule.appId, config.customAppPresets);
-        if (!preset) continue;
-
-        // 确定出站方式
-        let outbound = 'direct';
-        if (appRule.action === 'proxy') {
-          // rule-sel-app 恒存在（generateRuleSelectors 为所有 proxy appRule 生成）：outbound 恒指
-          // rule-sel-app-<appId>，「默认/跟全局」= default=proxy-selector（嵌套），「指定节点」= default=节点 tag。
-          // 使「节点↔默认」= rule-sel-app default 变（PUT 热切换），非 outbound 结构变（重启）。
-          outbound = `rule-sel-app-${appRule.appId}`;
-        } else if (appRule.action === 'block') {
-          outbound = 'block';
-        }
-
-        // 走代理的 app 分流也要配对 udp443 reject（这些是终止规则、在末尾兜底之前命中，否则 blockQuic
-        // 对该应用的 QUIC 失效）。direct/block 不配对。
-        const appOutIsProxy = outbound !== 'direct' && outbound !== 'block';
-
-        // a. 基于进程名的规则（最精准，适用于 macOS/Windows TUN 模式）
-        if (preset.processNames && preset.processNames.length > 0) {
-          if (appOutIsProxy) {
-            const r = proxyUdpRejectFor({ process_name: preset.processNames });
-            if (r) rules.push(r);
-          }
-          rules.push({
-            process_name: preset.processNames,
-            action: 'route',
-            outbound,
-          });
-        }
-
-        // b. 基于原有 rule_set 的规则（兜底，基于域名/IP 识别）
-        const ruleSets = [
-          ...preset.geositeTags.map((tag) => `geosite-${tag}`),
-          ...(preset.geoipTags || []).map((tag) => `geoip-${tag}`),
-        ];
-
-        if (ruleSets.length > 0) {
-          if (appOutIsProxy) {
-            const r = proxyUdpRejectFor({ rule_set: ruleSets });
-            if (r) rules.push(r);
-          }
-          rules.push({
-            rule_set: ruleSets,
-            action: 'route',
-            outbound,
-          });
-        }
-      }
-    }
-
-    // 【QUIC 阻断】：放在自定义规则和应用分流之后，确保用户的 direct/proxy 规则优先级更高
-    // 这样游戏设为直连时，进程名匹配在前，游戏的 UDP 流量不会被误拒。
-    // 仅阻断浏览器的 DoH over QUIC，迫使浏览器回退到系统 UDP 53 + hijack-dns 体系。
-    // 重要：不能全量 reject 所有 UDP 443，否则 Hysteria2/TUIC 等 QUIC 协议节点会被误伤。
-    rules.push(udp443RejectRule({ domain_keyword: DOH_LEAK_DOMAIN_KEYWORDS }));
-
-    // 【DNS 死循环防范】：sing-box 本地 DNS 解析器的请求必须强制直连，否则在全局代理模式下会产生死循环
-    // 兼容 Windows 1.12.x 版本，不使用 DNS 配置里的 detour
-    rules.push({
-      protocol: 'dns',
-      action: 'route',
-      outbound: 'direct',
-    });
-
-    rules.push({
-      // #57：1.12.12.12（DNSPod IP-DoH，节点域名解析器 DNSPod 档）与 223.5.5.5 同列直连放行
-      ip_cidr: ['223.5.5.5/32', '1.12.12.12/32'],
-      port: [53, 443],
-      action: 'route',
-      outbound: 'direct',
-    });
-
-    rules.push({
-      domain_suffix: ['doh.pub'],
-      action: 'route',
-      outbound: 'direct',
-    });
-
-    // 【通用修复：Chrome/Edge 心跳 beacon 域名强制直连 —— global 和 smart 模式均生效】
-    // gvt2.com / gvt1.com 是 Google CDN 心跳；clientservices / oauthaccountmanager /
-    // optimizationguide-pa 是 Chrome 账号同步、FCM Push 和优化引导的后台服务。
-    // 这些域名对代理节点出口通常限速或屏蔽（非浏览行为），一旦持续超时会耗尽连接池，
-    // 导致所有正常网页也超时 —— 即"过一会就断网"现象。
-    // 在 global 和 smart 两种模式下均强制直连，彻底消除对连接池的占用。
-    if (proxyMode !== 'direct') {
-      rules.push({
-        domain_suffix: [
-          // Google CDN 心跳
-          'gvt2.com',
-          'gvt1.com',
-          // Chrome 账号同步 / FCM Push 后台
-          'oauthaccountmanager.googleapis.com',
-          'clientservices.googleapis.com',
-          // Chrome 优化引导服务
-          'optimizationguide-pa.googleapis.com',
-          // Google FCM 推送 (port 5228)
-          'mtalk.google.com',
-          // Android 客户端服务
-          'android.clients.google.com',
-          // Chrome / GMS clients
-          'clients1.google.com',
-          'clients2.google.com',
-          'clients3.google.com',
-          'clients4.google.com',
-          'clients5.google.com',
-          'clients6.google.com',
-          // 自动更新检查
-          'update.googleapis.com',
-        ],
-        action: 'reject',
-      });
-    }
-
-    // 智能分流规则（仅在智能分流模式下启用）
-    if (proxyMode === 'smart') {
-      // 已移除 ::/0 block，因为 block 是静默丢包，会导致 Chrome 等浏览器在发起 TCP SYN 包时陷入漫长的 21 秒重传等待（Happy Eyeballs 假死），
-      // 从而让用户以为“所有的海外网站全都打不开了”。我们必须依靠浏览器的原生 fallback，或者直接让 Mac 本机关闭 IPv6 分配。
-
-      // 针对 Google 核心服务（搜索/YouTube/Gmail 等）的关键词兜底规则（仅在未专门设置应用分流时作为备份）
-      // 注意：这些规则在 AppRules 之后，所以不会覆盖用户手动指定的节点
-      const googleKeywords = ['google', 'gmail', 'youtube', 'gstatic', 'googleapis', 'googlevideo'];
-
-      // 代理向 UDP（smart）：在每条"→代理"规则之前配对一条 reject，使该走代理的 UDP 在被路由到代理
-      // 前就 reject——不能中继的节点拦全部 UDP，能中继+blockQuic 仅拦 QUIC(UDP443)。下方 CN 直连规则
-      // 不配对，故 CN/直连 UDP 不受影响（兜底见 generateRouteConfig 末尾）。
-      const googleUdpReject = proxyUdpRejectFor({ domain_keyword: googleKeywords });
-      if (googleUdpReject) rules.push(googleUdpReject);
-      rules.push({
-        domain_keyword: googleKeywords,
-        action: 'route',
-        outbound: selectedServerTag,
-      });
-
-      // 国外域名走代理
-      const foreignUdpReject = proxyUdpRejectFor({ rule_set: 'geosite-geolocation-!cn' });
-      if (foreignUdpReject) rules.push(foreignUdpReject);
-      rules.push({
-        rule_set: 'geosite-geolocation-!cn',
-        action: 'route',
-        outbound: selectedServerTag,
-      });
-      // 中国域名直连
-      rules.push({
-        rule_set: 'geosite-cn',
-        action: 'route',
-        outbound: 'direct',
-      });
-      // 中国 IP 直连
-      rules.push({
-        rule_set: 'geoip-cn',
-        action: 'route',
-        outbound: 'direct',
-      });
-    }
-
-    // 添加 rule_set（除非是直连模式）
-    // 直连模式下不需要 rule_set，因为全部走 direct
-    if (proxyMode !== 'direct') {
-      if (!routeConfig.rule_set) {
-        routeConfig.rule_set = [];
-      }
-      // 路径取自与 copyRuleSetsToUserData 同一真值表，杜绝目录/文件名漂移
-      const runtimeDir = this.getRuleSetRuntimeDir();
-      for (const rs of this.getLocalGeoRuleSets()) {
-        const filePath = path.join(runtimeDir, rs.fileName);
-        // 缺失/损坏即跳过：不引用不存在的本地文件（否则 sing-box `initialize rule-set` FATAL）。
-        // fail-closed：缺失不再远程兜底——引用该 tag 的路由规则由 generateRouteConfig 末尾「悬空引用剪枝」剪掉，
-        // 重新在「规则资源」页下载后定义恢复 → 规则自动恢复（应用分流仍按进程名生效）。
-        if (!isValidSrsFile(filePath)) continue;
-        routeConfig.rule_set.push({ tag: rs.tag, type: 'local', format: 'binary', path: filePath });
-      }
-    }
-
-    // 添加自定义规则和应用分流所需的 Geosite/GeoIP rule_set
-    const { geosite: customGeositeCategories, geoip: customGeoipCategories } =
-      this.getRequiredGeoCategories(
-        this.effectiveCustomRules(config),
-        this.effectiveAppRules(config),
-        config.customAppPresets || []
-      );
-
-    // fail-closed：自定义规则 / 应用分流引用的 geo（geosite-<cat>/geoip-<cat>）统一由「规则资源」管理——
-    // 随包内置（上方已注入本地定义）或用户在规则资源页下载的本地 .srs；运行期零远程下载、缺失绝不远程兜底。
-    // 缺失本地副本 → 不注入定义；末尾「悬空引用剪枝」剪掉引用该 tag 的规则（应用分流只掉 geo 半、进程名仍生效；
-    // 自定义规则按 AND/OR 合理坍缩）。重新在规则资源页下载后定义恢复 → 规则自动恢复（download 触发 core reload）。
-    if (
-      proxyMode !== 'direct' &&
-      (customGeositeCategories.size > 0 || customGeoipCategories.size > 0)
-    ) {
-      if (!routeConfig.rule_set) {
-        routeConfig.rule_set = [];
-      }
-      // 已下载进规则资源的本地副本（id 形如 geosite-<cat>/geoip-<cat>）→ 注入 type:'local'；缺失/损坏返回 null。
-      const localResPath = (tag: string): string | null => {
-        const r = (config.ruleResources || []).find((x) => x.id === tag);
-        if (!r) return null;
-        const p = path.join(getRuleResourcesPath(), r.fileName);
-        return isValidSrsFile(p) ? p : null;
-      };
-      // 已有本地定义（随包内置已在上方注入）→ 跳过；否则用规则资源页的本地副本；再否则缺失（不注入，末尾剪枝）。
-      const definedTags = new Set(routeConfig.rule_set.map((rs) => rs.tag));
-      const addLocalGeo = (tag: string): void => {
-        if (definedTags.has(tag)) return;
-        const local = localResPath(tag);
-        if (local) {
-          routeConfig.rule_set!.push({ tag, type: 'local', format: 'binary', path: local });
-          definedTags.add(tag);
-        }
-        // 缺失：不注入、不远程兜底 → 交末尾悬空引用剪枝（fail-closed）。
-      };
-      for (const category of Array.from(customGeositeCategories))
-        addLocalGeo(`geosite-${category}`);
-      for (const category of Array.from(customGeoipCategories)) addLocalGeo(`geoip-${category}`);
-    }
-
-    // 【代理向 QUIC 兜底】：放在所有直连/分流规则之后，拦截"会落到 final(代理)"的剩余 QUIC(udp443)。
-    // global 模式拦全部代理向 QUIC；smart 模式拦未被上方 →代理 配对 reject 命中的（CN 已直连豁免）。
-    // 只拦 QUIC——非 QUIC 的代理向 UDP 若节点不能中继，由 sing-box 出站层自动拒绝（见上方 blockProxyQuic）。
-    if (blockProxyQuic) {
-      rules.push(udp443RejectRule());
-    }
-
-    // rule_set 按 tag 去重（保留首次=本地 .srs 优先于远程）：用户加 geosite/geoip cn 等规则时，其远程
-    // rule_set tag 会与 getLocalGeoRuleSets 的本地 geosite-cn/geoip-cn 撞名 → sing-box 启动 FATAL
-    // (duplicate rule-set tag)。去重后撞名项复用本地 .srs，行为更优（无需下载）。(修 review P0)
-    if (routeConfig.rule_set && routeConfig.rule_set.length > 0) {
-      const seenTags = new Set<string>();
-      routeConfig.rule_set = routeConfig.rule_set.filter((rs) => {
-        if (seenTags.has(rs.tag)) return false;
-        seenTags.add(rs.tag);
-        return true;
-      });
-    }
-
-    // fail-closed 兜底：剪掉引用「未定义 rule_set tag」的路由规则——本地 geo 缺失/损坏未注入定义即成悬空引用，
-    // 否则 sing-box `initialize rule-set` FATAL 崩整个代理。复用 applyRuleSetPrune 的三态递归剪枝（string/array/logical）：
-    // smart geo 缺失→该方向规则 skip；app 规则只掉 geo 半（进程名规则独立保留）；自定义规则 AND/OR 合理坍缩。
-    // 重下缺失资源后定义恢复 → 规则自动恢复（RuleResourceManager.download 触发 core reload）。
-    {
-      const definedTags = new Set((routeConfig.rule_set ?? []).map((rs) => rs.tag));
-      const referenced = new Set<string>();
-      const collectRefs = (rules: SingBoxRouteRule[]): void => {
-        for (const rule of rules) {
-          if (Array.isArray(rule.rules)) collectRefs(rule.rules);
-          const rs = rule.rule_set;
-          if (typeof rs === 'string') referenced.add(rs);
-          else if (Array.isArray(rs)) for (const t of rs) referenced.add(t);
-        }
-      };
-      collectRefs(routeConfig.rules);
-      const dangling = new Set(Array.from(referenced).filter((t) => !definedTags.has(t)));
-      if (dangling.size > 0) {
-        this.applyRuleSetPrune({ route: routeConfig } as SingBoxConfig, dangling);
-        this.logToManager(
-          'warn',
-          `规则资源：${Array.from(dangling).join(', ')} 缺少本地副本，已跳过引用它的规则以避免代理启动失败` +
-            `（在「规则资源」页下载后自动恢复；应用分流仍按进程名生效）`
-        );
-      }
-    }
-
-    return routeConfig;
-  }
-
-  /**
-   * 收集自定义规则中使用的 Geosite 类别（同时扫描 customRules 和 appRules）
-   */
-  /**
-   * 收集自定义规则中使用的 Geosite 和 GeoIP 类别（同时扫描 customRules 和 appRules）
-   */
-  private getRequiredGeoCategories(
-    customRules: import('../../shared/types').Rule[],
-    appRules: import('../../shared/types').AppRule[] = [],
-    customAppPresets: import('../../shared/types').CustomAppPreset[] = []
-  ): { geosite: Set<string>; geoip: Set<string> } {
-    const geositeCategories = new Set<string>();
-    const geoipCategories = new Set<string>();
-
-    // 扫描 geosite / geoip 类型的自定义规则（values 即裸标签，如 'youtube'/'cn'）
-    for (const rule of customRules) {
-      if (!rule.enabled) continue;
-      // 扫所有条件（多条件规则的 logical 内可含 geosite/geoip → 否则其 remote rule_set 不被注入致引用缺失）
-      for (const cond of ruleConditions(rule)) {
-        if (cond.type === 'geosite') {
-          for (const t of cond.values) {
-            const tag = t.trim().toLowerCase();
-            if (tag) geositeCategories.add(tag);
-          }
-        } else if (cond.type === 'geoip') {
-          for (const t of cond.values) {
-            const tag = t.trim().toLowerCase();
-            if (tag) geoipCategories.add(tag);
-          }
-        }
-      }
-    }
-
-    // 扫描应用分流规则
-    for (const appRule of appRules) {
-      if (!appRule.enabled) continue;
-      const preset = getAppPreset(appRule.appId, customAppPresets);
-      if (preset) {
-        preset.geositeTags.forEach((tag) => geositeCategories.add(tag));
-        if (preset.geoipTags) {
-          preset.geoipTags.forEach((tag) => geoipCategories.add(tag));
-        }
-      }
-    }
-    return { geosite: geositeCategories, geoip: geoipCategories };
-  }
-
-  private generateCustomRules(
-    customRules: import('../../shared/types').Rule[],
-    customRuleSets: import('../../shared/types').CustomRuleSet[] = [],
-    selectedServerId?: string,
-    idToTagMap?: Map<string, string>,
-    selectedServerTag: string = 'proxy',
-    ruleResources: import('../../shared/types').RuleResource[] = [],
-    registerDnsBypass: boolean = false // FakeIP 启用时：为外化 bypassFakeIP 规则注册 <base>-dns rule_set 条目
-  ): { rules: SingBoxRouteRule[]; ruleSets: SingBoxRuleSet[] } {
-    const rules: SingBoxRouteRule[] = [];
-    const ruleSets: SingBoxRuleSet[] = [];
-
-    // 目的地 OR 组：这些 type 的字段在单条 default rule 内原生 OR（sing-box: domain||suffix||keyword||regex||ip_cidr）。
-    const OR_GROUP = new Set(['domain', 'domainSuffix', 'domainKeyword', 'domainRegex', 'ipCidr']);
-    const pushU = <T>(arr: T[] | undefined, items: T[]): T[] => [...(arr || []), ...items];
-
-    // 把一个条件的 type→字段累积到 target（值并集；geosite/geoip→rule_set tag；ruleSet→注册定义+tag）。返回 hasMatcher。
-    const applyConditionFields = (
-      cond: import('../../shared/types').RuleCondition,
-      target: SingBoxRouteRule
-    ): boolean => {
-      // EXT 类型（域名/IP/端口/进程 系）委托单一真值 condMatcherFields——与外化文件内容永不漂移。
-      // geosite/geoip/ruleSet 不可 headless 表达，留 inline（下方 switch）。
-      if (EXT_TYPES.has(cond.type)) {
-        const fields = condMatcherFields(cond);
-        if (!fields) return false;
-        const t = target as Record<string, unknown>;
-        for (const k of Object.keys(fields)) {
-          t[k] = pushU(t[k] as unknown[] | undefined, fields[k] as unknown[]);
-        }
-        return true;
-      }
-      const vals = (cond.values || []).map((v) => v.trim()).filter(Boolean);
-      if (vals.length === 0) return false;
-      switch (cond.type) {
-        case 'geosite':
-          // 裸标签 → geosite-<tag>（lowercase 对齐 getRequiredGeoCategories 与远程 .srs 文件名）
-          target.rule_set = pushU(
-            target.rule_set as string[] | undefined,
-            vals.map((t) => `geosite-${t.toLowerCase()}`)
-          );
-          return true;
-        case 'geoip':
-          target.rule_set = pushU(
-            target.rule_set as string[] | undefined,
-            vals.map((t) => `geoip-${t.toLowerCase()}`)
-          );
-          return true;
-        case 'ruleSet': {
-          // 远程 URL → custom-ruleset-N（format 按扩展名）；本地 res:<id> → 本地 rule_set。tag 去重（重复 tag 会 FATAL）。
-          const seen = new Set(
-            Array.isArray(target.rule_set)
-              ? target.rule_set
-              : target.rule_set != null
-                ? [target.rule_set]
-                : []
-          );
-          for (const v of vals) {
-            let tag: string;
-            if (v.startsWith('res:')) {
-              const resId = v.slice(4);
-              // 内置随包资源（res:builtin:<tag>）：复用 b.tag（geosite-netflix 等），runtime 路径取自单一真值表。
-              // proxy 模式下 getLocalGeoRuleSets 已注入同 tag 本地 rule_set；此处自注入 + route 装配末尾按 tag 去重
-              // （keep-first，两者同路径同 format，行为一致），故 direct 模式或本地缺失也能自洽。文件缺失/损坏 → 跳过（不 FATAL）。
-              const builtinMeta = resolveBuiltinRuleSetRefMeta(resId);
-              if (builtinMeta) {
-                const filePath = path.join(this.getRuleSetRuntimeDir(), builtinMeta.fileName);
-                if (!isValidSrsFile(filePath)) {
-                  this.logToManager(
-                    'warn',
-                    `ruleSet 规则引用的内置资源文件缺失/损坏，已跳过: ${builtinMeta.fileName}`
-                  );
-                  continue;
-                }
-                tag = builtinMeta.tag;
-                if (!ruleSets.some((rs) => rs.tag === tag)) {
-                  ruleSets.push({ tag, type: 'local', format: 'binary', path: filePath });
-                }
-              } else {
-                const res = ruleResources.find((rr) => rr.id === resId);
-                if (!res) {
-                  this.logToManager('warn', `ruleSet 规则引用的资源不存在，已跳过: ${v}`);
-                  continue;
-                }
-                const filePath = path.join(getRuleResourcesPath(), res.fileName);
-                // 与 builtin 分支对齐：用 isValidSrsFile（验 SRS 魔数）而非裸 existsSync——
-                // 损坏/半写文件存在但非法时注入 local rule_set 会让 sing-box 加载 FATAL，须 fail-closed 跳过。
-                if (!isValidSrsFile(filePath)) {
-                  this.logToManager(
-                    'warn',
-                    `ruleSet 规则引用的资源文件缺失/损坏，已跳过: ${res.fileName}`
-                  );
-                  continue;
-                }
-                tag = `local-rs-${res.id}`;
-                if (!ruleSets.some((rs) => rs.tag === tag)) {
-                  ruleSets.push({ tag, type: 'local', format: res.format, path: filePath });
-                }
-              }
-            } else {
-              // fail-closed：远程 URL rule_set 能力已移除——所有 srs 统一由「规则资源」页管理（res:<id> 引用本地副本）。
-              // 旧配置 / 旁路写入的裸 URL 值 → 跳过（不再生成运行期远程下载），日志告知改用规则资源。
-              this.logToManager(
-                'warn',
-                `ruleSet 规则的远程 URL 已不再支持，请改用「规则资源」下载后引用，已跳过: ${v}`
-              );
-              continue;
-            }
-            seen.add(tag);
-          }
-          if (seen.size > 0) {
-            target.rule_set = Array.from(seen);
-            return true;
-          }
-          return false;
-        }
-      }
-      return false;
-    };
-
-    // 每条 Rule → 单 default rule（目的地 OR 组合并：单条件 或 combineMode!=='and' 且全为 OR 组类型）
-    // 或 logical rule（跨维度/含 rule_set 的 OR、或 AND）。绝不把「应 OR」的条件塞进单 default rule 当 AND。
-    for (const rule of customRules) {
-      if (!rule.enabled) continue;
-
-      // ── L3 外化分流：全条件可 headless 表达 → 写独立 rule_set 文件，route 仅引用 tag（改值热重载零重启）。
-      const plan = planCustomRule(rule);
-      const extBase = customRuleFileBase(rule.id);
-      // DNS 文件注册须在 ext-skip 跳过之前：route 侧 fail-closed 跳过的 bypass 规则，DNS 侧今日仍消费其域名值。
-      if (registerDnsBypass && rule.bypassFakeIP && plan.kind !== 'inline' && plan.dnsRules) {
-        const dnsPath = path.join(getCustomRulesDir(), `${extBase}.dns.json`);
-        if (
-          require('fs').existsSync(dnsPath) &&
-          !ruleSets.some((rs) => rs.tag === `${extBase}-dns`)
-        ) {
-          ruleSets.push({ tag: `${extBase}-dns`, type: 'local', format: 'source', path: dnsPath });
-        }
-      }
-      if (plan.kind === 'ext-skip') continue; // 全 EXT 但 fail-closed → 无 route 规则（DNS 已上方处理）
-      if (plan.kind === 'ext') {
-        const filePath = path.join(getCustomRulesDir(), `${extBase}.json`);
-        if (require('fs').existsSync(filePath)) {
-          if (!ruleSets.some((rs) => rs.tag === extBase)) {
-            ruleSets.push({ tag: extBase, type: 'local', format: 'source', path: filePath });
-          }
-          const extRule: SingBoxRouteRule = { action: 'route', rule_set: [extBase] };
-          this.applyRuleAction(
-            extRule,
-            rule.action,
-            rule.targetServerId,
-            selectedServerId,
-            idToTagMap,
-            selectedServerTag,
-            rule.id
-          );
-          rules.push(extRule);
-          continue;
-        }
-        // 文件未落盘（预检/onRetry 重生成期）→ 标记降级，回落 inline 现状（值已知，功能不损）。
-        this.customRuleFilesDegraded = true;
-      }
-
-      const rawConds = ruleConditions(rule);
-      const conds = rawConds
-        .map((c) => ({
-          type: c.type,
-          values: (c.values || []).map((v) => v.trim()).filter(Boolean),
-        }))
-        .filter((c) => c.values.length > 0);
-      if (conds.length === 0) continue;
-      // AND 模式任一条件在预过滤被丢弃（值全空白/空数组）→ 整条跳过（fail-closed）：否则会以剩余条件的
-      // 「超集」下发（含坍缩成单 default rule 绕过 logical 分支的 fail-closed）。UI 已拦空值，此为旁路写/手改兜底。
-      if (rule.combineMode === 'and' && conds.length < rawConds.length) continue;
-
-      const mergeable =
-        conds.length === 1 ||
-        (rule.combineMode !== 'and' && conds.every((c) => OR_GROUP.has(c.type)));
-
-      let finalRule: SingBoxRouteRule | null = null;
-      if (mergeable) {
-        const singboxRule: SingBoxRouteRule = { action: 'route' };
-        let hasMatcher = false;
-        for (const c of conds) if (applyConditionFields(c, singboxRule)) hasMatcher = true;
-        if (hasMatcher) finalRule = singboxRule;
-      } else {
-        // logical：每条件一个纯 matcher 子规则（无 action/outbound，action/outbound 在外层 logical 规则）
-        const subRules: SingBoxRouteRule[] = [];
-        let dropped = false;
-        for (const c of conds) {
-          const sub: SingBoxRouteRule = {};
-          if (applyConditionFields(c, sub)) subRules.push(sub);
-          else dropped = true; // 该条件无法产出 matcher（如 ruleSet 资源缺失、端口全非法）
-        }
-        // AND 模式任一子条件被丢弃 → 整条跳过（fail-closed）：否则剩余 AND 子集会匹配「超集」（比用户意图更宽），
-        // 与最小权限/旧行为（整条 skip）相悖。OR 模式丢弃失败子条件无害（少一个候选项）。
-        if (rule.combineMode === 'and' && dropped) {
-          finalRule = null;
-        } else if (subRules.length === 1) {
-          finalRule = { ...subRules[0], action: 'route' };
-        } else if (subRules.length > 1) {
-          finalRule = {
-            action: 'route',
-            type: 'logical',
-            mode: rule.combineMode || 'or',
-            rules: subRules,
-          };
-        }
-      }
-      if (!finalRule) continue;
-
-      this.applyRuleAction(
-        finalRule,
-        rule.action,
-        rule.targetServerId,
-        selectedServerId,
-        idToTagMap,
-        selectedServerTag,
-        rule.id
-      );
-      rules.push(finalRule);
-    }
-
-    // fail-closed：legacy customRuleSets（迁移后通常为空，ConfigManager 已转成 customRules）原走运行期远程下载，
-    // 远程能力已移除 → 仅日志告知，不再生成 remote rule_set / 路由规则。
-    for (const ruleSet of customRuleSets) {
-      if (!ruleSet.enabled || !ruleSet.url) continue;
-      this.logToManager(
-        'warn',
-        `legacy 远程规则集已不再支持（请用「规则资源」），已跳过: ${ruleSet.url}`
-      );
-    }
-
-    return { rules, ruleSets };
-  }
-
-  /**
-   * 应用规则动作到 sing-box 规则对象
-   */
-  private applyRuleAction(
-    singboxRule: SingBoxRouteRule,
-    action: string,
-    targetServerId?: string,
-    _selectedServerId?: string, // 3b 后不再用于条件判断（保留位参以不动调用方），下划线跳过未用检查
-    idToTagMap?: Map<string, string>,
-    selectedServerTag: string = 'proxy',
-    ruleId?: string // rule-sel selector 命名所需（customRule.id）；缺省则回落旧行为
-  ): void {
-    // 设置出站
-    if (action === 'proxy') {
-      // anti-drift（铁律）：指定了目标节点的规则 → 指向独立 rule-sel-<ruleId> selector，**绝不直绑节点、
-      // 绝不经共享 proxy-selector**。直绑节点会让热切全局节点后这条固定规则漂走（复活旧 quirk）；经共享
-      // proxy-selector 会让「固定规则」随全局节点漂移。rule-sel 独立 selector 使「改规则目标节点」可经
-      // clash_api PUT /proxies/rule-sel-<id> 热切换、且互不干扰。selector 由 generateRuleSelectors 预建；
-      // 极端情况（selector 未生成，如降级/onRetry 中间态）→ 死引用由 fixRouteDeadReferences 兜底改写为
-      // proxy-selector，启动不 FATAL。
-      // rule-sel 恒存在（generateRuleSelectors 为所有 proxy 规则生成）：outbound 恒指 rule-sel-<ruleId>，
-      // 「默认/跟全局」= rule-sel default=proxy-selector（嵌套），「指定节点」= default=节点 tag。
-      if (ruleId) {
-        singboxRule.outbound = `rule-sel-${ruleId}`;
-      } else if (targetServerId) {
-        const targetTag = idToTagMap?.get(targetServerId);
-        singboxRule.outbound = targetTag || `proxy-${targetServerId}`;
-      } else {
-        singboxRule.outbound = selectedServerTag;
-      }
-    } else if (action === 'direct') {
-      singboxRule.outbound = 'direct';
-    } else if (action === 'block') {
-      singboxRule.outbound = 'block';
-    } else {
-      // 如果没有指定，默认使用主节点
-      singboxRule.outbound = selectedServerTag;
     }
   }
 
@@ -4991,7 +2133,7 @@ done
           throw new Error('没有可用的代理节点出站（全部节点未通过启动前配置校验）');
         }
         if (ob.default && toRemove.has(ob.default)) {
-          ob.default = this.prunedSelectorDefault(ob.tag, ob.outbounds);
+          ob.default = prunedSelectorDefault(ob.tag, ob.outbounds);
         }
       }
     }
@@ -5025,10 +2167,6 @@ done
    * - proxy-selector / urltest 自身：仍落剩余首成员（首存活节点是正确兜底）。
    * 注：rule-sel 的 outbounds 恒含 'proxy-selector'（generateRuleSelectors emit），故该回落目标必有效。
    */
-  private prunedSelectorDefault(tag: string | undefined, remainingOutbounds: string[]): string {
-    return tag?.startsWith('rule-sel-') ? 'proxy-selector' : remainingOutbounds[0];
-  }
-
   /**
    * route 规则指向「已被剔除/不存在的出站」（死引用）→ 统一修正为 proxy-selector，避免 sing-box 以
    * "outbound not found" 启动失败。从 generateSingBoxConfig 末尾抽出的纯等价重构（A5），gate 剔除后复用。
@@ -5086,7 +2224,7 @@ done
     ]);
     if (raceTimer) clearTimeout(raceTimer);
     const unreachable = new Set(checks.filter((c) => !c.reachable).map((c) => c.tag));
-    const dropped = this.applyRuleSetPrune(singboxConfig, unreachable);
+    const dropped = applyRuleSetPrune(singboxConfig, unreachable);
     if (dropped.length > 0) {
       this.logToManager(
         'warn',
@@ -5102,50 +2240,6 @@ done
    *   ② 路由规则的 `rule_set`（string | string[]）剔除这些 tag；剔空则整条规则丢弃（含 logical 子规则递归）。
    * 返回实际丢弃的 rule_set tag 列表。本地 bundled rule_set（geosite-cn 等，非 remote）永不入 `unreachable`，不受影响。
    */
-  private applyRuleSetPrune(singboxConfig: SingBoxConfig, unreachable: Set<string>): string[] {
-    if (unreachable.size === 0 || !singboxConfig.route) return [];
-    const dropped: string[] = [];
-    if (singboxConfig.route.rule_set) {
-      singboxConfig.route.rule_set = singboxConfig.route.rule_set.filter((rs) => {
-        if (rs.tag && unreachable.has(rs.tag)) {
-          dropped.push(rs.tag);
-          return false;
-        }
-        return true;
-      });
-    }
-    const pruneRules = (rules: SingBoxRouteRule[]): SingBoxRouteRule[] =>
-      rules.filter((rule) => {
-        if (Array.isArray(rule.rules)) {
-          const before = rule.rules.length;
-          rule.rules = pruneRules(rule.rules);
-          // logical 规则子条件被剪空 → 整条丢（空 logical 无意义且 sing-box 不接受）
-          if (rule.type === 'logical' && rule.rules.length === 0) return false;
-          // AND logical 任一子条件被剪掉 → 整条丢（fail-closed：缺失条件无法满足，留剩余子集会以「超集」匹配）。
-          // 关键：修「派生的嵌套 AND udp443 reject」——内层 geo logical 被剪空丢弃后，外层 AND 仅剩 {network,port}
-          // 否则坍缩成无差别全局 udp443 reject（误伤 Hysteria2/TUIC 握手与全局 QUIC）；与 generateCustomRules
-          // 的 AND fail-closed 语义一致（任一子条件丢→整条 skip）。OR / 无 mode 不受影响（少一候选项无害）。
-          if (rule.type === 'logical' && rule.mode === 'and' && rule.rules.length < before) {
-            return false;
-          }
-        }
-        const rs = rule.rule_set;
-        if (typeof rs === 'string') {
-          if (unreachable.has(rs)) return false; // 唯一 rule_set 缺失 → 该规则停止生效
-        } else if (Array.isArray(rs)) {
-          const kept = rs.filter((t) => !unreachable.has(t));
-          if (kept.length === 0) return false; // rule_set 全缺失 → 该规则停止生效
-          if (kept.length !== rs.length) rule.rule_set = kept; // 部分保留（如 geosite 留、geoip 丢）
-        }
-        return true;
-      });
-    // 以 unreachable.size 为闸（已 early-return size===0），不以 dropped.length——fail-closed 下复用此函数剪
-    // 「本地缺失、无 rule_set 定义」的悬空 tag 时 dropped 恒为 0，但仍须剪掉引用它的规则（否则悬空引用 FATAL）。
-    if (unreachable.size > 0) {
-      singboxConfig.route.rules = pruneRules(singboxConfig.route.rules ?? []);
-    }
-    return dropped;
-  }
 
   /**
    * HEAD 预检单个远程 rule_set URL 是否存在。仅"确定缺失"(404/403/410) 返回 false→剪枝；2xx/3xx 返回 true；
