+++ src/main/services/__tests__/singbox-inbounds-builder.test.ts
@@ -1,0 +1,106 @@
+/**
+ * buildInbounds 单测 —— 原 ProxyManager.generateInbounds 无单测（仅 config-snapshot 集成锁字节）。
+ * 锁：mixed inbound（listenAddr/端口/legacy sniff）/ probe inbound 注入 / TUN inbound（平台排除段/MTU/stack/
+ * IPv6/macOS http_proxy platform）/ allowLan。依赖（PRIVATE_IP_CIDRS 等）均为 config-helpers 纯函数。
+ */
+jest.mock('electron', () => ({
+  app: { getPath: () => '/fake/userData', getAppPath: () => '/fake/app', isPackaged: false },
+  net: {},
+}));
+
+import { buildInbounds, type InboundsDeps } from '../singbox-inbounds-builder';
+import type { SingBoxInbound } from '../singbox-config-types';
+import type { UserConfig } from '../../../shared/types';
+import { withPlatform } from './platform-test-utils';
+
+const deps = (over: Partial<InboundsDeps> = {}): InboundsDeps => ({
+  coreVersion: '1.13.0',
+  probeDirectPort: null,
+  probeProxyPort: null,
+  ...over,
+});
+
+const cfg = (over: Partial<UserConfig>): UserConfig =>
+  ({
+    proxyModeType: 'systemProxy',
+    servers: [],
+    selectedServerId: 's1',
+    ...over,
+  }) as unknown as UserConfig;
+
+/** 按 tag 取 inbound 并断言存在（消 strict undefined）。 */
+function byTag(ibs: SingBoxInbound[], tag: string): SingBoxInbound {
+  const ib = ibs.find((i) => i.tag === tag);
+  expect(ib).toBeTruthy();
+  return ib as SingBoxInbound;
+}
+
+describe('buildInbounds — mixed + probe', () => {
+  it('systemProxy：恒含 mixed-in（listen 127.0.0.1）；1.13 不带 legacy sniff', () => {
+    const ibs = withPlatform('linux', () => buildInbounds(cfg({}), undefined, deps()));
+    const mixed = byTag(ibs, 'mixed-in');
+    expect(mixed.listen).toBe('127.0.0.1');
+    expect(mixed.sniff).toBeUndefined(); // 1.13 路由层 sniff
+  });
+
+  it('allowLan=true → mixed listen=::（监听全部接口）', () => {
+    const ibs = withPlatform('linux', () =>
+      buildInbounds(cfg({ allowLan: true }), undefined, deps())
+    );
+    expect(byTag(ibs, 'mixed-in').listen).toBe('::');
+  });
+
+  it('coreVersion 1.12 → mixed 带 legacy sniff/sniff_override_destination', () => {
+    const ibs = withPlatform('linux', () =>
+      buildInbounds(cfg({}), undefined, deps({ coreVersion: '1.12.8' }))
+    );
+    const mixed = byTag(ibs, 'mixed-in');
+    expect(mixed.sniff).toBe(true);
+    expect(mixed.sniff_override_destination).toBe(true);
+  });
+
+  it('probe 端口注入 → 增 probe-direct-in/probe-proxy-in（对应端口）', () => {
+    const ibs = withPlatform('linux', () =>
+      buildInbounds(cfg({}), undefined, deps({ probeDirectPort: 21001, probeProxyPort: 21002 }))
+    );
+    expect(byTag(ibs, 'probe-direct-in').listen_port).toBe(21001);
+    expect(byTag(ibs, 'probe-proxy-in').listen_port).toBe(21002);
+  });
+
+  it('probe 端口缺失（null）→ 不注入探针', () => {
+    const ibs = withPlatform('linux', () => buildInbounds(cfg({}), undefined, deps()));
+    expect(ibs.map((i) => i.tag)).not.toContain('probe-direct-in');
+  });
+});
+
+describe('buildInbounds — TUN', () => {
+  it('TUN：增 tun-in；route_exclude_address 含回环；macOS gvisor 栈 + http_proxy platform', () => {
+    const ibs = withPlatform('darwin', () =>
+      buildInbounds(cfg({ proxyModeType: 'tun' }), undefined, deps())
+    );
+    const tun = byTag(ibs, 'tun-in');
+    expect(tun.stack).toBe('gvisor'); // macOS 强制 gvisor
+    expect(tun.route_exclude_address).toContain('127.0.0.0/8');
+    expect(tun.platform?.http_proxy?.enabled).toBe(true); // macOS http_proxy
+  });
+
+  it('Windows TUN + bypassLAN：排除段含私网 CIDR + 核心 DNS IP', () => {
+    const ibs = withPlatform('win32', () =>
+      buildInbounds(cfg({ proxyModeType: 'tun' }), undefined, deps())
+    );
+    const tun = byTag(ibs, 'tun-in');
+    expect(tun.stack).toBe('system'); // Windows system 栈
+    expect(tun.route_exclude_address).toContain('10.0.0.0/8'); // bypassLAN 默认 true
+    expect(tun.route_exclude_address).toContain('223.5.5.5/32'); // 核心 DNS 防回流
+  });
+
+  it('选中节点为 IP 字面量 → 排除该节点 IP（防回流死循环）', () => {
+    const c = cfg({
+      proxyModeType: 'tun',
+      servers: [{ id: 's1', address: '1.2.3.4', protocol: 'vless', port: 443 }],
+      selectedServerId: 's1',
+    } as unknown as Partial<UserConfig>);
+    const ibs = withPlatform('linux', () => buildInbounds(c, undefined, deps()));
+    expect(byTag(ibs, 'tun-in').route_exclude_address).toContain('1.2.3.4/32');
+  });
+});
