+++ src/main/services/singbox-outbound-builder.ts
@@ -1,0 +1,909 @@
+/**
+ * 代理 Outbound 构造（sing-box 1.12.x / 1.13.x）—— 从 ProxyManager 抽出（SingBoxConfigBuilder 抽取 Phase 2 step 2）。
+ * 纯函数簇、零实例状态：只读入参 server / idToTagMap / nodeResolverTag，被 config-gen（generateOutbounds）
+ * 与测速（buildSpeedTestOutbound）共用。逐协议字段映射 + TLS/Reality/传输层 + PR-6 抗封后处理。
+ */
+
+import * as path from 'path';
+import type { ServerConfig, UserConfig, InvalidNodeInfo } from '../../shared/types';
+import type { SingBoxOutbound, SingBoxEndpoint } from './singbox-config-types';
+import { resourceManager } from './ResourceManager';
+import { coreVersionAtLeast } from '../../shared/version';
+import { isEndpointProtocol } from '../../shared/endpoint-routes';
+import { getUserDataPath } from '../utils/paths';
+import {
+  effectiveCustomRules,
+  effectiveAppRules,
+  getNodeResolverTag,
+} from './singbox-config-helpers';
+
+/** 节点是否可用：naive 需要 libcronet 核心库，缺库时不可用（会被跳过、分流/选中回退到 selector）。 */
+export function isNodeUsable(server: ServerConfig): boolean {
+  if (server.protocol.toLowerCase() === 'naive' && !resourceManager.hasCronetLib()) {
+    return false;
+  }
+  return true;
+}
+
+/** selector default 被剔除后的兜底：rule-sel 回 proxy-selector（嵌套），proxy-selector 落剩余首节点（MED-1）。 */
+export function prunedSelectorDefault(
+  tag: string | undefined,
+  remainingOutbounds: string[]
+): string {
+  return tag?.startsWith('rule-sel-') ? 'proxy-selector' : remainingOutbounds[0];
+}
+
+/** WireGuard endpoint 构造（sing-box 1.11+ 顶层 endpoints[]）。config-gen 与测速共用。 */
+export function buildWireGuardEndpoint(server: ServerConfig, tag: string): SingBoxEndpoint {
+  const s = server.wireguardSettings;
+  if (!s || !s.privateKey || !s.peerPublicKey || !s.localAddress?.length) {
+    throw new Error('WireGuard 配置缺少 privateKey/peerPublicKey/localAddress');
+  }
+  return {
+    type: 'wireguard',
+    tag,
+    system: false,
+    mtu: s.mtu && s.mtu > 0 ? s.mtu : 1408,
+    address: s.localAddress,
+    private_key: s.privateKey,
+    peers: [
+      {
+        address: server.address,
+        port: server.port,
+        public_key: s.peerPublicKey,
+        allowed_ips: s.allowedIPs?.length ? s.allowedIPs : ['0.0.0.0/0', '::/0'],
+        ...(s.preSharedKey ? { pre_shared_key: s.preSharedKey } : {}),
+        // keepalive 缺省强制 25s：WireGuard 是无连接 UDP，client 多在 NAT 后，无 keepalive 则 NAT 映射超时
+        // → 入向不可达、隧道静默断连。故未显式设置时兜 25s 维持映射，避免断连（用户可在表单调大）。
+        persistent_keepalive_interval:
+          s.persistentKeepalive && s.persistentKeepalive > 0 ? s.persistentKeepalive : 25,
+        ...(s.reserved?.length === 3 ? { reserved: s.reserved } : {}),
+      },
+    ],
+  };
+}
+
+/**
+ * 生成代理 Outbound 配置（sing-box 1.12.x / 1.13.x 兼容格式）
+ */
+export function buildProxyOutbound(
+  server: ServerConfig,
+  idToTagMap: Map<string, string>,
+  // #57：节点域名 dial 解析器 tag，由调用方传 getNodeResolverTag(config,'dial')。
+  // 缺省 dns-bootstrap（AliDNS IP-DoH）= 现状，兼容无 config 上下文的兜底调用。
+  nodeResolverTag: string = 'dns-bootstrap'
+): SingBoxOutbound {
+  // sing-box 要求协议类型必须是小写
+  const protocol = server.protocol.toLowerCase();
+  const protocolLower = protocol;
+  const tlsProtocols = ['trojan', 'anytls', 'hysteria2', 'tuic'];
+
+  // 自定义协议（raw-JSON 透传）：原样下发用户的 outbound 对象，仅【强制覆盖 tag】（防撞/防注入），
+  // 语义与能否启用由 sing-box check（启动 gate / 表单 probe）判定，FlowZ 不解析 type。
+  if (protocol === 'custom') {
+    // 强制覆盖 tag、剥离 detour：内层 detour 会绕过 FlowZ 的代理链环检测（仅走 server.detour）且 sing-box check
+    // 测不出 detour 环 → 运行时拨号死循环。FlowZ 经 detour 字段统一管理链路，自定义透传不携带内层 detour。
+    const { detour: _drop, ...userOutbound } = (server.customSettings?.outbound || {}) as Record<
+      string,
+      unknown
+    >;
+    return {
+      ...userOutbound,
+      tag: idToTagMap.get(server.id) || `proxy-${server.id}`,
+    } as unknown as SingBoxOutbound;
+  }
+
+  const outbound: SingBoxOutbound = {
+    type: protocol,
+    tag: idToTagMap.get(server.id) || `proxy-${server.id}`,
+    server: server.address,
+    server_port: server.port,
+    // 代理节点域名经引导解析（默认 dns-bootstrap=AliDNS IP-DoH），免疫 UDP 53 限速/劫持，
+    // 避免节点解析失败导致全断流；同时防止 dns-local 死循环导致的连接挂起。
+    // #57：节点域名解析器档位可改为 dns-node(DNSPod)/dns-local(系统 DNS)，dial 与 rule1 同档（统一 tag）。
+    domain_resolver: nodeResolverTag,
+  };
+
+  // vless/vmess UDP 封装：默认 xudp；可经 server.packetEncoding 覆盖（兼容拒收 xudp 的旧核心/服务端，
+  // 否则 UDP 断流且无从调整）。显式设为空串则省略该字段（不下发 packet_encoding，由核心用其默认）。
+  const packetEncoding = server.packetEncoding ?? 'xudp';
+
+  // VLESS 特定配置
+  if (protocol === 'vless') {
+    outbound.uuid = server.uuid;
+    if (server.flow) {
+      outbound.flow = server.flow;
+    }
+    if (packetEncoding) {
+      outbound.packet_encoding = packetEncoding;
+    }
+  }
+
+  // VMess 特定配置
+  if (protocol === 'vmess') {
+    outbound.uuid = server.uuid;
+    outbound.security = server.vmessSecurity || 'auto';
+    outbound.alter_id = server.alterId || 0;
+    if (packetEncoding) {
+      outbound.packet_encoding = packetEncoding;
+    }
+  }
+
+  // Trojan 特定配置
+  if (protocol === 'trojan') {
+    outbound.password = server.password;
+  }
+
+  // Hysteria2 特定配置
+  if (protocol === 'hysteria2') {
+    outbound.password = server.password;
+
+    // 带宽限制
+    if (server.hysteria2Settings?.upMbps) {
+      outbound.up_mbps = server.hysteria2Settings.upMbps;
+    }
+    if (server.hysteria2Settings?.downMbps) {
+      outbound.down_mbps = server.hysteria2Settings.downMbps;
+    }
+
+    // 混淆配置
+    if (server.hysteria2Settings?.obfs?.type && server.hysteria2Settings?.obfs?.password) {
+      outbound.obfs = {
+        type: server.hysteria2Settings.obfs.type,
+        password: server.hysteria2Settings.obfs.password,
+      };
+    }
+
+    // 网络类型 (tcp/udp)
+    if (server.hysteria2Settings?.network) {
+      outbound.network = server.hysteria2Settings.network;
+    }
+  }
+
+  // AnyTLS 特定配置
+  if (protocol === 'anytls') {
+    outbound.password = server.password;
+    // AnyTLS 的 TLS 永远开启，这里不需要额外处理，类型检查结尾部分统一生成
+    // AnyTLS 会话参数
+    if (server.anyTlsSettings?.idleSessionCheckInterval) {
+      outbound.idle_session_check_interval = server.anyTlsSettings.idleSessionCheckInterval;
+    }
+    if (server.anyTlsSettings?.idleSessionTimeout) {
+      outbound.idle_session_timeout = server.anyTlsSettings.idleSessionTimeout;
+    }
+    if (server.anyTlsSettings?.minIdleSession !== undefined) {
+      outbound.min_idle_session = server.anyTlsSettings.minIdleSession;
+    }
+  }
+
+  // Shadowsocks 特定配置
+  if (protocol === 'shadowsocks') {
+    if (!server.shadowsocksSettings) {
+      throw new Error(`Shadowsocks server ${server.name} missing settings`);
+    }
+    outbound.method = server.shadowsocksSettings.method;
+    outbound.password = server.shadowsocksSettings.password;
+    if (server.shadowsocksSettings.plugin) {
+      outbound.plugin = server.shadowsocksSettings.plugin;
+      outbound.plugin_opts = server.shadowsocksSettings.pluginOptions;
+    }
+  }
+
+  // TUIC 特定配置
+  if (server.protocol === 'tuic') {
+    outbound.uuid = server.uuid;
+    outbound.password = server.password;
+
+    if (server.tuicSettings) {
+      if (server.tuicSettings.congestionControl) {
+        outbound.congestion_control = server.tuicSettings.congestionControl;
+      }
+      if (server.tuicSettings.udpRelayMode) {
+        outbound.udp_relay_mode = server.tuicSettings.udpRelayMode;
+      }
+      if (server.tuicSettings.zeroRttHandshake !== undefined) {
+        outbound.zero_rtt_handshake = server.tuicSettings.zeroRttHandshake;
+      }
+      if (server.tuicSettings.heartbeat) {
+        outbound.heartbeat = server.tuicSettings.heartbeat;
+      }
+    }
+  }
+
+  // NaiveProxy 特定配置
+  if (server.protocol === 'naive') {
+    outbound.username = server.username;
+    outbound.password = server.password;
+
+    // NaiveProxy specific configuration
+    // sing-box 的 naive outbound 由 Cronet 自管 TLS，仅支持 server_name / certificate /
+    // certificate_path / ech。下发 alpn 或 insecure:true 会让 sing-box 拒启：
+    //   FATAL: initialize outbound: alpn is not supported on naive outbound
+    //   FATAL: initialize outbound: insecure is not supported on naive outbound
+    // 故这里只下发 server_name（用户在节点上设置的 alpn / allowInsecure 对 naive 无效，忽略）。
+    // 参考：https://sing-box.sagernet.org/configuration/outbound/naive/ （TLS 字段限制）
+    //       SagerNet/sing-box protocol/naive/outbound.go (v1.13.x) ~L47 的 ALPN/insecure 校验
+    outbound.tls = {
+      enabled: true,
+      server_name: server.tlsSettings?.serverName || server.address,
+    };
+
+    // HTTP/3：naive 经 quic:true 走 h3(QUIC/UDP) 拨号传输，对应服务端 `--listen=quic://`。
+    // 注意这只改变"拨号传输"：naive 仍只过 TCP（HTTP CONNECT）、不能中继客户端 UDP（除非
+    // udp_over_tcp，FlowZ 不下发）。客户端 QUIC 若走到 naive，由 blockQuic（udp443 reject 逼回退 TCP）
+    // 或 sing-box 出站层（"UDP is not supported by outbound"）处理；该拨号本身受 fwmark 保护、
+    // 绕过 route 规则，不被 reject 误杀（已实测）。
+    if (server.naiveSettings?.useHttp3) {
+      outbound.quic = true;
+    }
+  }
+
+  // SOCKS 特定配置
+  if (server.protocol === 'socks') {
+    if (server.username) outbound.username = server.username;
+    if (server.password) outbound.password = server.password;
+    // 默认 SOCKS 版本
+    (outbound as any).version = '5';
+  }
+
+  // HTTP 特定配置
+  if (server.protocol === 'http') {
+    if (server.username) outbound.username = server.username;
+    if (server.password) outbound.password = server.password;
+
+    // HTTP outbound headers mapping can be added if needed via server.httpSettings.headers
+    if (server.httpSettings?.headers) {
+      if (!outbound.transport) outbound.transport = { type: 'http' };
+      outbound.transport.headers = server.httpSettings.headers;
+    }
+    if (server.httpSettings?.path) {
+      if (!outbound.transport) outbound.transport = { type: 'http' };
+      outbound.transport.path = server.httpSettings.path;
+    }
+  }
+
+  // SSH 特定配置
+  if (server.protocol === 'ssh') {
+    const ssh = server.sshSettings || {};
+    if (ssh.user) outbound.user = ssh.user;
+    if (ssh.password) outbound.password = ssh.password;
+    if (ssh.privateKey) outbound.private_key = ssh.privateKey;
+    if (ssh.privateKeyPath) outbound.private_key_path = ssh.privateKeyPath;
+    if (ssh.privateKeyPassphrase) outbound.private_key_passphrase = ssh.privateKeyPassphrase;
+    if (ssh.hostKey && ssh.hostKey.length > 0) outbound.host_key = ssh.hostKey;
+    if (ssh.hostKeyAlgorithms && ssh.hostKeyAlgorithms.length > 0)
+      outbound.host_key_algorithms = ssh.hostKeyAlgorithms;
+    if (ssh.clientVersion) outbound.client_version = ssh.clientVersion;
+
+    // SSH outbound 不需要 TLS 和传输层配置，直接返回
+    return outbound;
+  }
+
+  // TLS 配置 (非 Naive 协议，因为 Naive 已在前一段处理了 tls 结构)
+  if (
+    server.protocol !== 'naive' &&
+    (server.security === 'tls' || server.tlsSettings || tlsProtocols.includes(protocol))
+  ) {
+    // 为 Trojan 设置默认 ALPN ["http/1.1"] 以提高兼容性
+    let finalAlpn = server.tlsSettings?.alpn;
+    if (!finalAlpn && protocolLower === 'trojan') {
+      finalAlpn = ['http/1.1'];
+    }
+
+    outbound.tls = {
+      enabled: true,
+      server_name: server.tlsSettings?.serverName || server.address,
+      insecure: server.tlsSettings?.allowInsecure || false,
+      alpn: finalAlpn,
+    };
+
+    // uTLS 仅适用于基于 TCP 的协议，Hysteria2 和 TUIC 使用 QUIC (UDP) 不支持 uTLS
+    const fingerprint = server.tlsSettings?.fingerprint;
+
+    // 默认行为：VLESS 等协议默认开启 chrome 指纹，Trojan 默认不开启（none）以通过标准 TLS 握手
+    let finalFingerprint = fingerprint;
+    if (!finalFingerprint) {
+      if (protocolLower === 'vless' || protocolLower === 'anytls') {
+        finalFingerprint = 'chrome';
+      } else {
+        finalFingerprint = 'none';
+      }
+    }
+
+    if (
+      server.protocol !== 'hysteria2' &&
+      server.protocol !== 'tuic' &&
+      finalFingerprint !== 'none'
+    ) {
+      outbound.tls.utls = {
+        enabled: true,
+        fingerprint: finalFingerprint,
+      };
+    }
+
+    // ALPN 仅在支持的协议上设置
+    if (server.tlsSettings?.alpn) {
+      outbound.tls.alpn = server.tlsSettings.alpn;
+    }
+  }
+
+  // Reality 配置
+  if (server.security === 'reality' && server.realitySettings) {
+    outbound.tls = {
+      enabled: true,
+      server_name: server.tlsSettings?.serverName || undefined,
+      insecure: server.tlsSettings?.allowInsecure || false,
+      utls: {
+        enabled: true,
+        fingerprint: server.tlsSettings?.fingerprint || 'chrome',
+      },
+      reality: {
+        enabled: true,
+        public_key: server.realitySettings.publicKey,
+        short_id: server.realitySettings.shortId || '',
+      },
+    };
+  }
+
+  // 传输层配置（不适用于 hysteria2、anytls、naive）
+  if (
+    server.protocol !== 'hysteria2' &&
+    server.protocol !== 'anytls' &&
+    server.protocol !== 'naive' &&
+    server.network &&
+    server.network !== 'tcp'
+  ) {
+    outbound.transport = generateTransportConfig(server);
+  }
+
+  // PR-6 抗封增强（ECH / TLS 分片 / Multiplex / Hy2 端口跳跃）统一后处理
+  applyAntiCensorshipOptions(outbound, server);
+
+  return outbound;
+}
+
+/**
+ * 生成传输层配置
+ */
+function generateTransportConfig(server: ServerConfig): SingBoxOutbound['transport'] {
+  if (server.network === 'ws' && server.wsSettings) {
+    // 0-RTT early-data：订阅/分享链解析时已存入 wsSettings，此前未落运行时配置导致静默失效
+    return {
+      type: 'ws',
+      path: server.wsSettings.path || '/',
+      headers: server.wsSettings.headers,
+      max_early_data: server.wsSettings.maxEarlyData,
+      early_data_header_name: server.wsSettings.earlyDataHeaderName,
+    };
+  }
+
+  if (server.network === 'grpc' && server.grpcSettings) {
+    return {
+      type: 'grpc',
+      service_name: server.grpcSettings.serviceName || '',
+    };
+  }
+
+  // httpupgrade：较 ws 更隐蔽的 HTTP Upgrade 传输（复用 ws 的 path / Host）
+  if (server.network === 'httpupgrade') {
+    return {
+      type: 'httpupgrade',
+      path: server.wsSettings?.path || '/',
+      host: server.wsSettings?.headers?.['Host'] || server.tlsSettings?.serverName,
+    };
+  }
+
+  return undefined;
+}
+
+/**
+ * PR-6 抗封增强统一后处理：ECH、每节点 TLS 分片、Multiplex(reality+vision 跳过)、Hy2 端口跳跃。
+ * 放在 outbound 构建完成后统一处理，避免散落到各协议的 tls/transport 构建点。
+ */
+function applyAntiCensorshipOptions(outbound: SingBoxOutbound, server: ServerConfig): void {
+  const protocolLower = server.protocol.toLowerCase();
+
+  // fragment 仅对「标准 sing-box TCP-TLS 栈」有意义，以下协议必须排除（否则死配置或直接启动 FATAL）：
+  //   · hy2/tuic：TLS 在 QUIC 内、无 TCP ClientHello（死配置）；
+  //   · naive：TLS 由 Cronet 自管，naive 出站直接拒绝 fragment 字段（实测
+  //     "fragment is not supported on naive outbound" → 启动 FATAL），无论 h2/h3。
+  // 注：ECH 不受此限——QUIC 与 naive(Cronet) 均原生支持 ECH。
+  const fragmentUnsupported =
+    protocolLower === 'hysteria2' || protocolLower === 'tuic' || protocolLower === 'naive';
+
+  // ECH（隐藏 SNI）+ 每节点 TLS 分片（抗 SNI-DPI）：需已有 tls 块
+  if (outbound.tls) {
+    if (server.tlsSettings?.ech) {
+      // 可选 ECHConfigList：填了下发 tls.ech.config（PEM 按行拆数组）；留空则 sing-box 从 DNS(HTTPS RR) 自取。
+      const echCfg = server.tlsSettings.echConfig?.trim();
+      const echLines = echCfg
+        ? echCfg
+            .split(/\r?\n/)
+            .map((l) => l.trim())
+            .filter(Boolean)
+        : [];
+      outbound.tls.ech =
+        echLines.length > 0 ? { enabled: true, config: echLines } : { enabled: true };
+    }
+    if (server.tlsSettings?.fragment && !fragmentUnsupported) outbound.tls.fragment = true;
+  }
+
+  // Multiplex（vless/trojan/vmess/shadowsocks）；vision flow(xtls-rprx-vision) 自带流分帧、与 mux
+  // 不兼容（与是否 reality 无关，普通 TLS+vision 同样不兼容）→ 跳过
+  const mux = server.multiplexSettings;
+  const hasVisionFlow = (server.flow || '').toLowerCase().includes('vision');
+  if (
+    mux?.enabled &&
+    ['vless', 'trojan', 'vmess', 'shadowsocks'].includes(protocolLower) &&
+    !hasVisionFlow
+  ) {
+    outbound.multiplex = {
+      enabled: true,
+      protocol: mux.protocol || 'h2mux',
+      ...(mux.maxConnections ? { max_connections: mux.maxConnections } : {}),
+      ...(mux.minStreams ? { min_streams: mux.minStreams } : {}),
+      ...(mux.padding ? { padding: true } : {}),
+    };
+  }
+
+  // Hysteria2 端口跳跃（serverPorts 为逗号分隔的范围串，如 "20000:30000"，支持多段）
+  if (protocolLower === 'hysteria2' && server.hysteria2Settings?.serverPorts) {
+    const ports = server.hysteria2Settings.serverPorts
+      .split(',')
+      .map((s) => s.trim())
+      .filter(Boolean);
+    if (ports.length > 0) {
+      outbound.server_ports = ports;
+      if (server.hysteria2Settings.hopInterval) {
+        outbound.hop_interval = server.hysteria2Settings.hopInterval;
+      }
+    }
+  }
+}
+
+// ---- generateOutbounds 本体 + selector/Tailscale 簇（SingBoxConfigBuilder 抽取 Phase 2 step 10）----
+
+/** rule-sel 载体：ruleKey→(selectorTag,memberTag) 供 generateSingBoxConfig 回填 currentRuleTargetMap 热切换定位。 */
+export interface PendingRuleSelector {
+  ruleKey: string;
+  selectorTag: string;
+  memberTag: string;
+  targetServerId?: string;
+}
+
+/** 注入依赖：generateOutbounds 原读/写的实例态（coreVersion 值 / gateInvalidNodes Map 就地改 / log 回调）。 */
+export interface OutboundsDeps {
+  coreVersion: string;
+  gateInvalidNodes: Map<string, InvalidNodeInfo>;
+  log: (level: 'debug' | 'info' | 'warn' | 'error' | 'fatal', message: string) => void;
+}
+
+/** Tailscale state 目录：`<userData>/tailscale/<serverId>`，跨重启免重认证、删节点时清理。 */
+function tailscaleStateDir(serverId: string): string {
+  return path.join(getUserDataPath(), 'tailscale', serverId);
+}
+
+function tailscaleStateExists(serverId: string): boolean {
+  try {
+    // readdirSync 在目录缺失时抛 ENOENT → catch 返 false，无需先 existsSync（省一次 stat）。
+    return require('fs').readdirSync(tailscaleStateDir(serverId)).length > 0;
+  } catch {
+    return false;
+  }
+}
+
+/**
+ * 构造 Tailscale endpoint（sing-box 1.12+ 顶层 endpoints[]）。账号制 mesh，无 server address/port。
+ * 默认 tsnet 用户态（system_interface 省略=false，零提权）；auth_key 可选（无则核日志出登录 URL）。
+ */
+function buildTailscaleEndpoint(server: ServerConfig, tag: string): SingBoxEndpoint {
+  const ts = server.tailscaleSettings || {};
+  const stateDir = tailscaleStateDir(server.id);
+  try {
+    require('fs').mkdirSync(stateDir, { recursive: true });
+  } catch {
+    /* best-effort；目录建失败时 sing-box 自身会报错 */
+  }
+  const ep: SingBoxEndpoint = { type: 'tailscale', tag, state_directory: stateDir };
+  const authKey = ts.authKey?.trim();
+  if (authKey) ep.auth_key = authKey;
+  const controlUrl = ts.controlUrl?.trim();
+  if (controlUrl) ep.control_url = controlUrl;
+  const hostname = ts.hostname?.trim();
+  if (hostname) ep.hostname = hostname;
+  const exitNode = ts.exitNode?.trim();
+  if (exitNode) {
+    ep.exit_node = exitNode;
+    if (ts.exitNodeAllowLanAccess) ep.exit_node_allow_lan_access = true;
+  }
+  if (ts.acceptRoutes) ep.accept_routes = true;
+  if (ts.ephemeral) ep.ephemeral = true;
+  const adv = (ts.advertiseRoutes || []).map((c) => c.trim()).filter(Boolean);
+  if (adv.length) ep.advertise_routes = adv;
+  // Phase 2：ts.reverseMesh → ep.system_interface=true（+ 选中节点门控 + 强制 helper）。Phase 1 用户态、省略。
+  return ep;
+}
+
+/**
+ * rule-sel selector 生成：所有 proxy 规则（custom + app）各建独立 selector（anti-drift 热切换）。
+ * pendingRuleSelectors 为 buildOutbounds 传入的载体（push 收集），由 generateSingBoxConfig 回填 currentRuleTargetMap。
+ */
+function generateRuleSelectors(
+  config: UserConfig,
+  idToTagMap: Map<string, string>,
+  nodeTags: string[],
+  outbounds: SingBoxOutbound[],
+  pendingRuleSelectors: PendingRuleSelector[]
+): void {
+  const existingTags = new Set(outbounds.map((o) => o.tag).filter((t): t is string => !!t));
+  const interrupt = config.interruptConnectionsOnSwitch === true;
+  // (selectorTag, memberTag) 收集，供 generateSingBoxConfig 末尾回填 currentRuleTargetMap。
+  pendingRuleSelectors.length = 0;
+
+  // rule-sel 恒存在（所有 proxy 规则，无论 targetServerId 有无）：members 含 proxy-selector（嵌套全局），
+  // default=target 节点（有）或 proxy-selector（无=跟全局）。使「节点↔默认」= rule-sel default 变化（PUT 热切换），
+  // 非 outbound 结构变（重启）。anti-drift 仍保留：固定节点规则 default=节点，不随全局漂。
+  const emit = (ruleKey: string, selectorTag: string, targetServerId: string | undefined) => {
+    // tag 防撞：节点名可能恰好叫 rule-sel-xxx → 加 (n) 后缀（与 getUniqueTag 同形）
+    let tag = selectorTag;
+    let n = 1;
+    while (existingTags.has(tag)) {
+      tag = `${selectorTag} (${n})`;
+      n++;
+    }
+    existingTags.add(tag);
+    const targetTag = targetServerId ? idToTagMap.get(targetServerId) : undefined;
+    // default：target 有效→节点 tag；target 无/无效→proxy-selector（跟全局，嵌套）
+    const defaultTag = targetTag && nodeTags.includes(targetTag) ? targetTag : 'proxy-selector';
+    outbounds.push({
+      type: 'selector',
+      tag,
+      outbounds: [...nodeTags, 'proxy-selector'],
+      default: defaultTag,
+      interrupt_exist_connections: interrupt,
+    });
+    pendingRuleSelectors.push({
+      ruleKey,
+      selectorTag: tag,
+      memberTag: defaultTag,
+      targetServerId,
+    });
+  };
+
+  for (const rule of effectiveCustomRules(config)) {
+    if (!rule.enabled) continue;
+    if (rule.action !== 'proxy') continue;
+    emit(`custom:${rule.id}`, `rule-sel-${rule.id}`, rule.targetServerId);
+  }
+  for (const appRule of effectiveAppRules(config)) {
+    if (!appRule.enabled) continue;
+    if (appRule.action !== 'proxy') continue;
+    // appRule 按 appId 去重（effectiveAppRules），用 appId 作 key；selectorTag 加 app- 前缀避免与 customRule 混淆。
+    emit(`app:${appRule.appId}`, `rule-sel-app-${appRule.appId}`, appRule.targetServerId);
+  }
+}
+
+/**
+ * 生成 outbounds（节点出站 + selector + rule-sel + direct/block/shadow-tls + detour 死引用预校验）。
+ * 注入 coreVersion/gateInvalidNodes/log；返回 outbounds + 两载体 pendingEndpoints/pendingRuleSelectors
+ * （由 generateSingBoxConfig 回写 this.*，供顶层 endpoints[] 注入与 hotSwitch/clash_api 回读）。
+ */
+export function buildOutbounds(
+  selectedServer: ServerConfig,
+  config: UserConfig,
+  idToTagMap: Map<string, string>,
+  deps: OutboundsDeps
+): {
+  outbounds: SingBoxOutbound[];
+  pendingEndpoints: SingBoxEndpoint[];
+  pendingRuleSelectors: PendingRuleSelector[];
+} {
+  const pendingEndpoints: SingBoxEndpoint[] = [];
+  const pendingRuleSelectors: PendingRuleSelector[] = [];
+  const outbounds: SingBoxOutbound[] = [];
+  // endpoint（WireGuard 等）单独收进 pendingEndpoints（顶层 endpoints[]），但其 tag 仍与节点出站 tag 一同
+  // 进入 nodeTags → proxy-selector / rule-sel / route，与普通节点一视同仁（clash_api 热切换已实测兼容）。
+  pendingEndpoints.length = 0;
+  const nodeTags: string[] = [];
+
+  if (config) {
+    // 生成【全部】节点的 Outbound：selector 需要列出所有可切换节点；detour 前置节点亦在 config.servers
+    // 中，一并生成、通过 detour 字段链接。单个节点配置异常不应拖垮整体配置，逐节点 try/catch 跳过。
+    // （app/custom 分流规则指向的固定节点 tag 不变，仍直接命中其节点出站，不经 selector。）
+    for (const server of config.servers) {
+      const tag = idToTagMap.get(server.id) || `proxy-${server.id}`;
+      if (nodeTags.includes(tag)) continue; // 去重（含 endpoint）
+      // 不可用节点跳过：naive 缺 libcronet 时，sing-box 启动会预初始化全部出站、缺库的 naive 会让
+      // 整个代理启动 FATAL（连非 naive 节点也用不了）。跳过后，路由层对该节点 tag 的死引用会在
+      // generateSingBoxConfig 末尾被统一修正为 selector（见 H2 修复）。
+      if (!isNodeUsable(server)) {
+        deps.log('warn', `跳过不可用节点「${server.name}」：NaiveProxy 缺少 libcronet 核心库`);
+        continue;
+      }
+      // 启动前配置校验 gate 已标记为非法的节点：跳过、不进 outbounds/selector（防 onRetry 重生成复活）。
+      // 路由层对其 tag 的死引用由 generateSingBoxConfig 末尾 fixRouteDeadReferences 统一修正为 selector。
+      if (deps.gateInvalidNodes.has(server.id)) {
+        continue;
+      }
+      // WireGuard：endpoint（非 outbound）。建进 pendingEndpoints，tag 仍入 nodeTags 参与选择器/路由。
+      if (server.protocol.toLowerCase() === 'wireguard') {
+        try {
+          pendingEndpoints.push(buildWireGuardEndpoint(server, tag));
+          nodeTags.push(tag);
+        } catch (e: any) {
+          deps.log(
+            'warn',
+            `生成 WireGuard endpoint 失败，已跳过: ${server.name} (${e?.message ?? e})`
+          );
+        }
+        continue;
+      }
+      // Tailscale：endpoint（账号制 mesh）。就绪门控——非选中且未就绪（无 authKey、无持久 state）不发射，
+      // 避免拖慢启动 + 多个未登录节点登录 URL 刷屏；选中节点即便未就绪也发射（触发交互登录 URL 流）。
+      if (server.protocol.toLowerCase() === 'tailscale') {
+        const ts = server.tailscaleSettings;
+        const ready = !!ts?.authKey?.trim() || tailscaleStateExists(server.id);
+        if (server.id !== selectedServer.id && !ready) {
+          deps.log('info', `Tailscale 节点「${server.name}」未就绪(需登录)且非选中，已跳过`);
+          continue;
+        }
+        try {
+          pendingEndpoints.push(buildTailscaleEndpoint(server, tag));
+          nodeTags.push(tag);
+        } catch (e: any) {
+          deps.log(
+            'warn',
+            `生成 Tailscale endpoint 失败，已跳过: ${server.name} (${e?.message ?? e})`
+          );
+        }
+        continue;
+      }
+      // 自定义协议标记为 endpoint 类型（第三方类 wireguard/tailscale 实现）→ 进 endpoints[]；
+      // 否则走下方普通 outbound 路径。tag 由 generateProxyOutbound/此处统一注入。
+      if (server.protocol.toLowerCase() === 'custom' && server.customSettings?.isEndpoint) {
+        const { detour: _d, ...epOutbound } = (server.customSettings.outbound || {}) as Record<
+          string,
+          unknown
+        >;
+        pendingEndpoints.push({
+          ...(epOutbound as unknown as SingBoxEndpoint),
+          tag,
+        });
+        nodeTags.push(tag);
+        continue;
+      }
+      try {
+        const ob = buildProxyOutbound(server, idToTagMap, getNodeResolverTag(config, 'dial'));
+        ob.tag = tag;
+        if (server.detour && config.servers.some((s) => s.id === server.detour)) {
+          // 环检测：沿 detour 链行进，若回到本节点即成环 → 不设 detour，避免 sing-box 报循环引用启动失败
+          const seen = new Set<string>([server.id]);
+          let cur: string | undefined = server.detour;
+          let looped = false;
+          while (cur) {
+            if (seen.has(cur)) {
+              looped = true;
+              break;
+            }
+            seen.add(cur);
+            cur = config.servers.find((s) => s.id === cur)?.detour;
+          }
+          if (looped) {
+            deps.log('warn', `检测到代理链成环，已跳过 detour: ${server.name}`);
+          } else {
+            const detourSrv = config.servers.find((s) => s.id === server.detour);
+            if (detourSrv && isEndpointProtocol(detourSrv.protocol)) {
+              // endpoint（WireGuard/Tailscale）不作为 outbound 的前置代理(detour)目标（dialer detour→endpoint 语义
+              // 未定，且本节点会被下方 detour 死引用预校验误剔）。丢弃 detour（本节点仍直连可用）+ 警告。UI 亦不列
+              // endpoint 为 detour 候选（route 规则指向 endpoint 仍合法，仅 detour 不合法——两者语义不同）。
+              deps.log(
+                'warn',
+                `endpoint 节点（WireGuard/Tailscale）不支持作为前置代理(detour)目标，已忽略「${server.name}」的代理链`
+              );
+            } else {
+              ob.detour = idToTagMap.get(server.detour);
+            }
+          }
+        }
+        outbounds.push(ob);
+        nodeTags.push(tag);
+      } catch (e: any) {
+        deps.log('warn', `生成节点出站失败，已跳过: ${server.name} (${e?.message ?? e})`);
+      }
+    }
+
+    // 全局 TLS 分片（PR-6）：开启后对所有已生成的 TCP-TLS 节点出站切分 ClientHello，抗 SNI-DPI。
+    // 跳过 hy2/tuic（QUIC 内 TLS、无 TCP ClientHello，死配置）与 naive（Cronet 自管 TLS，拒绝
+    // fragment 字段 → 启动 FATAL）。
+    // 机制选型：用 outbound 的 `tls.fragment`（按 TCP 段切分代理自身 ClientHello）而非 route action
+    // `route-options.tls_fragment`——后者作用于被嗅探出的流量（切内层），非代理自身握手入口。默认仅注入
+    // fragment=true，不注入 fragment_fallback_delay（用核心默认 500ms）/record_fragment（保持纯开关）。
+    if (config.tlsFragment) {
+      for (const ob of outbounds) {
+        if (ob.tls && ob.type !== 'hysteria2' && ob.type !== 'tuic' && ob.type !== 'naive') {
+          ob.tls.fragment = true;
+        }
+      }
+    }
+
+    // selector：列出已生成的全部节点 tag，default 指向当前选中节点；interrupt_exist_connections 由用户
+    // 开关决定（默认 false=优雅切换，现有连接保留至自然关闭）。clash_api `PUT /proxies/proxy-selector`
+    // 据此热切换、无需重启 sing-box（详见 switchMode）。路由的 final 与「→代理」规则统一指向本 selector。
+    // nodeTags 已在节点循环中按序累积（含 endpoint tag）；空=所有节点生成失败/不可用。
+    if (nodeTags.length === 0) {
+      throw new Error('没有可用的代理节点出站（所有节点配置生成失败）');
+    }
+    const selectedServerTag = idToTagMap.get(selectedServer.id) || 'proxy';
+    outbounds.push({
+      type: 'selector',
+      tag: 'proxy-selector',
+      outbounds: nodeTags,
+      default: nodeTags.includes(selectedServerTag) ? selectedServerTag : nodeTags[0],
+      interrupt_exist_connections: config.interruptConnectionsOnSwitch === true,
+    });
+
+    // rule-sel：恒优化后「所有 proxy 规则」均生成独立 selector（无论 targetServerId 有无）。
+    //   有 target → default=目标节点（anti-drift：固定规则不经共享 proxy-selector，热切全局节点后不漂走）；
+    //   无 target → default=proxy-selector（跟全局，嵌套）。「节点↔默认」切换 = rule-sel default 变（PUT 热切换），
+    //   非 outbound 结构变（重启）。members 复用 proxy-selector 的可用 nodeTags（已排除 naive 缺库 / gate-invalid /
+    //   detour 死引用剔除的节点）；目标无效/被 gate 剔除/无 target → default 落 proxy-selector 兜底
+    //   （rule-sel 仍建、不 FATAL，与 fixRouteDeadReferences 兜底语义一致）。
+    // 仅 smart 生成：用户路由仅 smart 生效，global/direct 下 effectiveCustomRules/effectiveAppRules 为空 → 无 proxy 规则、无消费者。
+    if ((config.proxyMode || 'smart').toLowerCase() === 'smart') {
+      generateRuleSelectors(config, idToTagMap, nodeTags, outbounds, pendingRuleSelectors);
+    }
+  } else {
+    // Fallback if config is missing (shouldn't happen)
+    outbounds.push(
+      buildProxyOutbound(selectedServer, idToTagMap, getNodeResolverTag(config, 'dial'))
+    );
+  }
+
+  // 直连出站
+  outbounds.push({
+    type: 'direct',
+    tag: 'direct',
+  });
+
+  // 版本条件：sing-box 1.12.x 需要在 outbound 层面做 override_address
+  // 因为 1.12 的路由规则不支持 override_address 字段（会被静默忽略）。
+  // 1.13+ 已将此功能迁移到路由规则，不需要额外的 outbound。
+  if (!coreVersionAtLeast(deps.coreVersion, 1, 13)) {
+    outbounds.push({
+      type: 'direct',
+      tag: 'direct-loopback',
+      override_address: '127.0.0.1',
+    });
+  }
+
+  // 阻断出站
+  outbounds.push({
+    type: 'block',
+    tag: 'block',
+  });
+
+  // Shadow-TLS 后处理：如果主节点或任意辅助节点使用了 Shadow-TLS，
+  // 为每个使用 Shadow-TLS 的节点插入内层 SS outbound
+  const stlsOutbounds: SingBoxOutbound[] = [];
+  for (const ob of outbounds) {
+    // 根据 tag（节点名称）反查对应的 ServerConfig；selector/direct/block 等非节点出站匹配不到 → 跳过
+    const srv = config?.servers.find((s) => idToTagMap.get(s.id) === ob.tag);
+    if (srv?.shadowTlsSettings) {
+      // 创建独立的外层 ShadowTLS outbound
+      const stlsTag = `stls-out-${srv.id}`;
+      const stlsOutbound: SingBoxOutbound = {
+        type: 'shadowtls',
+        tag: stlsTag,
+        server: srv.address,
+        server_port: srv.shadowTlsSettings.port || srv.port,
+        version: 3,
+        password: srv.shadowTlsSettings.password,
+        tls: {
+          enabled: true,
+          server_name: srv.shadowTlsSettings.sni || undefined,
+          utls: {
+            enabled: true,
+            fingerprint: srv.shadowTlsSettings.fingerprint || 'chrome',
+          },
+        },
+      };
+      stlsOutbounds.push(stlsOutbound);
+
+      // 主 outbound (原本的 shadowsocks) 必须作为应用的路由目标
+      // 所以我们保留它为 proxy (shadowsocks)，但将其 detour 指向新增的 shadowtls outbound
+      ob.detour = stlsTag;
+
+      // 当配置了 detour 后，sing-box 通常期望主 outbound 的 server/port 被忽略
+      // 但为了规范，我们可以保留 shadowsocks 的原参数或统一指向实际伪装的地址
+      // 在 ShadowTLS 架构中，外层负责 TLS 握手连接真实服务器地址，内层 SS 则是被保护的流量
+    }
+  }
+  outbounds.push(...stlsOutbounds);
+
+  // detour 引用预校验（补 check 唯一盲区：start-stage 引用解析——check 过、run 时报
+  // `dependency[X] not found for outbound[Y]` FATAL）。任何 outbound.detour 指向「生成集合内不存在的
+  // tag」→ 剔除该引用方（**不删 detour 字段**，与 pruneTagsClosure 同隐私语义；选中节点则 throw），
+  // 从 selector 删成员并记录 gateInvalidNodes；selector 剔空 throw。仅 config 存在时有 selector，需 guard。
+  if (config) {
+    const tagToServerId = (tag: string): string | undefined => {
+      if (tag.startsWith('stls-out-')) return tag.slice('stls-out-'.length);
+      for (const [id, t] of idToTagMap) {
+        if (t === tag) return id;
+      }
+      return undefined;
+    };
+    const selector = outbounds.find((o) => o.tag === 'proxy-selector');
+    const selectedTag = idToTagMap.get(selectedServer.id);
+    let mutated = false;
+    // 反复扫描：剔一个引用方可能让别的 detour 链断裂，收敛到不再有死引用。
+
+    while (true) {
+      const validTags = new Set(outbounds.map((o) => o.tag).filter((t): t is string => !!t));
+      const dead = outbounds.find(
+        (ob) => ob.detour !== undefined && !validTags.has(ob.detour) && ob.tag !== 'proxy-selector'
+      );
+      if (!dead) break;
+      if (dead.tag === selectedTag) {
+        throw new Error(
+          `选中节点「${dead.tag}」的代理链依赖的前置节点不存在，无法启动，请更换节点后重试`
+        );
+      }
+      // 删该引用方 outbound + selector 成员；记录 gateInvalidNodes。
+      const removedTag = dead.tag;
+      const sid = tagToServerId(removedTag);
+      outbounds.splice(outbounds.indexOf(dead), 1);
+      // 同步剔除所有 selector（proxy-selector + 各 rule-sel）的成员引用：rule-sel 的 members 与
+      // proxy-selector 同源（generateRuleSelectors 复用 nodeTags 快照），节点被剔后须一并清理，否则
+      // rule-sel 引用幽灵 tag → sing-box 启动 FATAL。default 命中被剔 → rule-sel 回 proxy-selector、
+      // proxy-selector 落剩余首节点（MED-1，见 prunedSelectorDefault；不 FATAL）。
+      const allSelectors = outbounds.filter(
+        (o) => o.type === 'selector' && Array.isArray(o.outbounds)
+      );
+      for (const sel of allSelectors) {
+        sel.outbounds = (sel.outbounds as string[]).filter((t) => t !== removedTag);
+        if (sel.default === removedTag) {
+          sel.default = prunedSelectorDefault(sel.tag, sel.outbounds as string[]);
+        }
+      }
+      if (sid) {
+        idToTagMap.delete(sid);
+        if (!deps.gateInvalidNodes.has(sid)) {
+          deps.gateInvalidNodes.set(sid, {
+            id: sid,
+            tag: removedTag,
+            reason: '代理链依赖的前置节点不存在（detour 引用无效）',
+          });
+        }
+      }
+      deps.log('warn', `启动前配置校验：节点「${removedTag}」的 detour 引用无效，已剔除`);
+      mutated = true;
+    }
+    if (
+      mutated &&
+      selector &&
+      Array.isArray(selector.outbounds) &&
+      selector.outbounds.length === 0
+    ) {
+      throw new Error('没有可用的代理节点出站（节点代理链依赖无效）');
+    }
+    // rule-sel 剔空（members 全被 detour 死引用剔除）：删该 selector outbound，对应 route 规则的
+    // outbound（rule-sel-<id>）成死引用 → fixRouteDeadReferences 兜底改写为 proxy-selector，启动不 FATAL。
+    if (mutated) {
+      for (let i = outbounds.length - 1; i >= 0; i--) {
+        const o = outbounds[i];
+        if (
+          o.type === 'selector' &&
+          o.tag !== 'proxy-selector' &&
+          o.tag &&
+          o.tag.startsWith('rule-sel') &&
+          Array.isArray(o.outbounds) &&
+          o.outbounds.length === 0
+        ) {
+          outbounds.splice(i, 1);
+        }
+      }
+    }
+  }
+
+  return { outbounds, pendingEndpoints, pendingRuleSelectors };
+}
